The SEIT2L model for pomp

The SEIT2L model with constant population size, uniform prior and Poisson observation with reporting rate, as a pomp object.

Format

A pomp object.

Details

Look at the documentation of pomp for more details about each of these elements.

Examples

## load pomp package
require('pomp')

## load Tristan da Cunha data
data(FluTdC1971)

## load SEITL_pomp object
example(SEITL_pomp)

#> Warning: no help found for ‘SEITL_pomp’

## define deterministic skeleton
SEIT2L.skel.c <- '
    double trans[6];

    double beta = R0 / D_inf;
    double epsilon = 1 / D_lat;
    double nu = 1 / D_inf;
    double tau = 1 / D_imm;

    double N = S + E + I + T1 + T2 + L;

    trans[0] = beta * I / N * S;
    trans[1] = epsilon * E;
    trans[2] = nu * I;
    trans[3] = 2 * tau * T1;
    trans[4] = 2 * alpha * tau * T2;
    trans[5] = 2 * (1 - alpha) * tau * T2;

    DS = -trans[0] + trans[5];
    DE = trans[0] - trans[1];
    DI = trans[1] - trans[2];
    DT1 = trans[2] - trans[3];
    DT2 = trans[3] - trans[4] - trans[5];
    DL = trans[4];
    DInc = trans[1];
'

## define stochastic model, for use with euler.sim, see ?euler.sim
SEIT2L.sim.c <- '
    double rate[6];
    double dN[6];

    double beta = R0 / D_inf;
    double epsilon = 1 / D_lat;
    double nu = 1 / D_inf;
    double tau = 1 / D_imm;

    double N = S + E + I + T1 + T2 + L;

    rate[0] = beta * I / N;
    rate[1] = epsilon;
    rate[2] = nu;
    rate[3] = 2 * tau;
    rate[4] = 2 * alpha * tau;
    rate[5] = 2 * (1 - alpha) * tau;

    reulermultinom(1, S, &rate[0], dt, &dN[0]);
    reulermultinom(1, E, &rate[1], dt, &dN[1]);
    reulermultinom(1, I, &rate[2], dt, &dN[2]);
    reulermultinom(1, T1, &rate[3], dt, &dN[3]);
    reulermultinom(2, T2, &rate[4], dt, &dN[4]);

    S += -dN[0] + dN[5];
    E += dN[0] - dN[1];
    I += dN[1] - dN[2];
    T1 += dN[2] - dN[3];
    T2 += dN[3] - dN[4] - dN[5];
    L += dN[4];
    Inc += dN[1];
'

## construct pomp object
SEIT2L_pomp <- pomp(data = FluTdC1971[, c("time", "obs")],
                      skeleton = vectorfield(Csnippet(SEIT2L.skel.c)),
                      rprocess = euler.sim(step.fun = Csnippet(SEIT2L.sim.c),
                                           delta.t = 0.1),
                      rmeasure = Csnippet(SEITL.rmeas.c),
                      dmeasure = Csnippet(SEITL.dmeas.c),
                      toEstimationScale = Csnippet(SEITL.logtrans.c),
                      fromEstimationScale = Csnippet(SEITL.exptrans.c),
                      times = "time",
                      t0 = 1,
                      zeronames = "Inc",
                      paramnames = c("R0", "D_inf", "D_lat", "D_imm", "alpha", "rho"),
                      statenames = c("S", "E", "I", "T1", "T2", "L", "Inc"),
                      obsnames = c("obs"))

#> Error: in ‘pomp’: object 'SEITL.rmeas.c' not found