Burn and thin MCMC chain

Return a burned and thined trace of the chain.

burnAndThin(trace, burn = 0, thin = 0)

Arguments

trace

either a data.frame or a list of data.frame with all variables in column, as outputed by mcmcMh. Accept also an mcmc or mcmc.list object.

burn

proportion of the chain to burn.

thin

number of samples to discard per sample that is being kept

Value

an object with the same format as trace (data.frame or

list of data.frame or mcmc object or mcmc.list

object)

Examples

data(mcmcEpi)
burnAndThin(mcmcEpi1, burn = 100, thin = 4)
#>              R_0 D_inf
#>    [1,] 1.742161     2
#>    [2,] 1.738248     2
#>    [3,] 1.735560     2
#>    [4,] 1.744548     2
#>    [5,] 1.744548     2
#>    [6,] 1.729030     2
#>    [7,] 1.733658     2
#>    [8,] 1.752673     2
#>    [9,] 1.754241     2
#>   [10,] 1.740464     2
#>   [11,] 1.751359     2
#>   [12,] 1.746166     2
#>   [13,] 1.746166     2
#>   [14,] 1.742011     2
#>   [15,] 1.724842     2
#>   [16,] 1.729246     2
#>   [17,] 1.751541     2
#>   [18,] 1.747286     2
#>   [19,] 1.748210     2
#>   [20,] 1.749184     2
#>   [21,] 1.756158     2
#>   [22,] 1.751707     2
#>   [23,] 1.739172     2
#>   [24,] 1.747882     2
#>   [25,] 1.734385     2
#>   [26,] 1.734385     2
#>   [27,] 1.760027     2
#>   [28,] 1.760027     2
#>   [29,] 1.732266     2
#>   [30,] 1.750032     2
#>   [31,] 1.731701     2
#>   [32,] 1.733433     2
#>   [33,] 1.752613     2
#>   [34,] 1.752613     2
#>   [35,] 1.748789     2
#>   [36,] 1.762280     2
#>   [37,] 1.745583     2
#>   [38,] 1.745583     2
#>   [39,] 1.745583     2
#>   [40,] 1.744722     2
#>   [41,] 1.750881     2
#>   [42,] 1.740169     2
#>   [43,] 1.751171     2
#>   [44,] 1.749390     2
#>   [45,] 1.749390     2
#>   [46,] 1.745493     2
#>   [47,] 1.739589     2
#>   [48,] 1.739589     2
#>   [49,] 1.739589     2
#>   [50,] 1.744530     2
#>   [51,] 1.754325     2
#>   [52,] 1.747846     2
#>   [53,] 1.747846     2
#>   [54,] 1.729019     2
#>   [55,] 1.734450     2
#>   [56,] 1.754521     2
#>   [57,] 1.740570     2
#>   [58,] 1.740570     2
#>   [59,] 1.738599     2
#>   [60,] 1.750214     2
#>   [61,] 1.750214     2
#>   [62,] 1.750214     2
#>   [63,] 1.741850     2
#>   [64,] 1.765903     2
#>   [65,] 1.765238     2
#>   [66,] 1.760246     2
#>   [67,] 1.745256     2
#>   [68,] 1.745256     2
#>   [69,] 1.755453     2
#>   [70,] 1.752199     2
#>   [71,] 1.736958     2
#>   [72,] 1.740405     2
#>   [73,] 1.751198     2
#>   [74,] 1.741092     2
#>   [75,] 1.750044     2
#>   [76,] 1.750044     2
#>   [77,] 1.750044     2
#>   [78,] 1.741487     2
#>   [79,] 1.741487     2
#>   [80,] 1.745712     2
#>   [81,] 1.745712     2
#>   [82,] 1.746939     2
#>   [83,] 1.744632     2
#>   [84,] 1.722905     2
#>   [85,] 1.722195     2
#>   [86,] 1.751054     2
#>   [87,] 1.722654     2
#>   [88,] 1.722654     2
#>   [89,] 1.741651     2
#>   [90,] 1.741651     2
#>   [91,] 1.748509     2
#>   [92,] 1.735568     2
#>   [93,] 1.738480     2
#>   [94,] 1.738480     2
#>   [95,] 1.729600     2
#>   [96,] 1.738196     2
#>   [97,] 1.738196     2
#>   [98,] 1.738196     2
#>   [99,] 1.741789     2
#>  [100,] 1.736058     2
#>  [101,] 1.742394     2
#>  [102,] 1.752380     2
#>  [103,] 1.735778     2
#>  [104,] 1.735778     2
#>  [105,] 1.748693     2
#>  [106,] 1.752023     2
#>  [107,] 1.734888     2
#>  [108,] 1.740655     2
#>  [109,] 1.730672     2
#>  [110,] 1.733822     2
#>  [111,] 1.750557     2
#>  [112,] 1.731487     2
#>  [113,] 1.727681     2
#>  [114,] 1.730430     2
#>  [115,] 1.730430     2
#>  [116,] 1.737285     2
#>  [117,] 1.736172     2
#>  [118,] 1.753182     2
#>  [119,] 1.753182     2
#>  [120,] 1.741108     2
#>  [121,] 1.742762     2
#>  [122,] 1.747467     2
#>  [123,] 1.745120     2
#>  [124,] 1.745120     2
#>  [125,] 1.745120     2
#>  [126,] 1.766163     2
#>  [127,] 1.749168     2
#>  [128,] 1.744737     2
#>  [129,] 1.738396     2
#>  [130,] 1.754829     2
#>  [131,] 1.758030     2
#>  [132,] 1.738152     2
#>  [133,] 1.738152     2
#>  [134,] 1.738152     2
#>  [135,] 1.745273     2
#>  [136,] 1.745273     2
#>  [137,] 1.748370     2
#>  [138,] 1.748370     2
#>  [139,] 1.748370     2
#>  [140,] 1.750319     2
#>  [141,] 1.750319     2
#>  [142,] 1.750319     2
#>  [143,] 1.742111     2
#>  [144,] 1.742111     2
#>  [145,] 1.742111     2
#>  [146,] 1.740071     2
#>  [147,] 1.766131     2
#>  [148,] 1.748352     2
#>  [149,] 1.748352     2
#>  [150,] 1.741162     2
#>  [151,] 1.741162     2
#>  [152,] 1.749332     2
#>  [153,] 1.738150     2
#>  [154,] 1.738150     2
#>  [155,] 1.736415     2
#>  [156,] 1.736415     2
#>  [157,] 1.731558     2
#>  [158,] 1.732997     2
#>  [159,] 1.738702     2
#>  [160,] 1.745111     2
#>  [161,] 1.735555     2
#>  [162,] 1.735555     2
#>  [163,] 1.748488     2
#>  [164,] 1.748488     2
#>  [165,] 1.748118     2
#>  [166,] 1.748118     2
#>  [167,] 1.746856     2
#>  [168,] 1.746696     2
#>  [169,] 1.732946     2
#>  [170,] 1.753764     2
#>  [171,] 1.735805     2
#>  [172,] 1.739546     2
#>  [173,] 1.739546     2
#>  [174,] 1.739546     2
#>  [175,] 1.739546     2
#>  [176,] 1.739546     2
#>  [177,] 1.747987     2
#>  [178,] 1.747237     2
#>  [179,] 1.751532     2
#>  [180,] 1.743410     2
#>  [181,] 1.752209     2
#>  [182,] 1.736217     2
#>  [183,] 1.736217     2
#>  [184,] 1.736217     2
#>  [185,] 1.757556     2
#>  [186,] 1.734363     2
#>  [187,] 1.748452     2
#>  [188,] 1.726441     2
#>  [189,] 1.726441     2
#>  [190,] 1.744410     2
#>  [191,] 1.744410     2
#>  [192,] 1.744410     2
#>  [193,] 1.744410     2
#>  [194,] 1.744410     2
#>  [195,] 1.744410     2
#>  [196,] 1.744410     2
#>  [197,] 1.744410     2
#>  [198,] 1.751387     2
#>  [199,] 1.730296     2
#>  [200,] 1.746508     2
#>  [201,] 1.741186     2
#>  [202,] 1.741186     2
#>  [203,] 1.741186     2
#>  [204,] 1.741186     2
#>  [205,] 1.747725     2
#>  [206,] 1.747725     2
#>  [207,] 1.746004     2
#>  [208,] 1.741752     2
#>  [209,] 1.732732     2
#>  [210,] 1.734391     2
#>  [211,] 1.742252     2
#>  [212,] 1.756476     2
#>  [213,] 1.737282     2
#>  [214,] 1.737324     2
#>  [215,] 1.747762     2
#>  [216,] 1.747762     2
#>  [217,] 1.758257     2
#>  [218,] 1.740850     2
#>  [219,] 1.740850     2
#>  [220,] 1.740850     2
#>  [221,] 1.761294     2
#>  [222,] 1.731908     2
#>  [223,] 1.754637     2
#>  [224,] 1.725249     2
#>  [225,] 1.748915     2
#>  [226,] 1.751880     2
#>  [227,] 1.749444     2
#>  [228,] 1.741071     2
#>  [229,] 1.741071     2
#>  [230,] 1.741071     2
#>  [231,] 1.758267     2
#>  [232,] 1.732640     2
#>  [233,] 1.750234     2
#>  [234,] 1.741551     2
#>  [235,] 1.741551     2
#>  [236,] 1.747382     2
#>  [237,] 1.737657     2
#>  [238,] 1.737657     2
#>  [239,] 1.758588     2
#>  [240,] 1.752215     2
#>  [241,] 1.752215     2
#>  [242,] 1.741025     2
#>  [243,] 1.741025     2
#>  [244,] 1.741025     2
#>  [245,] 1.747312     2
#>  [246,] 1.753002     2
#>  [247,] 1.738771     2
#>  [248,] 1.748593     2
#>  [249,] 1.741068     2
#>  [250,] 1.734232     2
#>  [251,] 1.745328     2
#>  [252,] 1.745328     2
#>  [253,] 1.746183     2
#>  [254,] 1.752597     2
#>  [255,] 1.752597     2
#>  [256,] 1.737160     2
#>  [257,] 1.737160     2
#>  [258,] 1.737160     2
#>  [259,] 1.737160     2
#>  [260,] 1.745428     2
#>  [261,] 1.745428     2
#>  [262,] 1.745428     2
#>  [263,] 1.740217     2
#>  [264,] 1.750094     2
#>  [265,] 1.735515     2
#>  [266,] 1.736068     2
#>  [267,] 1.746023     2
#>  [268,] 1.746023     2
#>  [269,] 1.760725     2
#>  [270,] 1.749755     2
#>  [271,] 1.752864     2
#>  [272,] 1.747027     2
#>  [273,] 1.741787     2
#>  [274,] 1.741787     2
#>  [275,] 1.741787     2
#>  [276,] 1.731993     2
#>  [277,] 1.740336     2
#>  [278,] 1.737095     2
#>  [279,] 1.742092     2
#>  [280,] 1.735282     2
#>  [281,] 1.736883     2
#>  [282,] 1.748017     2
#>  [283,] 1.753814     2
#>  [284,] 1.732711     2
#>  [285,] 1.746138     2
#>  [286,] 1.739613     2
#>  [287,] 1.739613     2
#>  [288,] 1.742059     2
#>  [289,] 1.736222     2
#>  [290,] 1.755142     2
#>  [291,] 1.745362     2
#>  [292,] 1.745362     2
#>  [293,] 1.746918     2
#>  [294,] 1.728917     2
#>  [295,] 1.745848     2
#>  [296,] 1.737904     2
#>  [297,] 1.737904     2
#>  [298,] 1.738787     2
#>  [299,] 1.754616     2
#>  [300,] 1.731171     2
#>  [301,] 1.749997     2
#>  [302,] 1.738571     2
#>  [303,] 1.728934     2
#>  [304,] 1.738574     2
#>  [305,] 1.740640     2
#>  [306,] 1.740640     2
#>  [307,] 1.740640     2
#>  [308,] 1.742324     2
#>  [309,] 1.744076     2
#>  [310,] 1.744076     2
#>  [311,] 1.736876     2
#>  [312,] 1.736876     2
#>  [313,] 1.736876     2
#>  [314,] 1.736876     2
#>  [315,] 1.738363     2
#>  [316,] 1.761664     2
#>  [317,] 1.718407     2
#>  [318,] 1.751773     2
#>  [319,] 1.754177     2
#>  [320,] 1.754177     2
#>  [321,] 1.738477     2
#>  [322,] 1.747755     2
#>  [323,] 1.747792     2
#>  [324,] 1.742062     2
#>  [325,] 1.743948     2
#>  [326,] 1.750760     2
#>  [327,] 1.750760     2
#>  [328,] 1.750760     2
#>  [329,] 1.753006     2
#>  [330,] 1.729535     2
#>  [331,] 1.744160     2
#>  [332,] 1.744160     2
#>  [333,] 1.749309     2
#>  [334,] 1.754293     2
#>  [335,] 1.753313     2
#>  [336,] 1.745834     2
#>  [337,] 1.742008     2
#>  [338,] 1.742238     2
#>  [339,] 1.742238     2
#>  [340,] 1.740152     2
#>  [341,] 1.746839     2
#>  [342,] 1.746839     2
#>  [343,] 1.733877     2
#>  [344,] 1.748789     2
#>  [345,] 1.748789     2
#>  [346,] 1.740632     2
#>  [347,] 1.738326     2
#>  [348,] 1.746036     2
#>  [349,] 1.733572     2
#>  [350,] 1.733572     2
#>  [351,] 1.733572     2
#>  [352,] 1.741788     2
#>  [353,] 1.741788     2
#>  [354,] 1.741788     2
#>  [355,] 1.741788     2
#>  [356,] 1.737463     2
#>  [357,] 1.742891     2
#>  [358,] 1.742891     2
#>  [359,] 1.751520     2
#>  [360,] 1.744920     2
#>  [361,] 1.726633     2
#>  [362,] 1.727443     2
#>  [363,] 1.727443     2
#>  [364,] 1.753064     2
#>  [365,] 1.753064     2
#>  [366,] 1.752076     2
#>  [367,] 1.758291     2
#>  [368,] 1.758291     2
#>  [369,] 1.741683     2
#>  [370,] 1.741683     2
#>  [371,] 1.741683     2
#>  [372,] 1.762394     2
#>  [373,] 1.737061     2
#>  [374,] 1.738440     2
#>  [375,] 1.745116     2
#>  [376,] 1.733578     2
#>  [377,] 1.730309     2
#>  [378,] 1.745630     2
#>  [379,] 1.743779     2
#>  [380,] 1.749385     2
#>  [381,] 1.742510     2
#>  [382,] 1.742510     2
#>  [383,] 1.742510     2
#>  [384,] 1.747664     2
#>  [385,] 1.744903     2
#>  [386,] 1.755346     2
#>  [387,] 1.738310     2
#>  [388,] 1.736956     2
#>  [389,] 1.739600     2
#>  [390,] 1.738414     2
#>  [391,] 1.735844     2
#>  [392,] 1.737769     2
#>  [393,] 1.737769     2
#>  [394,] 1.742678     2
#>  [395,] 1.742678     2
#>  [396,] 1.757031     2
#>  [397,] 1.729719     2
#>  [398,] 1.740906     2
#>  [399,] 1.745591     2
#>  [400,] 1.751072     2
#>  [401,] 1.740018     2
#>  [402,] 1.750761     2
#>  [403,] 1.750761     2
#>  [404,] 1.740266     2
#>  [405,] 1.738033     2
#>  [406,] 1.756476     2
#>  [407,] 1.743850     2
#>  [408,] 1.726745     2
#>  [409,] 1.774255     2
#>  [410,] 1.750842     2
#>  [411,] 1.750842     2
#>  [412,] 1.750842     2
#>  [413,] 1.737785     2
#>  [414,] 1.728374     2
#>  [415,] 1.728374     2
#>  [416,] 1.743815     2
#>  [417,] 1.743815     2
#>  [418,] 1.743815     2
#>  [419,] 1.752881     2
#>  [420,] 1.749766     2
#>  [421,] 1.741170     2
#>  [422,] 1.734801     2
#>  [423,] 1.734801     2
#>  [424,] 1.741003     2
#>  [425,] 1.746339     2
#>  [426,] 1.746339     2
#>  [427,] 1.746339     2
#>  [428,] 1.746339     2
#>  [429,] 1.743056     2
#>  [430,] 1.743056     2
#>  [431,] 1.738733     2
#>  [432,] 1.752957     2
#>  [433,] 1.752957     2
#>  [434,] 1.753354     2
#>  [435,] 1.748573     2
#>  [436,] 1.742247     2
#>  [437,] 1.753408     2
#>  [438,] 1.738981     2
#>  [439,] 1.746809     2
#>  [440,] 1.746374     2
#>  [441,] 1.739926     2
#>  [442,] 1.733064     2
#>  [443,] 1.749948     2
#>  [444,] 1.730798     2
#>  [445,] 1.731623     2
#>  [446,] 1.728947     2
#>  [447,] 1.756953     2
#>  [448,] 1.756953     2
#>  [449,] 1.756953     2
#>  [450,] 1.762999     2
#>  [451,] 1.762999     2
#>  [452,] 1.731732     2
#>  [453,] 1.727537     2
#>  [454,] 1.756403     2
#>  [455,] 1.757829     2
#>  [456,] 1.751737     2
#>  [457,] 1.717107     2
#>  [458,] 1.732590     2
#>  [459,] 1.734992     2
#>  [460,] 1.743022     2
#>  [461,] 1.734658     2
#>  [462,] 1.734658     2
#>  [463,] 1.729413     2
#>  [464,] 1.729413     2
#>  [465,] 1.741245     2
#>  [466,] 1.741245     2
#>  [467,] 1.741245     2
#>  [468,] 1.741245     2
#>  [469,] 1.745583     2
#>  [470,] 1.753576     2
#>  [471,] 1.753576     2
#>  [472,] 1.753576     2
#>  [473,] 1.737211     2
#>  [474,] 1.754686     2
#>  [475,] 1.754686     2
#>  [476,] 1.754686     2
#>  [477,] 1.730665     2
#>  [478,] 1.732731     2
#>  [479,] 1.728217     2
#>  [480,] 1.736011     2
#>  [481,] 1.736011     2
#>  [482,] 1.734727     2
#>  [483,] 1.750572     2
#>  [484,] 1.745867     2
#>  [485,] 1.745867     2
#>  [486,] 1.737945     2
#>  [487,] 1.737945     2
#>  [488,] 1.735704     2
#>  [489,] 1.736114     2
#>  [490,] 1.754592     2
#>  [491,] 1.744853     2
#>  [492,] 1.747068     2
#>  [493,] 1.743860     2
#>  [494,] 1.743479     2
#>  [495,] 1.739068     2
#>  [496,] 1.736724     2
#>  [497,] 1.736724     2
#>  [498,] 1.736724     2
#>  [499,] 1.739622     2
#>  [500,] 1.724835     2
#>  [501,] 1.745789     2
#>  [502,] 1.745789     2
#>  [503,] 1.761660     2
#>  [504,] 1.734666     2
#>  [505,] 1.752014     2
#>  [506,] 1.751536     2
#>  [507,] 1.725154     2
#>  [508,] 1.734992     2
#>  [509,] 1.734992     2
#>  [510,] 1.740422     2
#>  [511,] 1.750818     2
#>  [512,] 1.750818     2
#>  [513,] 1.746989     2
#>  [514,] 1.744473     2
#>  [515,] 1.744473     2
#>  [516,] 1.744473     2
#>  [517,] 1.740410     2
#>  [518,] 1.740410     2
#>  [519,] 1.740410     2
#>  [520,] 1.740410     2
#>  [521,] 1.740410     2
#>  [522,] 1.740410     2
#>  [523,] 1.740410     2
#>  [524,] 1.744541     2
#>  [525,] 1.744541     2
#>  [526,] 1.757457     2
#>  [527,] 1.757442     2
#>  [528,] 1.729792     2
#>  [529,] 1.740697     2
#>  [530,] 1.739461     2
#>  [531,] 1.754264     2
#>  [532,] 1.742233     2
#>  [533,] 1.742233     2
#>  [534,] 1.744227     2
#>  [535,] 1.744227     2
#>  [536,] 1.744227     2
#>  [537,] 1.744227     2
#>  [538,] 1.744227     2
#>  [539,] 1.740486     2
#>  [540,] 1.740486     2
#>  [541,] 1.746804     2
#>  [542,] 1.745650     2
#>  [543,] 1.734125     2
#>  [544,] 1.730134     2
#>  [545,] 1.751344     2
#>  [546,] 1.747575     2
#>  [547,] 1.739848     2
#>  [548,] 1.743210     2
#>  [549,] 1.743210     2
#>  [550,] 1.740907     2
#>  [551,] 1.748624     2
#>  [552,] 1.733766     2
#>  [553,] 1.738118     2
#>  [554,] 1.753260     2
#>  [555,] 1.748928     2
#>  [556,] 1.733635     2
#>  [557,] 1.733635     2
#>  [558,] 1.751355     2
#>  [559,] 1.751355     2
#>  [560,] 1.751634     2
#>  [561,] 1.741494     2
#>  [562,] 1.739498     2
#>  [563,] 1.735583     2
#>  [564,] 1.735583     2
#>  [565,] 1.735583     2
#>  [566,] 1.735583     2
#>  [567,] 1.735583     2
#>  [568,] 1.732706     2
#>  [569,] 1.750072     2
#>  [570,] 1.750365     2
#>  [571,] 1.743841     2
#>  [572,] 1.743387     2
#>  [573,] 1.743387     2
#>  [574,] 1.743387     2
#>  [575,] 1.733981     2
#>  [576,] 1.746957     2
#>  [577,] 1.746957     2
#>  [578,] 1.746957     2
#>  [579,] 1.747625     2
#>  [580,] 1.744607     2
#>  [581,] 1.744607     2
#>  [582,] 1.740036     2
#>  [583,] 1.756868     2
#>  [584,] 1.756868     2
#>  [585,] 1.736266     2
#>  [586,] 1.749010     2
#>  [587,] 1.749010     2
#>  [588,] 1.745401     2
#>  [589,] 1.739947     2
#>  [590,] 1.753829     2
#>  [591,] 1.740642     2
#>  [592,] 1.738986     2
#>  [593,] 1.751362     2
#>  [594,] 1.750643     2
#>  [595,] 1.750643     2
#>  [596,] 1.744456     2
#>  [597,] 1.755164     2
#>  [598,] 1.730439     2
#>  [599,] 1.746191     2
#>  [600,] 1.746972     2
#>  [601,] 1.745770     2
#>  [602,] 1.748432     2
#>  [603,] 1.734883     2
#>  [604,] 1.739673     2
#>  [605,] 1.743097     2
#>  [606,] 1.743097     2
#>  [607,] 1.743097     2
#>  [608,] 1.743097     2
#>  [609,] 1.744421     2
#>  [610,] 1.762397     2
#>  [611,] 1.762397     2
#>  [612,] 1.752711     2
#>  [613,] 1.746193     2
#>  [614,] 1.744691     2
#>  [615,] 1.751344     2
#>  [616,] 1.751344     2
#>  [617,] 1.751344     2
#>  [618,] 1.751344     2
#>  [619,] 1.725803     2
#>  [620,] 1.742397     2
#>  [621,] 1.734765     2
#>  [622,] 1.739114     2
#>  [623,] 1.739114     2
#>  [624,] 1.748398     2
#>  [625,] 1.748398     2
#>  [626,] 1.755003     2
#>  [627,] 1.729953     2
#>  [628,] 1.729953     2
#>  [629,] 1.756568     2
#>  [630,] 1.724520     2
#>  [631,] 1.750860     2
#>  [632,] 1.760182     2
#>  [633,] 1.738152     2
#>  [634,] 1.738152     2
#>  [635,] 1.738152     2
#>  [636,] 1.733425     2
#>  [637,] 1.731519     2
#>  [638,] 1.743213     2
#>  [639,] 1.747297     2
#>  [640,] 1.745173     2
#>  [641,] 1.735963     2
#>  [642,] 1.750869     2
#>  [643,] 1.742423     2
#>  [644,] 1.755886     2
#>  [645,] 1.739111     2
#>  [646,] 1.740438     2
#>  [647,] 1.746457     2
#>  [648,] 1.730830     2
#>  [649,] 1.728655     2
#>  [650,] 1.728655     2
#>  [651,] 1.742288     2
#>  [652,] 1.742288     2
#>  [653,] 1.732916     2
#>  [654,] 1.731891     2
#>  [655,] 1.748054     2
#>  [656,] 1.733696     2
#>  [657,] 1.733696     2
#>  [658,] 1.754138     2
#>  [659,] 1.754138     2
#>  [660,] 1.738950     2
#>  [661,] 1.736695     2
#>  [662,] 1.750234     2
#>  [663,] 1.737081     2
#>  [664,] 1.737081     2
#>  [665,] 1.739097     2
#>  [666,] 1.739893     2
#>  [667,] 1.735590     2
#>  [668,] 1.758458     2
#>  [669,] 1.755499     2
#>  [670,] 1.755499     2
#>  [671,] 1.737067     2
#>  [672,] 1.737067     2
#>  [673,] 1.732193     2
#>  [674,] 1.732193     2
#>  [675,] 1.730885     2
#>  [676,] 1.729715     2
#>  [677,] 1.727084     2
#>  [678,] 1.727084     2
#>  [679,] 1.766403     2
#>  [680,] 1.736414     2
#>  [681,] 1.733936     2
#>  [682,] 1.754134     2
#>  [683,] 1.735273     2
#>  [684,] 1.742784     2
#>  [685,] 1.736841     2
#>  [686,] 1.733261     2
#>  [687,] 1.746995     2
#>  [688,] 1.744991     2
#>  [689,] 1.740115     2
#>  [690,] 1.750094     2
#>  [691,] 1.750094     2
#>  [692,] 1.728702     2
#>  [693,] 1.741287     2
#>  [694,] 1.741287     2
#>  [695,] 1.741287     2
#>  [696,] 1.741287     2
#>  [697,] 1.741287     2
#>  [698,] 1.741287     2
#>  [699,] 1.739216     2
#>  [700,] 1.729590     2
#>  [701,] 1.736402     2
#>  [702,] 1.735663     2
#>  [703,] 1.753221     2
#>  [704,] 1.753221     2
#>  [705,] 1.745937     2
#>  [706,] 1.748506     2
#>  [707,] 1.748506     2
#>  [708,] 1.748506     2
#>  [709,] 1.740342     2
#>  [710,] 1.740342     2
#>  [711,] 1.745213     2
#>  [712,] 1.745213     2
#>  [713,] 1.742465     2
#>  [714,] 1.742465     2
#>  [715,] 1.742465     2
#>  [716,] 1.759638     2
#>  [717,] 1.759638     2
#>  [718,] 1.751628     2
#>  [719,] 1.751628     2
#>  [720,] 1.727929     2
#>  [721,] 1.729921     2
#>  [722,] 1.749996     2
#>  [723,] 1.743439     2
#>  [724,] 1.743439     2
#>  [725,] 1.752274     2
#>  [726,] 1.735107     2
#>  [727,] 1.750929     2
#>  [728,] 1.750929     2
#>  [729,] 1.737374     2
#>  [730,] 1.741858     2
#>  [731,] 1.748257     2
#>  [732,] 1.748257     2
#>  [733,] 1.724613     2
#>  [734,] 1.721416     2
#>  [735,] 1.721416     2
#>  [736,] 1.750525     2
#>  [737,] 1.735960     2
#>  [738,] 1.735960     2
#>  [739,] 1.757417     2
#>  [740,] 1.748830     2
#>  [741,] 1.740254     2
#>  [742,] 1.743208     2
#>  [743,] 1.743208     2
#>  [744,] 1.732165     2
#>  [745,] 1.729188     2
#>  [746,] 1.743770     2
#>  [747,] 1.749773     2
#>  [748,] 1.747178     2
#>  [749,] 1.743391     2
#>  [750,] 1.737902     2
#>  [751,] 1.740837     2
#>  [752,] 1.740837     2
#>  [753,] 1.740837     2
#>  [754,] 1.740837     2
#>  [755,] 1.734164     2
#>  [756,] 1.738839     2
#>  [757,] 1.738839     2
#>  [758,] 1.739169     2
#>  [759,] 1.744936     2
#>  [760,] 1.744936     2
#>  [761,] 1.744936     2
#>  [762,] 1.720937     2
#>  [763,] 1.742194     2
#>  [764,] 1.749195     2
#>  [765,] 1.749195     2
#>  [766,] 1.749195     2
#>  [767,] 1.749195     2
#>  [768,] 1.734784     2
#>  [769,] 1.734784     2
#>  [770,] 1.734784     2
#>  [771,] 1.742951     2
#>  [772,] 1.742951     2
#>  [773,] 1.742951     2
#>  [774,] 1.742951     2
#>  [775,] 1.742951     2
#>  [776,] 1.745326     2
#>  [777,] 1.732386     2
#>  [778,] 1.732386     2
#>  [779,] 1.732386     2
#>  [780,] 1.738415     2
#>  [781,] 1.756563     2
#>  [782,] 1.741663     2
#>  [783,] 1.739732     2
#>  [784,] 1.739732     2
#>  [785,] 1.739732     2
#>  [786,] 1.739732     2
#>  [787,] 1.742540     2
#>  [788,] 1.742540     2
#>  [789,] 1.743683     2
#>  [790,] 1.743683     2
#>  [791,] 1.743683     2
#>  [792,] 1.739859     2
#>  [793,] 1.739859     2
#>  [794,] 1.739859     2
#>  [795,] 1.741213     2
#>  [796,] 1.739140     2
#>  [797,] 1.738961     2
#>  [798,] 1.738961     2
#>  [799,] 1.754733     2
#>  [800,] 1.741255     2
#>  [801,] 1.756613     2
#>  [802,] 1.756613     2
#>  [803,] 1.743285     2
#>  [804,] 1.734626     2
#>  [805,] 1.744331     2
#>  [806,] 1.740091     2
#>  [807,] 1.744540     2
#>  [808,] 1.754002     2
#>  [809,] 1.744674     2
#>  [810,] 1.737116     2
#>  [811,] 1.737977     2
#>  [812,] 1.751036     2
#>  [813,] 1.743710     2
#>  [814,] 1.729028     2
#>  [815,] 1.726905     2
#>  [816,] 1.760903     2
#>  [817,] 1.741208     2
#>  [818,] 1.741954     2
#>  [819,] 1.734364     2
#>  [820,] 1.734364     2
#>  [821,] 1.737775     2
#>  [822,] 1.735396     2
#>  [823,] 1.745850     2
#>  [824,] 1.751190     2
#>  [825,] 1.745207     2
#>  [826,] 1.744010     2
#>  [827,] 1.744010     2
#>  [828,] 1.744043     2
#>  [829,] 1.744043     2
#>  [830,] 1.744043     2
#>  [831,] 1.744043     2
#>  [832,] 1.744043     2
#>  [833,] 1.744043     2
#>  [834,] 1.760091     2
#>  [835,] 1.750192     2
#>  [836,] 1.750192     2
#>  [837,] 1.756699     2
#>  [838,] 1.756699     2
#>  [839,] 1.746105     2
#>  [840,] 1.751738     2
#>  [841,] 1.748182     2
#>  [842,] 1.745117     2
#>  [843,] 1.745117     2
#>  [844,] 1.745117     2
#>  [845,] 1.754255     2
#>  [846,] 1.750318     2
#>  [847,] 1.752750     2
#>  [848,] 1.742649     2
#>  [849,] 1.741437     2
#>  [850,] 1.732169     2
#>  [851,] 1.732169     2
#>  [852,] 1.739818     2
#>  [853,] 1.739818     2
#>  [854,] 1.752822     2
#>  [855,] 1.742041     2
#>  [856,] 1.743329     2
#>  [857,] 1.746835     2
#>  [858,] 1.746835     2
#>  [859,] 1.746835     2
#>  [860,] 1.746835     2
#>  [861,] 1.746835     2
#>  [862,] 1.746835     2
#>  [863,] 1.737950     2
#>  [864,] 1.737950     2
#>  [865,] 1.737950     2
#>  [866,] 1.737950     2
#>  [867,] 1.747800     2
#>  [868,] 1.747800     2
#>  [869,] 1.747800     2
#>  [870,] 1.737920     2
#>  [871,] 1.737920     2
#>  [872,] 1.737920     2
#>  [873,] 1.737920     2
#>  [874,] 1.748851     2
#>  [875,] 1.737430     2
#>  [876,] 1.744818     2
#>  [877,] 1.721231     2
#>  [878,] 1.734706     2
#>  [879,] 1.734706     2
#>  [880,] 1.734706     2
#>  [881,] 1.744455     2
#>  [882,] 1.748757     2
#>  [883,] 1.738065     2
#>  [884,] 1.742830     2
#>  [885,] 1.725385     2
#>  [886,] 1.725385     2
#>  [887,] 1.732276     2
#>  [888,] 1.737427     2
#>  [889,] 1.737427     2
#>  [890,] 1.738088     2
#>  [891,] 1.743805     2
#>  [892,] 1.743805     2
#>  [893,] 1.743805     2
#>  [894,] 1.743805     2
#>  [895,] 1.762812     2
#>  [896,] 1.724860     2
#>  [897,] 1.762574     2
#>  [898,] 1.762574     2
#>  [899,] 1.737484     2
#>  [900,] 1.737484     2
#>  [901,] 1.741885     2
#>  [902,] 1.741885     2
#>  [903,] 1.741885     2
#>  [904,] 1.741885     2
#>  [905,] 1.753869     2
#>  [906,] 1.753869     2
#>  [907,] 1.753869     2
#>  [908,] 1.753869     2
#>  [909,] 1.721779     2
#>  [910,] 1.745376     2
#>  [911,] 1.745121     2
#>  [912,] 1.730877     2
#>  [913,] 1.743085     2
#>  [914,] 1.743085     2
#>  [915,] 1.743085     2
#>  [916,] 1.743085     2
#>  [917,] 1.743085     2
#>  [918,] 1.743085     2
#>  [919,] 1.743085     2
#>  [920,] 1.744461     2
#>  [921,] 1.742717     2
#>  [922,] 1.759873     2
#>  [923,] 1.759873     2
#>  [924,] 1.756047     2
#>  [925,] 1.734293     2
#>  [926,] 1.734293     2
#>  [927,] 1.739043     2
#>  [928,] 1.752842     2
#>  [929,] 1.736195     2
#>  [930,] 1.750779     2
#>  [931,] 1.750779     2
#>  [932,] 1.750779     2
#>  [933,] 1.736384     2
#>  [934,] 1.744602     2
#>  [935,] 1.740127     2
#>  [936,] 1.736213     2
#>  [937,] 1.750410     2
#>  [938,] 1.750410     2
#>  [939,] 1.739593     2
#>  [940,] 1.754295     2
#>  [941,] 1.754295     2
#>  [942,] 1.737430     2
#>  [943,] 1.737430     2
#>  [944,] 1.737430     2
#>  [945,] 1.729448     2
#>  [946,] 1.757537     2
#>  [947,] 1.757537     2
#>  [948,] 1.736873     2
#>  [949,] 1.728796     2
#>  [950,] 1.742931     2
#>  [951,] 1.742931     2
#>  [952,] 1.732875     2
#>  [953,] 1.741043     2
#>  [954,] 1.730094     2
#>  [955,] 1.738887     2
#>  [956,] 1.737819     2
#>  [957,] 1.746324     2
#>  [958,] 1.752466     2
#>  [959,] 1.731762     2
#>  [960,] 1.746141     2
#>  [961,] 1.746141     2
#>  [962,] 1.743397     2
#>  [963,] 1.728426     2
#>  [964,] 1.728426     2
#>  [965,] 1.757914     2
#>  [966,] 1.723543     2
#>  [967,] 1.727955     2
#>  [968,] 1.727955     2
#>  [969,] 1.727955     2
#>  [970,] 1.749085     2
#>  [971,] 1.749085     2
#>  [972,] 1.749085     2
#>  [973,] 1.749085     2
#>  [974,] 1.749085     2
#>  [975,] 1.749085     2
#>  [976,] 1.755530     2
#>  [977,] 1.742413     2
#>  [978,] 1.738967     2
#>  [979,] 1.743098     2
#>  [980,] 1.748795     2
#>  [981,] 1.748795     2
#>  [982,] 1.748795     2
#>  [983,] 1.748795     2
#>  [984,] 1.750947     2
#>  [985,] 1.750947     2
#>  [986,] 1.741803     2
#>  [987,] 1.728854     2
#>  [988,] 1.732961     2
#>  [989,] 1.732961     2
#>  [990,] 1.746728     2
#>  [991,] 1.753462     2
#>  [992,] 1.756821     2
#>  [993,] 1.735507     2
#>  [994,] 1.735507     2
#>  [995,] 1.744816     2
#>  [996,] 1.744816     2
#>  [997,] 1.752637     2
#>  [998,] 1.750301     2
#>  [999,] 1.744026     2
#> [1000,] 1.740511     2
#> [1001,] 1.740511     2
#> [1002,] 1.735024     2
#> [1003,] 1.737191     2
#> [1004,] 1.737191     2
#> [1005,] 1.736253     2
#> [1006,] 1.750262     2
#> [1007,] 1.753664     2
#> [1008,] 1.737251     2
#> [1009,] 1.744500     2
#> [1010,] 1.732717     2
#> [1011,] 1.738041     2
#> [1012,] 1.745131     2
#> [1013,] 1.749854     2
#> [1014,] 1.749080     2
#> [1015,] 1.750233     2
#> [1016,] 1.744010     2
#> [1017,] 1.744010     2
#> [1018,] 1.740239     2
#> [1019,] 1.732542     2
#> [1020,] 1.737276     2
#> [1021,] 1.737672     2
#> [1022,] 1.737672     2
#> [1023,] 1.737672     2
#> [1024,] 1.737672     2
#> [1025,] 1.729928     2
#> [1026,] 1.743122     2
#> [1027,] 1.743122     2
#> [1028,] 1.743122     2
#> [1029,] 1.743122     2
#> [1030,] 1.743122     2
#> [1031,] 1.744401     2
#> [1032,] 1.751704     2
#> [1033,] 1.753077     2
#> [1034,] 1.753077     2
#> [1035,] 1.753077     2
#> [1036,] 1.750282     2
#> [1037,] 1.737072     2
#> [1038,] 1.737072     2
#> [1039,] 1.737072     2
#> [1040,] 1.745075     2
#> [1041,] 1.745075     2
#> [1042,] 1.738018     2
#> [1043,] 1.729775     2
#> [1044,] 1.747351     2
#> [1045,] 1.747351     2
#> [1046,] 1.748597     2
#> [1047,] 1.738365     2
#> [1048,] 1.738365     2
#> [1049,] 1.757392     2
#> [1050,] 1.757392     2
#> [1051,] 1.735199     2
#> [1052,] 1.741223     2
#> [1053,] 1.751526     2
#> [1054,] 1.746333     2
#> [1055,] 1.745250     2
#> [1056,] 1.757437     2
#> [1057,] 1.737896     2
#> [1058,] 1.745781     2
#> [1059,] 1.745781     2
#> [1060,] 1.752375     2
#> [1061,] 1.748794     2
#> [1062,] 1.741550     2
#> [1063,] 1.741550     2
#> [1064,] 1.741550     2
#> [1065,] 1.742379     2
#> [1066,] 1.742379     2
#> [1067,] 1.739242     2
#> [1068,] 1.754325     2
#> [1069,] 1.754325     2
#> [1070,] 1.754325     2
#> [1071,] 1.743070     2
#> [1072,] 1.739562     2
#> [1073,] 1.734343     2
#> [1074,] 1.734343     2
#> [1075,] 1.733358     2
#> [1076,] 1.733358     2
#> [1077,] 1.743329     2
#> [1078,] 1.745137     2
#> [1079,] 1.745137     2
#> [1080,] 1.737793     2
#> [1081,] 1.746396     2
#> [1082,] 1.735549     2
#> [1083,] 1.733135     2
#> [1084,] 1.733135     2
#> [1085,] 1.747095     2
#> [1086,] 1.740901     2
#> [1087,] 1.761825     2
#> [1088,] 1.757100     2
#> [1089,] 1.747710     2
#> [1090,] 1.752332     2
#> [1091,] 1.749143     2
#> [1092,] 1.749143     2
#> [1093,] 1.749143     2
#> [1094,] 1.749143     2
#> [1095,] 1.738166     2
#> [1096,] 1.746305     2
#> [1097,] 1.758729     2
#> [1098,] 1.728449     2
#> [1099,] 1.758713     2
#> [1100,] 1.732365     2
#> [1101,] 1.757832     2
#> [1102,] 1.746693     2
#> [1103,] 1.740051     2
#> [1104,] 1.731280     2
#> [1105,] 1.732849     2
#> [1106,] 1.747141     2
#> [1107,] 1.721956     2
#> [1108,] 1.721956     2
#> [1109,] 1.746114     2
#> [1110,] 1.748309     2
#> [1111,] 1.738464     2
#> [1112,] 1.740612     2
#> [1113,] 1.748959     2
#> [1114,] 1.748959     2
#> [1115,] 1.748959     2
#> [1116,] 1.748959     2
#> [1117,] 1.751361     2
#> [1118,] 1.734127     2
#> [1119,] 1.755686     2
#> [1120,] 1.752635     2
#> [1121,] 1.752635     2
#> [1122,] 1.744280     2
#> [1123,] 1.744280     2
#> [1124,] 1.744280     2
#> [1125,] 1.737854     2
#> [1126,] 1.737854     2
#> [1127,] 1.736739     2
#> [1128,] 1.732558     2
#> [1129,] 1.732558     2
#> [1130,] 1.740737     2
#> [1131,] 1.738347     2
#> [1132,] 1.738347     2
#> [1133,] 1.749454     2
#> [1134,] 1.755609     2
#> [1135,] 1.734085     2
#> [1136,] 1.747963     2
#> [1137,] 1.747963     2
#> [1138,] 1.741012     2
#> [1139,] 1.741012     2
#> [1140,] 1.741012     2
#> [1141,] 1.742632     2
#> [1142,] 1.742632     2
#> [1143,] 1.742632     2
#> [1144,] 1.747543     2
#> [1145,] 1.722823     2
#> [1146,] 1.722823     2
#> [1147,] 1.739852     2
#> [1148,] 1.741464     2
#> [1149,] 1.740160     2
#> [1150,] 1.740447     2
#> [1151,] 1.736912     2
#> [1152,] 1.730733     2
#> [1153,] 1.761265     2
#> [1154,] 1.734103     2
#> [1155,] 1.734103     2
#> [1156,] 1.751101     2
#> [1157,] 1.749378     2
#> [1158,] 1.749378     2
#> [1159,] 1.745185     2
#> [1160,] 1.737241     2
#> [1161,] 1.737241     2
#> [1162,] 1.756853     2
#> [1163,] 1.756853     2
#> [1164,] 1.733660     2
#> [1165,] 1.736084     2
#> [1166,] 1.742414     2
#> [1167,] 1.742414     2
#> [1168,] 1.747721     2
#> [1169,] 1.746932     2
#> [1170,] 1.746932     2
#> [1171,] 1.743411     2
#> [1172,] 1.743411     2
#> [1173,] 1.743411     2
#> [1174,] 1.740544     2
#> [1175,] 1.741871     2
#> [1176,] 1.738749     2
#> [1177,] 1.752512     2
#> [1178,] 1.748451     2
#> [1179,] 1.736290     2
#> [1180,] 1.746118     2
#> [1181,] 1.746118     2
#> [1182,] 1.739205     2
#> [1183,] 1.746655     2
#> [1184,] 1.740251     2
#> [1185,] 1.745038     2
#> [1186,] 1.745038     2
#> [1187,] 1.752433     2
#> [1188,] 1.752433     2
#> [1189,] 1.756922     2
#> [1190,] 1.749829     2
#> [1191,] 1.734029     2
#> [1192,] 1.760127     2
#> [1193,] 1.760127     2
#> [1194,] 1.744869     2
#> [1195,] 1.744869     2
#> [1196,] 1.729053     2
#> [1197,] 1.739108     2
#> [1198,] 1.739108     2
#> [1199,] 1.739108     2
#> [1200,] 1.752653     2
#> [1201,] 1.721601     2
#> [1202,] 1.731687     2
#> [1203,] 1.744632     2
#> [1204,] 1.761373     2
#> [1205,] 1.753357     2
#> [1206,] 1.725281     2
#> [1207,] 1.751425     2
#> [1208,] 1.754102     2
#> [1209,] 1.754102     2
#> [1210,] 1.752905     2
#> [1211,] 1.718919     2
#> [1212,] 1.746058     2
#> [1213,] 1.746058     2
#> [1214,] 1.746058     2
#> [1215,] 1.740125     2
#> [1216,] 1.740125     2
#> [1217,] 1.743287     2
#> [1218,] 1.743287     2
#> [1219,] 1.750593     2
#> [1220,] 1.750593     2
#> [1221,] 1.754318     2
#> [1222,] 1.754318     2
#> [1223,] 1.754318     2
#> [1224,] 1.747328     2
#> [1225,] 1.748842     2
#> [1226,] 1.755173     2
#> [1227,] 1.749408     2
#> [1228,] 1.749729     2
#> [1229,] 1.749729     2
#> [1230,] 1.751783     2
#> [1231,] 1.749026     2
#> [1232,] 1.738152     2
#> [1233,] 1.737916     2
#> [1234,] 1.737916     2
#> [1235,] 1.742703     2
#> [1236,] 1.736179     2
#> [1237,] 1.744333     2
#> [1238,] 1.744333     2
#> [1239,] 1.750828     2
#> [1240,] 1.749357     2
#> [1241,] 1.742169     2
#> [1242,] 1.742169     2
#> [1243,] 1.732418     2
#> [1244,] 1.733281     2
#> [1245,] 1.733281     2
#> [1246,] 1.754923     2
#> [1247,] 1.739545     2
#> [1248,] 1.739545     2
#> [1249,] 1.728904     2
#> [1250,] 1.728904     2
#> [1251,] 1.729646     2
#> [1252,] 1.739141     2
#> [1253,] 1.739141     2
#> [1254,] 1.739141     2
#> [1255,] 1.744946     2
#> [1256,] 1.756211     2
#> [1257,] 1.756211     2
#> [1258,] 1.742656     2
#> [1259,] 1.741186     2
#> [1260,] 1.737191     2
#> [1261,] 1.737191     2
#> [1262,] 1.745182     2
#> [1263,] 1.745182     2
#> [1264,] 1.736812     2
#> [1265,] 1.736812     2
#> [1266,] 1.736812     2
#> [1267,] 1.736623     2
#> [1268,] 1.736623     2
#> [1269,] 1.731196     2
#> [1270,] 1.731196     2
#> [1271,] 1.753120     2
#> [1272,] 1.734806     2
#> [1273,] 1.746794     2
#> [1274,] 1.743929     2
#> [1275,] 1.730103     2
#> [1276,] 1.757131     2
#> [1277,] 1.757131     2
#> [1278,] 1.761457     2
#> [1279,] 1.757300     2
#> [1280,] 1.757300     2
#> [1281,] 1.746784     2
#> [1282,] 1.746784     2
#> [1283,] 1.746784     2
#> [1284,] 1.741802     2
#> [1285,] 1.739434     2
#> [1286,] 1.739875     2
#> [1287,] 1.744785     2
#> [1288,] 1.744785     2
#> [1289,] 1.750579     2
#> [1290,] 1.750579     2
#> [1291,] 1.748485     2
#> [1292,] 1.748485     2
#> [1293,] 1.748485     2
#> [1294,] 1.739204     2
#> [1295,] 1.743904     2
#> [1296,] 1.743904     2
#> [1297,] 1.743904     2
#> [1298,] 1.748656     2
#> [1299,] 1.739570     2
#> [1300,] 1.750783     2
#> [1301,] 1.750783     2
#> [1302,] 1.750783     2
#> [1303,] 1.729501     2
#> [1304,] 1.728761     2
#> [1305,] 1.728761     2
#> [1306,] 1.715186     2
#> [1307,] 1.761301     2
#> [1308,] 1.736416     2
#> [1309,] 1.740663     2
#> [1310,] 1.740663     2
#> [1311,] 1.740111     2
#> [1312,] 1.734366     2
#> [1313,] 1.730690     2
#> [1314,] 1.747357     2
#> [1315,] 1.747357     2
#> [1316,] 1.747357     2
#> [1317,] 1.747357     2
#> [1318,] 1.747357     2
#> [1319,] 1.727806     2
#> [1320,] 1.738313     2
#> [1321,] 1.740435     2
#> [1322,] 1.731993     2
#> [1323,] 1.731993     2
#> [1324,] 1.723601     2
#> [1325,] 1.748974     2
#> [1326,] 1.749245     2
#> [1327,] 1.749245     2
#> [1328,] 1.735552     2
#> [1329,] 1.742715     2
#> [1330,] 1.751408     2
#> [1331,] 1.757882     2
#> [1332,] 1.730753     2
#> [1333,] 1.736319     2
#> [1334,] 1.738432     2
#> [1335,] 1.757884     2
#> [1336,] 1.757884     2
#> [1337,] 1.754835     2
#> [1338,] 1.729385     2
#> [1339,] 1.733961     2
#> [1340,] 1.733961     2
#> [1341,] 1.755884     2
#> [1342,] 1.745941     2
#> [1343,] 1.749018     2
#> [1344,] 1.750269     2
#> [1345,] 1.750703     2
#> [1346,] 1.734394     2
#> [1347,] 1.749812     2
#> [1348,] 1.738187     2
#> [1349,] 1.752450     2
#> [1350,] 1.740409     2
#> [1351,] 1.740409     2
#> [1352,] 1.740409     2
#> [1353,] 1.740409     2
#> [1354,] 1.739234     2
#> [1355,] 1.746726     2
#> [1356,] 1.749886     2
#> [1357,] 1.736622     2
#> [1358,] 1.736622     2
#> [1359,] 1.743257     2
#> [1360,] 1.741381     2
#> [1361,] 1.740991     2
#> [1362,] 1.745988     2
#> [1363,] 1.745988     2
#> [1364,] 1.742732     2
#> [1365,] 1.742732     2
#> [1366,] 1.751243     2
#> [1367,] 1.751243     2
#> [1368,] 1.728690     2
#> [1369,] 1.743759     2
#> [1370,] 1.739139     2
#> [1371,] 1.740252     2
#> [1372,] 1.736294     2
#> [1373,] 1.746526     2
#> [1374,] 1.746526     2
#> [1375,] 1.746526     2
#> [1376,] 1.747864     2
#> [1377,] 1.725013     2
#> [1378,] 1.742906     2
#> [1379,] 1.742906     2
#> [1380,] 1.728198     2
#> [1381,] 1.740370     2
#> [1382,] 1.740370     2
#> [1383,] 1.740370     2
#> [1384,] 1.736855     2
#> [1385,] 1.735393     2
#> [1386,] 1.741436     2
#> [1387,] 1.733496     2
#> [1388,] 1.738061     2
#> [1389,] 1.740888     2
#> [1390,] 1.751431     2
#> [1391,] 1.751431     2
#> [1392,] 1.750716     2
#> [1393,] 1.750716     2
#> [1394,] 1.747164     2
#> [1395,] 1.747164     2
#> [1396,] 1.736364     2
#> [1397,] 1.729562     2
#> [1398,] 1.740829     2
#> [1399,] 1.740829     2
#> [1400,] 1.740829     2
#> [1401,] 1.738594     2
#> [1402,] 1.741273     2
#> [1403,] 1.755059     2
#> [1404,] 1.755059     2
#> [1405,] 1.755059     2
#> [1406,] 1.740426     2
#> [1407,] 1.751820     2
#> [1408,] 1.730255     2
#> [1409,] 1.738502     2
#> [1410,] 1.739484     2
#> [1411,] 1.746706     2
#> [1412,] 1.749660     2
#> [1413,] 1.749660     2
#> [1414,] 1.739983     2
#> [1415,] 1.739983     2
#> [1416,] 1.739983     2
#> [1417,] 1.739983     2
#> [1418,] 1.734889     2
#> [1419,] 1.758902     2
#> [1420,] 1.729388     2
#> [1421,] 1.729388     2
#> [1422,] 1.744428     2
#> [1423,] 1.736501     2
#> [1424,] 1.736501     2
#> [1425,] 1.744116     2
#> [1426,] 1.744116     2
#> [1427,] 1.736555     2
#> [1428,] 1.745176     2
#> [1429,] 1.745176     2
#> [1430,] 1.745176     2
#> [1431,] 1.745176     2
#> [1432,] 1.745176     2
#> [1433,] 1.749036     2
#> [1434,] 1.731693     2
#> [1435,] 1.744685     2
#> [1436,] 1.747837     2
#> [1437,] 1.747837     2
#> [1438,] 1.747837     2
#> [1439,] 1.747837     2
#> [1440,] 1.747837     2
#> [1441,] 1.746000     2
#> [1442,] 1.746889     2
#> [1443,] 1.747803     2
#> [1444,] 1.747803     2
#> [1445,] 1.743141     2
#> [1446,] 1.743141     2
#> [1447,] 1.743141     2
#> [1448,] 1.737854     2
#> [1449,] 1.736788     2
#> [1450,] 1.754833     2
#> [1451,] 1.749249     2
#> [1452,] 1.744753     2
#> [1453,] 1.765007     2
#> [1454,] 1.756465     2
#> [1455,] 1.761359     2
#> [1456,] 1.745344     2
#> [1457,] 1.744451     2
#> [1458,] 1.757696     2
#> [1459,] 1.763153     2
#> [1460,] 1.765039     2
#> [1461,] 1.748441     2
#> [1462,] 1.746170     2
#> [1463,] 1.733518     2
#> [1464,] 1.733518     2
#> [1465,] 1.739295     2
#> [1466,] 1.739295     2
#> [1467,] 1.739295     2
#> [1468,] 1.753155     2
#> [1469,] 1.753155     2
#> [1470,] 1.737076     2
#> [1471,] 1.737076     2
#> [1472,] 1.762077     2
#> [1473,] 1.748539     2
#> [1474,] 1.748539     2
#> [1475,] 1.748539     2
#> [1476,] 1.748539     2
#> [1477,] 1.754699     2
#> [1478,] 1.736864     2
#> [1479,] 1.750253     2
#> [1480,] 1.760932     2
#> [1481,] 1.752761     2
#> [1482,] 1.752761     2
#> [1483,] 1.752761     2
#> [1484,] 1.752761     2
#> [1485,] 1.744454     2
#> [1486,] 1.750538     2
#> [1487,] 1.747479     2
#> [1488,] 1.747479     2
#> [1489,] 1.735668     2
#> [1490,] 1.737154     2
#> [1491,] 1.750531     2
#> [1492,] 1.749744     2
#> [1493,] 1.749523     2
#> [1494,] 1.747504     2
#> [1495,] 1.747504     2
#> [1496,] 1.744452     2
#> [1497,] 1.742969     2
#> [1498,] 1.742969     2
#> [1499,] 1.740303     2
#> [1500,] 1.749826     2
#> [1501,] 1.753319     2
#> [1502,] 1.741984     2
#> [1503,] 1.741984     2
#> [1504,] 1.730650     2
#> [1505,] 1.751864     2
#> [1506,] 1.749856     2
#> [1507,] 1.744399     2
#> [1508,] 1.744399     2
#> [1509,] 1.739159     2
#> [1510,] 1.739159     2
#> [1511,] 1.739159     2
#> [1512,] 1.755061     2
#> [1513,] 1.731336     2
#> [1514,] 1.750233     2
#> [1515,] 1.755563     2
#> [1516,] 1.745669     2
#> [1517,] 1.745398     2
#> [1518,] 1.745398     2
#> [1519,] 1.745398     2
#> [1520,] 1.740024     2
#> [1521,] 1.748393     2
#> [1522,] 1.739061     2
#> [1523,] 1.750279     2
#> [1524,] 1.751105     2
#> [1525,] 1.751105     2
#> [1526,] 1.752279     2
#> [1527,] 1.751478     2
#> [1528,] 1.743402     2
#> [1529,] 1.739139     2
#> [1530,] 1.739139     2
#> [1531,] 1.749537     2
#> [1532,] 1.733694     2
#> [1533,] 1.733694     2
#> [1534,] 1.753440     2
#> [1535,] 1.745514     2
#> [1536,] 1.745514     2
#> [1537,] 1.754864     2
#> [1538,] 1.734491     2
#> [1539,] 1.743390     2
#> [1540,] 1.743390     2
#> [1541,] 1.743390     2
#> [1542,] 1.742008     2
#> [1543,] 1.742008     2
#> [1544,] 1.740596     2
#> [1545,] 1.750800     2
#> [1546,] 1.750800     2
#> [1547,] 1.750800     2
#> [1548,] 1.750800     2
#> [1549,] 1.743196     2
#> [1550,] 1.741072     2
#> [1551,] 1.751340     2
#> [1552,] 1.751340     2
#> [1553,] 1.751340     2
#> [1554,] 1.750470     2
#> [1555,] 1.735135     2
#> [1556,] 1.737329     2
#> [1557,] 1.748633     2
#> [1558,] 1.746803     2
#> [1559,] 1.725899     2
#> [1560,] 1.745390     2
#> [1561,] 1.759037     2
#> [1562,] 1.759037     2
#> [1563,] 1.741220     2
#> [1564,] 1.739975     2
#> [1565,] 1.739975     2
#> [1566,] 1.741601     2
#> [1567,] 1.751681     2
#> [1568,] 1.754960     2
#> [1569,] 1.754960     2
#> [1570,] 1.754960     2
#> [1571,] 1.734293     2
#> [1572,] 1.740113     2
#> [1573,] 1.738999     2
#> [1574,] 1.738999     2
#> [1575,] 1.746041     2
#> [1576,] 1.758546     2
#> [1577,] 1.756369     2
#> [1578,] 1.756369     2
#> [1579,] 1.754205     2
#> [1580,] 1.739848     2
#> [1581,] 1.739848     2
#> [1582,] 1.739841     2
#> [1583,] 1.743948     2
#> [1584,] 1.751886     2
#> [1585,] 1.758836     2
#> [1586,] 1.740251     2
#> [1587,] 1.740251     2
#> [1588,] 1.738871     2
#> [1589,] 1.741077     2
#> [1590,] 1.743969     2
#> [1591,] 1.743969     2
#> [1592,] 1.722710     2
#> [1593,] 1.756764     2
#> [1594,] 1.756764     2
#> [1595,] 1.745173     2
#> [1596,] 1.745173     2
#> [1597,] 1.747139     2
#> [1598,] 1.747340     2
#> [1599,] 1.729368     2
#> [1600,] 1.742999     2
#> [1601,] 1.743687     2
#> [1602,] 1.743687     2
#> [1603,] 1.743687     2
#> [1604,] 1.743687     2
#> [1605,] 1.742152     2
#> [1606,] 1.742152     2
#> [1607,] 1.742152     2
#> [1608,] 1.742248     2
#> [1609,] 1.747448     2
#> [1610,] 1.744767     2
#> [1611,] 1.743696     2
#> [1612,] 1.743696     2
#> [1613,] 1.743696     2
#> [1614,] 1.743696     2
#> [1615,] 1.732972     2
#> [1616,] 1.732972     2
#> [1617,] 1.734022     2
#> [1618,] 1.758089     2
#> [1619,] 1.752582     2
#> [1620,] 1.751498     2
#> [1621,] 1.740442     2
#> [1622,] 1.750319     2
#> [1623,] 1.762262     2
#> [1624,] 1.735355     2
#> [1625,] 1.763177     2
#> [1626,] 1.741920     2
#> [1627,] 1.741920     2
#> [1628,] 1.731082     2
#> [1629,] 1.743240     2
#> [1630,] 1.745992     2
#> [1631,] 1.752270     2
#> [1632,] 1.750501     2
#> [1633,] 1.745290     2
#> [1634,] 1.745290     2
#> [1635,] 1.745290     2
#> [1636,] 1.745290     2
#> [1637,] 1.748413     2
#> [1638,] 1.745637     2
#> [1639,] 1.738565     2
#> [1640,] 1.738565     2
#> [1641,] 1.753125     2
#> [1642,] 1.753125     2
#> [1643,] 1.735508     2
#> [1644,] 1.748438     2
#> [1645,] 1.743204     2
#> [1646,] 1.743204     2
#> [1647,] 1.743204     2
#> [1648,] 1.737160     2
#> [1649,] 1.737160     2
#> [1650,] 1.748278     2
#> [1651,] 1.748278     2
#> [1652,] 1.748278     2
#> [1653,] 1.766391     2
#> [1654,] 1.732728     2
#> [1655,] 1.747939     2
#> [1656,] 1.737048     2
#> [1657,] 1.746081     2
#> [1658,] 1.737493     2
#> [1659,] 1.749958     2
#> [1660,] 1.741327     2
#> [1661,] 1.741327     2
#> [1662,] 1.747041     2
#> [1663,] 1.732572     2
#> [1664,] 1.752363     2
#> [1665,] 1.751326     2
#> [1666,] 1.756544     2
#> [1667,] 1.744236     2
#> [1668,] 1.744316     2
#> [1669,] 1.730654     2
#> [1670,] 1.723927     2
#> [1671,] 1.739427     2
#> [1672,] 1.739427     2
#> [1673,] 1.744121     2
#> [1674,] 1.739762     2
#> [1675,] 1.739762     2
#> [1676,] 1.732421     2
#> [1677,] 1.745639     2
#> [1678,] 1.750178     2
#> [1679,] 1.750178     2
#> [1680,] 1.751658     2
#> [1681,] 1.731843     2
#> [1682,] 1.731843     2
#> [1683,] 1.735670     2
#> [1684,] 1.748571     2
#> [1685,] 1.730608     2
#> [1686,] 1.752089     2
#> [1687,] 1.740314     2
#> [1688,] 1.740314     2
#> [1689,] 1.741854     2
#> [1690,] 1.736858     2
#> [1691,] 1.748607     2
#> [1692,] 1.748607     2
#> [1693,] 1.748607     2
#> [1694,] 1.728877     2
#> [1695,] 1.747929     2
#> [1696,] 1.742972     2
#> [1697,] 1.742972     2
#> [1698,] 1.742972     2
#> [1699,] 1.745599     2
#> [1700,] 1.745599     2
#> [1701,] 1.754850     2
#> [1702,] 1.737661     2
#> [1703,] 1.742664     2
#> [1704,] 1.748988     2
#> [1705,] 1.748988     2
#> [1706,] 1.741713     2
#> [1707,] 1.739503     2
#> [1708,] 1.739503     2
#> [1709,] 1.739503     2
#> [1710,] 1.751061     2
#> [1711,] 1.741917     2
#> [1712,] 1.735750     2
#> [1713,] 1.735750     2
#> [1714,] 1.740882     2
#> [1715,] 1.745599     2
#> [1716,] 1.745599     2
#> [1717,] 1.745599     2
#> [1718,] 1.745599     2
#> [1719,] 1.742962     2
#> [1720,] 1.737546     2
#> [1721,] 1.730052     2
#> [1722,] 1.759721     2
#> [1723,] 1.757714     2
#> [1724,] 1.748035     2
#> [1725,] 1.752000     2
#> [1726,] 1.752000     2
#> [1727,] 1.752000     2
#> [1728,] 1.756624     2
#> [1729,] 1.754557     2
#> [1730,] 1.754557     2
#> [1731,] 1.754557     2
#> [1732,] 1.750429     2
#> [1733,] 1.745308     2
#> [1734,] 1.735884     2
#> [1735,] 1.737072     2
#> [1736,] 1.741321     2
#> [1737,] 1.741321     2
#> [1738,] 1.741321     2
#> [1739,] 1.754902     2
#> [1740,] 1.755661     2
#> [1741,] 1.750926     2
#> [1742,] 1.760727     2
#> [1743,] 1.734486     2
#> [1744,] 1.734486     2
#> [1745,] 1.734486     2
#> [1746,] 1.737288     2
#> [1747,] 1.737288     2
#> [1748,] 1.740638     2
#> [1749,] 1.740638     2
#> [1750,] 1.746226     2
#> [1751,] 1.746226     2
#> [1752,] 1.728106     2
#> [1753,] 1.728106     2
#> [1754,] 1.727314     2
#> [1755,] 1.744983     2
#> [1756,] 1.745623     2
#> [1757,] 1.737323     2
#> [1758,] 1.737323     2
#> [1759,] 1.765282     2
#> [1760,] 1.742790     2
#> [1761,] 1.742790     2
#> [1762,] 1.748541     2
#> [1763,] 1.757319     2
#> [1764,] 1.757319     2
#> [1765,] 1.757319     2
#> [1766,] 1.744608     2
#> [1767,] 1.744324     2
#> [1768,] 1.744324     2
#> [1769,] 1.744031     2
#> [1770,] 1.744031     2
#> [1771,] 1.744031     2
#> [1772,] 1.748337     2
#> [1773,] 1.736014     2
#> [1774,] 1.747536     2
#> [1775,] 1.758056     2
#> [1776,] 1.734680     2
#> [1777,] 1.734680     2
#> [1778,] 1.728393     2
#> [1779,] 1.744897     2
#> [1780,] 1.745132     2
#> [1781,] 1.739970     2
#> [1782,] 1.755188     2
#> [1783,] 1.734388     2
#> [1784,] 1.753854     2
#> [1785,] 1.765579     2
#> [1786,] 1.719797     2
#> [1787,] 1.719797     2
#> [1788,] 1.747354     2
#> [1789,] 1.746190     2
#> [1790,] 1.727547     2
#> [1791,] 1.727547     2
#> [1792,] 1.750033     2
#> [1793,] 1.750033     2
#> [1794,] 1.750033     2
#> [1795,] 1.728064     2
#> [1796,] 1.738063     2
#> [1797,] 1.742642     2
#> [1798,] 1.742642     2
#> [1799,] 1.735416     2
#> [1800,] 1.735416     2
#> [1801,] 1.735416     2
#> [1802,] 1.735416     2
#> [1803,] 1.731153     2
#> [1804,] 1.737935     2
#> [1805,] 1.737935     2
#> [1806,] 1.765640     2
#> [1807,] 1.732026     2
#> [1808,] 1.732026     2
#> [1809,] 1.733098     2
#> [1810,] 1.733098     2
#> [1811,] 1.733098     2
#> [1812,] 1.744593     2
#> [1813,] 1.744593     2
#> [1814,] 1.753701     2
#> [1815,] 1.753701     2
#> [1816,] 1.743928     2
#> [1817,] 1.742553     2
#> [1818,] 1.738362     2
#> [1819,] 1.746062     2
#> [1820,] 1.746062     2
#> [1821,] 1.738759     2
#> [1822,] 1.738759     2
#> [1823,] 1.731621     2
#> [1824,] 1.731621     2
#> [1825,] 1.724611     2
#> [1826,] 1.762827     2
#> [1827,] 1.739496     2
#> [1828,] 1.739496     2
#> [1829,] 1.739496     2
#> [1830,] 1.735892     2
#> [1831,] 1.747170     2
#> [1832,] 1.753973     2
#> [1833,] 1.731862     2
#> [1834,] 1.748989     2
#> [1835,] 1.739126     2
#> [1836,] 1.739126     2
#> [1837,] 1.739126     2
#> [1838,] 1.744611     2
#> [1839,] 1.744611     2
#> [1840,] 1.744611     2
#> [1841,] 1.745173     2
#> [1842,] 1.745173     2
#> [1843,] 1.730413     2
#> [1844,] 1.738986     2
#> [1845,] 1.738986     2
#> [1846,] 1.738986     2
#> [1847,] 1.734807     2
#> [1848,] 1.734807     2
#> [1849,] 1.742749     2
#> [1850,] 1.742749     2
#> [1851,] 1.742749     2
#> [1852,] 1.742749     2
#> [1853,] 1.742749     2
#> [1854,] 1.742749     2
#> [1855,] 1.742749     2
#> [1856,] 1.744334     2
#> [1857,] 1.744334     2
#> [1858,] 1.744334     2
#> [1859,] 1.738455     2
#> [1860,] 1.732570     2
#> [1861,] 1.755146     2
#> [1862,] 1.743650     2
#> [1863,] 1.743650     2
#> [1864,] 1.748477     2
#> [1865,] 1.751992     2
#> [1866,] 1.740524     2
#> [1867,] 1.755698     2
#> [1868,] 1.727710     2
#> [1869,] 1.736960     2
#> [1870,] 1.747605     2
#> [1871,] 1.733415     2
#> [1872,] 1.745082     2
#> [1873,] 1.745031     2
#> [1874,] 1.745031     2
#> [1875,] 1.748734     2
#> [1876,] 1.748734     2
#> [1877,] 1.742898     2
#> [1878,] 1.742898     2
#> [1879,] 1.738429     2
#> [1880,] 1.738429     2
#> [1881,] 1.747475     2
#> [1882,] 1.733016     2
#> [1883,] 1.733016     2
#> [1884,] 1.730029     2
#> [1885,] 1.752633     2
#> [1886,] 1.750711     2
#> [1887,] 1.750711     2
#> [1888,] 1.750711     2
#> [1889,] 1.746523     2
#> [1890,] 1.715604     2
#> [1891,] 1.741995     2
#> [1892,] 1.728865     2
#> [1893,] 1.749187     2
#> [1894,] 1.731703     2
#> [1895,] 1.746034     2
#> [1896,] 1.739792     2
#> [1897,] 1.739792     2
#> [1898,] 1.739792     2
#> [1899,] 1.746366     2
#> [1900,] 1.753157     2
#> [1901,] 1.753157     2
#> [1902,] 1.746591     2
#> [1903,] 1.746591     2
#> [1904,] 1.732358     2
#> [1905,] 1.752418     2
#> [1906,] 1.752418     2
#> [1907,] 1.762241     2
#> [1908,] 1.735269     2
#> [1909,] 1.749092     2
#> [1910,] 1.738226     2
#> [1911,] 1.725620     2
#> [1912,] 1.737591     2
#> [1913,] 1.746956     2
#> [1914,] 1.746956     2
#> [1915,] 1.731614     2
#> [1916,] 1.731614     2
#> [1917,] 1.743490     2
#> [1918,] 1.747640     2
#> [1919,] 1.747640     2
#> [1920,] 1.729234     2
#> [1921,] 1.737353     2
#> [1922,] 1.735078     2
#> [1923,] 1.733760     2
#> [1924,] 1.741989     2
#> [1925,] 1.736651     2
#> [1926,] 1.738615     2
#> [1927,] 1.738615     2
#> [1928,] 1.748012     2
#> [1929,] 1.739590     2
#> [1930,] 1.739590     2
#> [1931,] 1.749713     2
#> [1932,] 1.742923     2
#> [1933,] 1.729669     2
#> [1934,] 1.739562     2
#> [1935,] 1.741618     2
#> [1936,] 1.741618     2
#> [1937,] 1.746872     2
#> [1938,] 1.745176     2
#> [1939,] 1.745176     2
#> [1940,] 1.744957     2
#> [1941,] 1.741505     2
#> [1942,] 1.741505     2
#> [1943,] 1.741505     2
#> [1944,] 1.740420     2
#> [1945,] 1.733490     2
#> [1946,] 1.734281     2
#> [1947,] 1.743936     2
#> [1948,] 1.750826     2
#> [1949,] 1.730482     2
#> [1950,] 1.743157     2
#> [1951,] 1.743157     2
#> [1952,] 1.736660     2
#> [1953,] 1.750916     2
#> [1954,] 1.736242     2
#> [1955,] 1.751758     2
#> [1956,] 1.725553     2
#> [1957,] 1.726613     2
#> [1958,] 1.723525     2
#> [1959,] 1.749994     2
#> [1960,] 1.729046     2
#> [1961,] 1.756726     2
#> [1962,] 1.741334     2
#> [1963,] 1.741334     2
#> [1964,] 1.732286     2
#> [1965,] 1.742510     2
#> [1966,] 1.742510     2
#> [1967,] 1.742510     2
#> [1968,] 1.743417     2
#> [1969,] 1.746070     2
#> [1970,] 1.750963     2
#> [1971,] 1.741565     2
#> [1972,] 1.741565     2
#> [1973,] 1.750654     2
#> [1974,] 1.745433     2
#> [1975,] 1.744170     2
#> [1976,] 1.745439     2
#> [1977,] 1.745439     2
#> [1978,] 1.745439     2
#> [1979,] 1.745439     2
#> [1980,] 1.749350     2
#> [1981,] 1.760467     2