Constructor of fitmodel object

A fitmodel object is a list that stores some variables and functions that will be useful to simulate and fit your model during the course.

fitmodel(name = NULL, state.names = NULL, theta.names = NULL,
  simulate = NULL, rPointObs = NULL, dprior = NULL, dPointObs = NULL)

Arguments

name
character. Name of the model (required).

state.names
character vector. Names of the state variables i.e. c("S","I","R") (required).

theta.names
character vector. Names of the parameters i.e. c("R0","infectious.period") (required).

simulate
R-function to simulate forward the model (required). This function takes 3 arguments:

theta named numeric vector. Values of the parameters. Names should match theta.names.

init.state named numeric vector. Initial values of the state variables. Names should match state.names.

times numeric vector. Time sequence for which the state of the model is wanted; the first value of times must be the initial time, i.e. the time of init.state.

and returns a data.fame containing the simulated trajectories that is the values of the state variables (1 per column) at each observation time (1 per row). The first column is time.

rPointObs
R-function that generates a (randomly sampled) observation point from a model point, using an observation model (optional). It thus acts as an inverse of dPointObs (see below). This function takes 2 arguments

model.point named numeric vector. State of the model at a given point in time.

theta named numeric vector. Values of the parameters. Names should match theta.names.

and returns an observation point

dprior
R-function that evaluates the prior density of the parameters at a given theta (optional). The function should take 2 arguments:

theta named numeric vector. Values of the parameters. Names should match theta.names.

log boolean. determines whether the logarithm of the prior density should be returned.

and returns the (logged, if requested) value of the prior density distribution.

dPointObs
R-function that evaluates the likelihood of one data point given the state of the model at the same time point. This function takes 4 arguments:

data.point named numeric vector. Observation time and observed data point.

model.point named numeric vector containing the state of the model at the observation time point.

theta named numeric vector. Parameter values. Useful since parameters are usually needed to compute the likelihood (i.e. reporting rate).

log boolean. determines whether the logarithm of the likelihood should be returned.

and returns the (log-)likelihood. (optional)

Value

a fitmodel object that is a list of 7 elements:

name character, name of the model

state.names vector, names of the state variables.

theta.names vector, names of the parameters.

simulate R-function to simulate forward the model; usage: simulate(theta,init.state,times).

rPointObs R-function to generate simulated observations; usage: rPointObs(model.point, theta).

dprior R-function to evaluate the log-prior of the parameter values; usage: dprior(theta).

dPointObs R-function to evaluate the log-likelihood of one data point; usage: dPointObs(data.point, model.point, theta, log).

See also

testFitmodel

Examples

## create a simple deterministic SIR model with constant population size

SIR_name <- "SIR with constant population size"
SIR_state.names <- c("S","I","R")
SIR_theta.names <- c("R0","D_inf")

SIR_simulateDeterministic <- function(theta,init.state,times) {

    SIR_ode <- function(time, state, parameters) {

        ## parameters
        beta <- parameters[["R0"]] / parameters[["D_inf"]]
        nu <- 1 / parameters[["D_inf"]]

        ## states
        S <- state[["S"]]
        I <- state[["I"]]
        R <- state[["R"]]

        N <- S + I + R

        dS <- -beta * S * I/N
        dI <- beta * S * I/N - nu * I
        dR <- nu * I

        return(list(c(dS, dI, dR)))
    }

    trajectory <- data.frame(ode(y = init.state,
                                 times = times,
                                 func = SIR_ode,
                                 parms = theta,
                                 method = "ode45"))

    return(trajectory)
}

## function to compute log-prior
SIR_prior <- function(theta, log = FALSE) {

    ## uniform prior on R0: U[1,100]
    log.prior.R0 <- dunif(theta[["R0"]], min = 1, max = 100, log = TRUE)
    ## uniform prior on infectious period: U[0,30]
    log.prior.D <- dunif(theta[["D_inf"]], min = 0, max = 30, log = TRUE)

    log.sum <- log.prior.R0 + log.prior.D

    return(ifelse(log, log.sum, exp(log.sum)))
}

## function to compute the likelihood of one data point
SIR_pointLike <- function(data.point, model.point, theta, log = FALSE){

    ## the prevalence is observed through a Poisson process
    return(dpois(x = data.point[["obs"]],
                 lambda = model.point[["I"]],
                 log = log))
}

## function to generate observation from a model simulation
SIR_genObsPoint <- function(model.point, theta){

    ## the prevalence is observed through a Poisson process
    obs.point <- rpois(n = 1, lambda = model.point[["I"]])

    return(c(obs = obs.point))
}

## create deterministic SIR fitmodel
SIR <- fitmodel(
    name = SIR_name,
    state.names = SIR_state.names,
    theta.names = SIR_theta.names,
    simulate = SIR_simulateDeterministic,
    dprior = SIR_prior,
    rPointObs = SIR_genObsPoint,
    dPointObs = SIR_pointLike)

## test
## theta <- c(R0 = 3, D_inf = 2)
## init.state <- c(S = 999, I = 1, R = 0)
## data(epi)