Metropolis-Hasting MCMC

Run n.iterations of a Metropolis-Hasting MCMC to sample from the target distribution using a gaussian proposal kernel. Two optional optimizations are also implemented: truncated gaussian proposal (to match the support of the target distribution, i.e. boundary of the parameters) and adaptative gaussian proposal (to match the size and the shape of the target distribution).

mcmcMH(target, init.theta, proposal.sd = NULL, n.iterations, covmat = NULL,
  limits = list(lower = NULL, upper = NULL), adapt.size.start = NULL,
  adapt.size.cooling = 0.99, adapt.shape.start = NULL,
  print.info.every = n.iterations/100, verbose = FALSE,
  max.scaling.sd = 50, acceptance.rate.weight = NULL,
  acceptance.window = NULL)

Arguments

target
R-function that takes a single argument: theta (named numeric vector of parameter values) and returns a list of 2 elements:

log.density the logged value of the target density, evaluated at theta.

trace a named numeric vector of values to be printed in the trace data.frame returned by mcmcMH.

init.theta
named vector of initial parameter values to start the chain.

proposal.sd
vector of standard deviations. If this is given and covmat is not, a diagonal matrix will be built from this to use as covariance matrix of the multivariate Gaussian proposal distribution. By default, this is set to init.theta/10.

n.iterations
number of iterations to run the MCMC chain.

covmat
named numeric covariance matrix of the multivariate Gaussian proposal distribution. Must have named rows and columns with at least all estimated theta. If proposal.sd is given, this is ignored.

limits
limits for the - potentially truncated - multi-variate normal proposal distribution of the MCMC. Contains 2 elements:

lower named numeric vector. Lower truncation points in each dimension of the Gaussian proposal distribution. By default they are set to -Inf.

upper named numeric vector. Upper truncation points in each dimension of the Gaussian proposal distribution. By default they are set to Inf.

adapt.size.start
number of iterations to run before adapting the size of the proposal covariance matrix (see note below). Set to 0 (default) if size is not to be adapted.

adapt.size.cooling
cooling factor for the scaling factor of the covariance matrix during size adaptation (see note below).

adapt.shape.start
number of accepted jumps before adapting the shape of the proposal covariance matrix (see note below). Set to 0 (default) if shape is not to be adapted

print.info.every
frequency of information on the chain: acceptance rate and state of the chain. Default value to n.iterations/100. Set to NULL to avoid any info.

verbose
logical. If TRUE, information are printed.

max.scaling.sd
numeric. Maximum value for the scaling factor of the covariance matrix. Avoid too high values for the scaling factor, which might happen due to the exponential update scheme. In this case, the covariance matrix becomes too wide and the sampling from the truncated proposal kernel becomes highly inefficient

acceptance.rate.weight
if this is non-NULL, the acceptance rate is calculated as a moving average. This makes sure the acceptance rate is a reflection of recent proposals, and is useful for adaptive MCMC to prevent overshoot. If this is set to a value between 0 and 1, the acceptance rate will be calculated as \(a = a_\mathrm{old} (1 - w) + a_\mathrm{new} w\), where \(a_\mathrm{old}\) is the (moving) average of the acceptance rate at the previous iteration, \(a_\mathrm{new}\) is 1 if the current proposal is accepted and 0 if it is rejected, and \(w\) is the value of acceptance.rate.weight. Values of \(w\) close to 1 give a lot of weight to recent proposals, while values closer to 0 give more weight to older proposals. A reasonable starting value is in the order of 1e-2.

acceptance.window
if given, how many acceptances to store

Value

a list with 3 elements:

trace a data.frame. Each row contains a state of the chain (as returned by target).

acceptance.rate acceptance rate of the MCMC chain.

covmat.empirical empirical covariance matrix of the target sample.

Note

The size of the proposal covariance matrix is adapted using the following formulae: $$\Sigma_{n+1}=\sigma_n * \Sigma_n$$ with \(\sigma_n=\sigma_{n-1}*exp(\alpha^n*(acc - 0.234))\), where \(\alpha\) is equal to adapt.size.cooling and \(acc\) is the acceptance rate of the chain.

The shape of the proposal covariance matrix is adapted using the following formulae: $$\Sigma_{n+1}=2.38^2/d * \Sigma_n$$ with \(\Sigma_n\) the empirical covariance matrix and \(d\) is the number of estimated parameters in the model.

References

Roberts GO, Rosenthal JS. Examples of adaptive MCMC. Journal of Computational and Graphical Statistics. Taylor & Francis; 2009;18(2):349-67.