Burn and thin MCMC chain

Return a burned and thinned trace of the chain.

burnAndThin(trace, burn = 0, thin = 0)

Arguments

trace

either a data.frame or a list of data.frame with all variables in column, as returned by mcmcMh. Accept also an mcmc or mcmc.list object.

burn

proportion of the chain to burn.

thin

number of samples to discard per sample that is being kept

Value

an object with the same format as trace (data.frame or

list of data.frame or mcmc object or mcmc.list

object)

Examples

data(mcmcEpi)
burnAndThin(mcmcEpi1, burn = 100, thin = 4)
#>              R_0 D_inf
#>    [1,] 1.757972     2
#>    [2,] 1.750043     2
#>    [3,] 1.742533     2
#>    [4,] 1.742533     2
#>    [5,] 1.742533     2
#>    [6,] 1.742533     2
#>    [7,] 1.737102     2
#>    [8,] 1.737102     2
#>    [9,] 1.735231     2
#>   [10,] 1.735231     2
#>   [11,] 1.728730     2
#>   [12,] 1.737132     2
#>   [13,] 1.737132     2
#>   [14,] 1.737132     2
#>   [15,] 1.739638     2
#>   [16,] 1.749975     2
#>   [17,] 1.726927     2
#>   [18,] 1.738775     2
#>   [19,] 1.761082     2
#>   [20,] 1.728091     2
#>   [21,] 1.743725     2
#>   [22,] 1.741245     2
#>   [23,] 1.728156     2
#>   [24,] 1.764595     2
#>   [25,] 1.748363     2
#>   [26,] 1.748363     2
#>   [27,] 1.738934     2
#>   [28,] 1.738934     2
#>   [29,] 1.738934     2
#>   [30,] 1.738934     2
#>   [31,] 1.737555     2
#>   [32,] 1.736955     2
#>   [33,] 1.736955     2
#>   [34,] 1.748080     2
#>   [35,] 1.739196     2
#>   [36,] 1.737472     2
#>   [37,] 1.745609     2
#>   [38,] 1.745609     2
#>   [39,] 1.737957     2
#>   [40,] 1.737561     2
#>   [41,] 1.737561     2
#>   [42,] 1.732946     2
#>   [43,] 1.732946     2
#>   [44,] 1.753312     2
#>   [45,] 1.745990     2
#>   [46,] 1.745990     2
#>   [47,] 1.738649     2
#>   [48,] 1.752414     2
#>   [49,] 1.745448     2
#>   [50,] 1.737965     2
#>   [51,] 1.735846     2
#>   [52,] 1.735846     2
#>   [53,] 1.757273     2
#>   [54,] 1.752221     2
#>   [55,] 1.752221     2
#>   [56,] 1.734686     2
#>   [57,] 1.735612     2
#>   [58,] 1.736973     2
#>   [59,] 1.751439     2
#>   [60,] 1.761245     2
#>   [61,] 1.723561     2
#>   [62,] 1.723561     2
#>   [63,] 1.721610     2
#>   [64,] 1.749057     2
#>   [65,] 1.749057     2
#>   [66,] 1.747912     2
#>   [67,] 1.750524     2
#>   [68,] 1.750524     2
#>   [69,] 1.737440     2
#>   [70,] 1.740618     2
#>   [71,] 1.735159     2
#>   [72,] 1.735159     2
#>   [73,] 1.735147     2
#>   [74,] 1.737897     2
#>   [75,] 1.737897     2
#>   [76,] 1.729595     2
#>   [77,] 1.752688     2
#>   [78,] 1.748413     2
#>   [79,] 1.748413     2
#>   [80,] 1.748413     2
#>   [81,] 1.738976     2
#>   [82,] 1.738976     2
#>   [83,] 1.746989     2
#>   [84,] 1.747990     2
#>   [85,] 1.755223     2
#>   [86,] 1.738166     2
#>   [87,] 1.751146     2
#>   [88,] 1.730495     2
#>   [89,] 1.741881     2
#>   [90,] 1.741881     2
#>   [91,] 1.741881     2
#>   [92,] 1.746099     2
#>   [93,] 1.746099     2
#>   [94,] 1.752716     2
#>   [95,] 1.740105     2
#>   [96,] 1.733125     2
#>   [97,] 1.733125     2
#>   [98,] 1.747457     2
#>   [99,] 1.747457     2
#>  [100,] 1.733568     2
#>  [101,] 1.753142     2
#>  [102,] 1.753142     2
#>  [103,] 1.753142     2
#>  [104,] 1.754061     2
#>  [105,] 1.754061     2
#>  [106,] 1.740041     2
#>  [107,] 1.741213     2
#>  [108,] 1.741213     2
#>  [109,] 1.741213     2
#>  [110,] 1.743213     2
#>  [111,] 1.754054     2
#>  [112,] 1.754054     2
#>  [113,] 1.754054     2
#>  [114,] 1.754054     2
#>  [115,] 1.752206     2
#>  [116,] 1.753140     2
#>  [117,] 1.731539     2
#>  [118,] 1.736912     2
#>  [119,] 1.747294     2
#>  [120,] 1.744455     2
#>  [121,] 1.738579     2
#>  [122,] 1.750705     2
#>  [123,] 1.750705     2
#>  [124,] 1.741140     2
#>  [125,] 1.741140     2
#>  [126,] 1.741140     2
#>  [127,] 1.747355     2
#>  [128,] 1.747355     2
#>  [129,] 1.751503     2
#>  [130,] 1.739258     2
#>  [131,] 1.739258     2
#>  [132,] 1.743797     2
#>  [133,] 1.743797     2
#>  [134,] 1.743797     2
#>  [135,] 1.745360     2
#>  [136,] 1.745360     2
#>  [137,] 1.740580     2
#>  [138,] 1.739763     2
#>  [139,] 1.749674     2
#>  [140,] 1.749674     2
#>  [141,] 1.747366     2
#>  [142,] 1.743959     2
#>  [143,] 1.748303     2
#>  [144,] 1.752429     2
#>  [145,] 1.740962     2
#>  [146,] 1.740962     2
#>  [147,] 1.740962     2
#>  [148,] 1.735938     2
#>  [149,] 1.754166     2
#>  [150,] 1.754166     2
#>  [151,] 1.764980     2
#>  [152,] 1.738930     2
#>  [153,] 1.738930     2
#>  [154,] 1.744273     2
#>  [155,] 1.752414     2
#>  [156,] 1.752414     2
#>  [157,] 1.744641     2
#>  [158,] 1.744641     2
#>  [159,] 1.739591     2
#>  [160,] 1.739591     2
#>  [161,] 1.737147     2
#>  [162,] 1.737147     2
#>  [163,] 1.737147     2
#>  [164,] 1.749295     2
#>  [165,] 1.739443     2
#>  [166,] 1.739443     2
#>  [167,] 1.739443     2
#>  [168,] 1.755744     2
#>  [169,] 1.736178     2
#>  [170,] 1.741985     2
#>  [171,] 1.741985     2
#>  [172,] 1.728024     2
#>  [173,] 1.726326     2
#>  [174,] 1.725668     2
#>  [175,] 1.758518     2
#>  [176,] 1.721335     2
#>  [177,] 1.741061     2
#>  [178,] 1.741061     2
#>  [179,] 1.741061     2
#>  [180,] 1.741061     2
#>  [181,] 1.741108     2
#>  [182,] 1.744219     2
#>  [183,] 1.744219     2
#>  [184,] 1.752419     2
#>  [185,] 1.745339     2
#>  [186,] 1.739157     2
#>  [187,] 1.748290     2
#>  [188,] 1.748290     2
#>  [189,] 1.748290     2
#>  [190,] 1.748818     2
#>  [191,] 1.748818     2
#>  [192,] 1.748818     2
#>  [193,] 1.737566     2
#>  [194,] 1.730397     2
#>  [195,] 1.739564     2
#>  [196,] 1.752033     2
#>  [197,] 1.752457     2
#>  [198,] 1.752457     2
#>  [199,] 1.748516     2
#>  [200,] 1.748516     2
#>  [201,] 1.738875     2
#>  [202,] 1.736347     2
#>  [203,] 1.729736     2
#>  [204,] 1.744410     2
#>  [205,] 1.744410     2
#>  [206,] 1.733833     2
#>  [207,] 1.740062     2
#>  [208,] 1.740062     2
#>  [209,] 1.740062     2
#>  [210,] 1.740062     2
#>  [211,] 1.743120     2
#>  [212,] 1.748312     2
#>  [213,] 1.748996     2
#>  [214,] 1.748996     2
#>  [215,] 1.748996     2
#>  [216,] 1.728491     2
#>  [217,] 1.727356     2
#>  [218,] 1.727356     2
#>  [219,] 1.734076     2
#>  [220,] 1.741106     2
#>  [221,] 1.741106     2
#>  [222,] 1.741106     2
#>  [223,] 1.747608     2
#>  [224,] 1.749033     2
#>  [225,] 1.743408     2
#>  [226,] 1.740792     2
#>  [227,] 1.743813     2
#>  [228,] 1.743813     2
#>  [229,] 1.743813     2
#>  [230,] 1.743813     2
#>  [231,] 1.755485     2
#>  [232,] 1.763249     2
#>  [233,] 1.739281     2
#>  [234,] 1.740136     2
#>  [235,] 1.740136     2
#>  [236,] 1.740136     2
#>  [237,] 1.742020     2
#>  [238,] 1.742020     2
#>  [239,] 1.742020     2
#>  [240,] 1.745634     2
#>  [241,] 1.745634     2
#>  [242,] 1.745266     2
#>  [243,] 1.726317     2
#>  [244,] 1.726317     2
#>  [245,] 1.749564     2
#>  [246,] 1.749564     2
#>  [247,] 1.728824     2
#>  [248,] 1.741907     2
#>  [249,] 1.741907     2
#>  [250,] 1.737259     2
#>  [251,] 1.735461     2
#>  [252,] 1.738649     2
#>  [253,] 1.737186     2
#>  [254,] 1.741241     2
#>  [255,] 1.743554     2
#>  [256,] 1.743554     2
#>  [257,] 1.750745     2
#>  [258,] 1.750745     2
#>  [259,] 1.750745     2
#>  [260,] 1.750745     2
#>  [261,] 1.750745     2
#>  [262,] 1.730831     2
#>  [263,] 1.730831     2
#>  [264,] 1.738796     2
#>  [265,] 1.738796     2
#>  [266,] 1.739709     2
#>  [267,] 1.739709     2
#>  [268,] 1.744228     2
#>  [269,] 1.748888     2
#>  [270,] 1.734121     2
#>  [271,] 1.740567     2
#>  [272,] 1.740567     2
#>  [273,] 1.750057     2
#>  [274,] 1.746785     2
#>  [275,] 1.746785     2
#>  [276,] 1.741433     2
#>  [277,] 1.741433     2
#>  [278,] 1.738149     2
#>  [279,] 1.764308     2
#>  [280,] 1.735954     2
#>  [281,] 1.735954     2
#>  [282,] 1.754449     2
#>  [283,] 1.748491     2
#>  [284,] 1.748491     2
#>  [285,] 1.748201     2
#>  [286,] 1.742471     2
#>  [287,] 1.742471     2
#>  [288,] 1.742471     2
#>  [289,] 1.746787     2
#>  [290,] 1.746787     2
#>  [291,] 1.746787     2
#>  [292,] 1.744424     2
#>  [293,] 1.738443     2
#>  [294,] 1.738728     2
#>  [295,] 1.738728     2
#>  [296,] 1.748023     2
#>  [297,] 1.737245     2
#>  [298,] 1.749153     2
#>  [299,] 1.744927     2
#>  [300,] 1.736841     2
#>  [301,] 1.737532     2
#>  [302,] 1.748747     2
#>  [303,] 1.739614     2
#>  [304,] 1.739614     2
#>  [305,] 1.745021     2
#>  [306,] 1.745021     2
#>  [307,] 1.745021     2
#>  [308,] 1.750913     2
#>  [309,] 1.748039     2
#>  [310,] 1.751803     2
#>  [311,] 1.758217     2
#>  [312,] 1.730983     2
#>  [313,] 1.757876     2
#>  [314,] 1.757876     2
#>  [315,] 1.757876     2
#>  [316,] 1.742611     2
#>  [317,] 1.742611     2
#>  [318,] 1.742416     2
#>  [319,] 1.757742     2
#>  [320,] 1.747116     2
#>  [321,] 1.747116     2
#>  [322,] 1.730545     2
#>  [323,] 1.748564     2
#>  [324,] 1.748564     2
#>  [325,] 1.743763     2
#>  [326,] 1.753842     2
#>  [327,] 1.753842     2
#>  [328,] 1.738446     2
#>  [329,] 1.743994     2
#>  [330,] 1.743994     2
#>  [331,] 1.743994     2
#>  [332,] 1.747555     2
#>  [333,] 1.750393     2
#>  [334,] 1.746215     2
#>  [335,] 1.737891     2
#>  [336,] 1.737891     2
#>  [337,] 1.737891     2
#>  [338,] 1.729585     2
#>  [339,] 1.742676     2
#>  [340,] 1.740327     2
#>  [341,] 1.740327     2
#>  [342,] 1.746432     2
#>  [343,] 1.746432     2
#>  [344,] 1.746432     2
#>  [345,] 1.746432     2
#>  [346,] 1.746432     2
#>  [347,] 1.746432     2
#>  [348,] 1.747142     2
#>  [349,] 1.747142     2
#>  [350,] 1.755812     2
#>  [351,] 1.755812     2
#>  [352,] 1.753767     2
#>  [353,] 1.738310     2
#>  [354,] 1.757663     2
#>  [355,] 1.747771     2
#>  [356,] 1.742968     2
#>  [357,] 1.742968     2
#>  [358,] 1.728160     2
#>  [359,] 1.737850     2
#>  [360,] 1.733331     2
#>  [361,] 1.733331     2
#>  [362,] 1.744586     2
#>  [363,] 1.745542     2
#>  [364,] 1.737218     2
#>  [365,] 1.737218     2
#>  [366,] 1.742799     2
#>  [367,] 1.730235     2
#>  [368,] 1.754952     2
#>  [369,] 1.750877     2
#>  [370,] 1.750877     2
#>  [371,] 1.735414     2
#>  [372,] 1.748035     2
#>  [373,] 1.732229     2
#>  [374,] 1.732229     2
#>  [375,] 1.756019     2
#>  [376,] 1.742241     2
#>  [377,] 1.742241     2
#>  [378,] 1.746443     2
#>  [379,] 1.748906     2
#>  [380,] 1.741610     2
#>  [381,] 1.746184     2
#>  [382,] 1.736351     2
#>  [383,] 1.736351     2
#>  [384,] 1.736351     2
#>  [385,] 1.738082     2
#>  [386,] 1.740189     2
#>  [387,] 1.745304     2
#>  [388,] 1.745304     2
#>  [389,] 1.745304     2
#>  [390,] 1.745304     2
#>  [391,] 1.737643     2
#>  [392,] 1.740398     2
#>  [393,] 1.740398     2
#>  [394,] 1.728813     2
#>  [395,] 1.728813     2
#>  [396,] 1.728808     2
#>  [397,] 1.745779     2
#>  [398,] 1.745779     2
#>  [399,] 1.751069     2
#>  [400,] 1.745832     2
#>  [401,] 1.745832     2
#>  [402,] 1.740883     2
#>  [403,] 1.739024     2
#>  [404,] 1.734361     2
#>  [405,] 1.734361     2
#>  [406,] 1.735141     2
#>  [407,] 1.749264     2
#>  [408,] 1.755205     2
#>  [409,] 1.751936     2
#>  [410,] 1.748718     2
#>  [411,] 1.746193     2
#>  [412,] 1.746193     2
#>  [413,] 1.746263     2
#>  [414,] 1.736026     2
#>  [415,] 1.736026     2
#>  [416,] 1.745195     2
#>  [417,] 1.750422     2
#>  [418,] 1.742996     2
#>  [419,] 1.742799     2
#>  [420,] 1.742799     2
#>  [421,] 1.751070     2
#>  [422,] 1.741546     2
#>  [423,] 1.760987     2
#>  [424,] 1.737070     2
#>  [425,] 1.754299     2
#>  [426,] 1.755651     2
#>  [427,] 1.734088     2
#>  [428,] 1.734088     2
#>  [429,] 1.754636     2
#>  [430,] 1.737075     2
#>  [431,] 1.737075     2
#>  [432,] 1.737075     2
#>  [433,] 1.744377     2
#>  [434,] 1.744377     2
#>  [435,] 1.752756     2
#>  [436,] 1.752756     2
#>  [437,] 1.752756     2
#>  [438,] 1.749939     2
#>  [439,] 1.749939     2
#>  [440,] 1.733225     2
#>  [441,] 1.733225     2
#>  [442,] 1.754200     2
#>  [443,] 1.754200     2
#>  [444,] 1.740348     2
#>  [445,] 1.773701     2
#>  [446,] 1.737132     2
#>  [447,] 1.745301     2
#>  [448,] 1.745301     2
#>  [449,] 1.754007     2
#>  [450,] 1.754007     2
#>  [451,] 1.742398     2
#>  [452,] 1.739257     2
#>  [453,] 1.739902     2
#>  [454,] 1.744822     2
#>  [455,] 1.753386     2
#>  [456,] 1.753386     2
#>  [457,] 1.747461     2
#>  [458,] 1.758765     2
#>  [459,] 1.753108     2
#>  [460,] 1.758581     2
#>  [461,] 1.740433     2
#>  [462,] 1.740433     2
#>  [463,] 1.740433     2
#>  [464,] 1.753742     2
#>  [465,] 1.735294     2
#>  [466,] 1.750887     2
#>  [467,] 1.743942     2
#>  [468,] 1.743942     2
#>  [469,] 1.752762     2
#>  [470,] 1.758427     2
#>  [471,] 1.758427     2
#>  [472,] 1.737006     2
#>  [473,] 1.737006     2
#>  [474,] 1.747964     2
#>  [475,] 1.741053     2
#>  [476,] 1.741053     2
#>  [477,] 1.741053     2
#>  [478,] 1.735088     2
#>  [479,] 1.735088     2
#>  [480,] 1.738961     2
#>  [481,] 1.738961     2
#>  [482,] 1.742577     2
#>  [483,] 1.742577     2
#>  [484,] 1.754178     2
#>  [485,] 1.723040     2
#>  [486,] 1.723040     2
#>  [487,] 1.768595     2
#>  [488,] 1.768595     2
#>  [489,] 1.740634     2
#>  [490,] 1.740634     2
#>  [491,] 1.742515     2
#>  [492,] 1.748458     2
#>  [493,] 1.748458     2
#>  [494,] 1.737958     2
#>  [495,] 1.737958     2
#>  [496,] 1.748190     2
#>  [497,] 1.750185     2
#>  [498,] 1.738474     2
#>  [499,] 1.738412     2
#>  [500,] 1.748696     2
#>  [501,] 1.757606     2
#>  [502,] 1.742787     2
#>  [503,] 1.735290     2
#>  [504,] 1.743228     2
#>  [505,] 1.746194     2
#>  [506,] 1.735644     2
#>  [507,] 1.731311     2
#>  [508,] 1.760233     2
#>  [509,] 1.728249     2
#>  [510,] 1.733280     2
#>  [511,] 1.733280     2
#>  [512,] 1.746266     2
#>  [513,] 1.741088     2
#>  [514,] 1.734540     2
#>  [515,] 1.734540     2
#>  [516,] 1.747770     2
#>  [517,] 1.747770     2
#>  [518,] 1.743463     2
#>  [519,] 1.729222     2
#>  [520,] 1.729741     2
#>  [521,] 1.740994     2
#>  [522,] 1.728287     2
#>  [523,] 1.747302     2
#>  [524,] 1.739819     2
#>  [525,] 1.739819     2
#>  [526,] 1.746471     2
#>  [527,] 1.738361     2
#>  [528,] 1.738361     2
#>  [529,] 1.738361     2
#>  [530,] 1.738361     2
#>  [531,] 1.738361     2
#>  [532,] 1.744579     2
#>  [533,] 1.744579     2
#>  [534,] 1.745469     2
#>  [535,] 1.745469     2
#>  [536,] 1.745469     2
#>  [537,] 1.750546     2
#>  [538,] 1.748527     2
#>  [539,] 1.748527     2
#>  [540,] 1.738077     2
#>  [541,] 1.744850     2
#>  [542,] 1.744850     2
#>  [543,] 1.728344     2
#>  [544,] 1.752360     2
#>  [545,] 1.752360     2
#>  [546,] 1.741667     2
#>  [547,] 1.743364     2
#>  [548,] 1.758707     2
#>  [549,] 1.738645     2
#>  [550,] 1.733412     2
#>  [551,] 1.739881     2
#>  [552,] 1.739881     2
#>  [553,] 1.739647     2
#>  [554,] 1.740306     2
#>  [555,] 1.749108     2
#>  [556,] 1.753744     2
#>  [557,] 1.741180     2
#>  [558,] 1.735641     2
#>  [559,] 1.756289     2
#>  [560,] 1.760115     2
#>  [561,] 1.766503     2
#>  [562,] 1.718428     2
#>  [563,] 1.743004     2
#>  [564,] 1.741497     2
#>  [565,] 1.741497     2
#>  [566,] 1.741497     2
#>  [567,] 1.736680     2
#>  [568,] 1.737690     2
#>  [569,] 1.738520     2
#>  [570,] 1.733736     2
#>  [571,] 1.748988     2
#>  [572,] 1.749368     2
#>  [573,] 1.760372     2
#>  [574,] 1.741932     2
#>  [575,] 1.741932     2
#>  [576,] 1.741932     2
#>  [577,] 1.743305     2
#>  [578,] 1.737134     2
#>  [579,] 1.749639     2
#>  [580,] 1.749639     2
#>  [581,] 1.743934     2
#>  [582,] 1.735605     2
#>  [583,] 1.736049     2
#>  [584,] 1.744596     2
#>  [585,] 1.739148     2
#>  [586,] 1.739148     2
#>  [587,] 1.739148     2
#>  [588,] 1.750090     2
#>  [589,] 1.738234     2
#>  [590,] 1.738234     2
#>  [591,] 1.745251     2
#>  [592,] 1.745251     2
#>  [593,] 1.745251     2
#>  [594,] 1.745251     2
#>  [595,] 1.745251     2
#>  [596,] 1.757033     2
#>  [597,] 1.753370     2
#>  [598,] 1.754643     2
#>  [599,] 1.754643     2
#>  [600,] 1.754643     2
#>  [601,] 1.746982     2
#>  [602,] 1.743470     2
#>  [603,] 1.742476     2
#>  [604,] 1.742195     2
#>  [605,] 1.748262     2
#>  [606,] 1.748262     2
#>  [607,] 1.748262     2
#>  [608,] 1.748262     2
#>  [609,] 1.752117     2
#>  [610,] 1.752117     2
#>  [611,] 1.742751     2
#>  [612,] 1.742751     2
#>  [613,] 1.742751     2
#>  [614,] 1.742751     2
#>  [615,] 1.742751     2
#>  [616,] 1.745020     2
#>  [617,] 1.743097     2
#>  [618,] 1.743097     2
#>  [619,] 1.752807     2
#>  [620,] 1.733081     2
#>  [621,] 1.754848     2
#>  [622,] 1.741079     2
#>  [623,] 1.741079     2
#>  [624,] 1.753671     2
#>  [625,] 1.753671     2
#>  [626,] 1.730732     2
#>  [627,] 1.739421     2
#>  [628,] 1.738894     2
#>  [629,] 1.746054     2
#>  [630,] 1.746054     2
#>  [631,] 1.756553     2
#>  [632,] 1.756553     2
#>  [633,] 1.743732     2
#>  [634,] 1.759415     2
#>  [635,] 1.740201     2
#>  [636,] 1.740201     2
#>  [637,] 1.751332     2
#>  [638,] 1.753259     2
#>  [639,] 1.741965     2
#>  [640,] 1.741965     2
#>  [641,] 1.755467     2
#>  [642,] 1.739514     2
#>  [643,] 1.739514     2
#>  [644,] 1.739514     2
#>  [645,] 1.747659     2
#>  [646,] 1.750674     2
#>  [647,] 1.730377     2
#>  [648,] 1.760207     2
#>  [649,] 1.747641     2
#>  [650,] 1.747641     2
#>  [651,] 1.742252     2
#>  [652,] 1.750992     2
#>  [653,] 1.716826     2
#>  [654,] 1.764585     2
#>  [655,] 1.754973     2
#>  [656,] 1.756466     2
#>  [657,] 1.756466     2
#>  [658,] 1.742805     2
#>  [659,] 1.735937     2
#>  [660,] 1.735937     2
#>  [661,] 1.735937     2
#>  [662,] 1.747927     2
#>  [663,] 1.737719     2
#>  [664,] 1.759602     2
#>  [665,] 1.737558     2
#>  [666,] 1.737558     2
#>  [667,] 1.737558     2
#>  [668,] 1.743397     2
#>  [669,] 1.743397     2
#>  [670,] 1.743397     2
#>  [671,] 1.751082     2
#>  [672,] 1.742369     2
#>  [673,] 1.742369     2
#>  [674,] 1.742369     2
#>  [675,] 1.742369     2
#>  [676,] 1.750176     2
#>  [677,] 1.737409     2
#>  [678,] 1.749325     2
#>  [679,] 1.749325     2
#>  [680,] 1.749325     2
#>  [681,] 1.749908     2
#>  [682,] 1.731347     2
#>  [683,] 1.731347     2
#>  [684,] 1.734718     2
#>  [685,] 1.742844     2
#>  [686,] 1.742844     2
#>  [687,] 1.743809     2
#>  [688,] 1.748508     2
#>  [689,] 1.750988     2
#>  [690,] 1.756747     2
#>  [691,] 1.749787     2
#>  [692,] 1.742257     2
#>  [693,] 1.743326     2
#>  [694,] 1.755972     2
#>  [695,] 1.742188     2
#>  [696,] 1.756666     2
#>  [697,] 1.752038     2
#>  [698,] 1.737718     2
#>  [699,] 1.734815     2
#>  [700,] 1.730728     2
#>  [701,] 1.730728     2
#>  [702,] 1.741069     2
#>  [703,] 1.741069     2
#>  [704,] 1.734069     2
#>  [705,] 1.734069     2
#>  [706,] 1.750152     2
#>  [707,] 1.746065     2
#>  [708,] 1.733230     2
#>  [709,] 1.747065     2
#>  [710,] 1.747065     2
#>  [711,] 1.730764     2
#>  [712,] 1.730764     2
#>  [713,] 1.734661     2
#>  [714,] 1.734661     2
#>  [715,] 1.738281     2
#>  [716,] 1.744044     2
#>  [717,] 1.736926     2
#>  [718,] 1.736926     2
#>  [719,] 1.743142     2
#>  [720,] 1.743142     2
#>  [721,] 1.745005     2
#>  [722,] 1.736681     2
#>  [723,] 1.733224     2
#>  [724,] 1.750748     2
#>  [725,] 1.751866     2
#>  [726,] 1.751866     2
#>  [727,] 1.751866     2
#>  [728,] 1.738404     2
#>  [729,] 1.738404     2
#>  [730,] 1.738404     2
#>  [731,] 1.738404     2
#>  [732,] 1.736607     2
#>  [733,] 1.741236     2
#>  [734,] 1.741236     2
#>  [735,] 1.741236     2
#>  [736,] 1.741236     2
#>  [737,] 1.743909     2
#>  [738,] 1.740123     2
#>  [739,] 1.740123     2
#>  [740,] 1.766070     2
#>  [741,] 1.743272     2
#>  [742,] 1.743272     2
#>  [743,] 1.742433     2
#>  [744,] 1.747773     2
#>  [745,] 1.747773     2
#>  [746,] 1.738267     2
#>  [747,] 1.738183     2
#>  [748,] 1.738183     2
#>  [749,] 1.748466     2
#>  [750,] 1.741556     2
#>  [751,] 1.745679     2
#>  [752,] 1.738725     2
#>  [753,] 1.738725     2
#>  [754,] 1.738725     2
#>  [755,] 1.750072     2
#>  [756,] 1.736352     2
#>  [757,] 1.736352     2
#>  [758,] 1.746726     2
#>  [759,] 1.737302     2
#>  [760,] 1.754706     2
#>  [761,] 1.763095     2
#>  [762,] 1.720590     2
#>  [763,] 1.742206     2
#>  [764,] 1.754928     2
#>  [765,] 1.746143     2
#>  [766,] 1.743122     2
#>  [767,] 1.749706     2
#>  [768,] 1.749205     2
#>  [769,] 1.736313     2
#>  [770,] 1.749751     2
#>  [771,] 1.749751     2
#>  [772,] 1.744680     2
#>  [773,] 1.752943     2
#>  [774,] 1.742297     2
#>  [775,] 1.761730     2
#>  [776,] 1.734389     2
#>  [777,] 1.745404     2
#>  [778,] 1.745619     2
#>  [779,] 1.748430     2
#>  [780,] 1.748430     2
#>  [781,] 1.750015     2
#>  [782,] 1.731567     2
#>  [783,] 1.749153     2
#>  [784,] 1.749153     2
#>  [785,] 1.749153     2
#>  [786,] 1.749153     2
#>  [787,] 1.748650     2
#>  [788,] 1.742555     2
#>  [789,] 1.755673     2
#>  [790,] 1.745351     2
#>  [791,] 1.745288     2
#>  [792,] 1.745288     2
#>  [793,] 1.737962     2
#>  [794,] 1.758050     2
#>  [795,] 1.734314     2
#>  [796,] 1.738558     2
#>  [797,] 1.738558     2
#>  [798,] 1.738558     2
#>  [799,] 1.737907     2
#>  [800,] 1.749602     2
#>  [801,] 1.749602     2
#>  [802,] 1.743633     2
#>  [803,] 1.729148     2
#>  [804,] 1.763043     2
#>  [805,] 1.763043     2
#>  [806,] 1.739308     2
#>  [807,] 1.733846     2
#>  [808,] 1.740125     2
#>  [809,] 1.747077     2
#>  [810,] 1.747077     2
#>  [811,] 1.756777     2
#>  [812,] 1.731979     2
#>  [813,] 1.754588     2
#>  [814,] 1.740697     2
#>  [815,] 1.740697     2
#>  [816,] 1.740697     2
#>  [817,] 1.755900     2
#>  [818,] 1.723248     2
#>  [819,] 1.740131     2
#>  [820,] 1.740131     2
#>  [821,] 1.740131     2
#>  [822,] 1.740131     2
#>  [823,] 1.740407     2
#>  [824,] 1.740407     2
#>  [825,] 1.746040     2
#>  [826,] 1.747887     2
#>  [827,] 1.741396     2
#>  [828,] 1.741396     2
#>  [829,] 1.749210     2
#>  [830,] 1.749210     2
#>  [831,] 1.749210     2
#>  [832,] 1.749210     2
#>  [833,] 1.741602     2
#>  [834,] 1.749704     2
#>  [835,] 1.744819     2
#>  [836,] 1.736830     2
#>  [837,] 1.746836     2
#>  [838,] 1.746836     2
#>  [839,] 1.746836     2
#>  [840,] 1.741451     2
#>  [841,] 1.735343     2
#>  [842,] 1.735343     2
#>  [843,] 1.735343     2
#>  [844,] 1.735343     2
#>  [845,] 1.752775     2
#>  [846,] 1.738476     2
#>  [847,] 1.748085     2
#>  [848,] 1.748085     2
#>  [849,] 1.748085     2
#>  [850,] 1.757715     2
#>  [851,] 1.741790     2
#>  [852,] 1.749651     2
#>  [853,] 1.749651     2
#>  [854,] 1.738295     2
#>  [855,] 1.749926     2
#>  [856,] 1.751301     2
#>  [857,] 1.744856     2
#>  [858,] 1.747165     2
#>  [859,] 1.734813     2
#>  [860,] 1.747958     2
#>  [861,] 1.747958     2
#>  [862,] 1.747159     2
#>  [863,] 1.742715     2
#>  [864,] 1.746948     2
#>  [865,] 1.746948     2
#>  [866,] 1.750102     2
#>  [867,] 1.750102     2
#>  [868,] 1.742250     2
#>  [869,] 1.743574     2
#>  [870,] 1.747680     2
#>  [871,] 1.753486     2
#>  [872,] 1.736732     2
#>  [873,] 1.749463     2
#>  [874,] 1.738099     2
#>  [875,] 1.743238     2
#>  [876,] 1.754674     2
#>  [877,] 1.757137     2
#>  [878,] 1.728822     2
#>  [879,] 1.738763     2
#>  [880,] 1.735748     2
#>  [881,] 1.739596     2
#>  [882,] 1.739797     2
#>  [883,] 1.731398     2
#>  [884,] 1.731398     2
#>  [885,] 1.740469     2
#>  [886,] 1.752821     2
#>  [887,] 1.733177     2
#>  [888,] 1.733177     2
#>  [889,] 1.720435     2
#>  [890,] 1.740539     2
#>  [891,] 1.734341     2
#>  [892,] 1.736113     2
#>  [893,] 1.736113     2
#>  [894,] 1.735408     2
#>  [895,] 1.747027     2
#>  [896,] 1.747420     2
#>  [897,] 1.743442     2
#>  [898,] 1.745993     2
#>  [899,] 1.745993     2
#>  [900,] 1.745993     2
#>  [901,] 1.743799     2
#>  [902,] 1.743799     2
#>  [903,] 1.743799     2
#>  [904,] 1.752764     2
#>  [905,] 1.744880     2
#>  [906,] 1.744880     2
#>  [907,] 1.730733     2
#>  [908,] 1.751467     2
#>  [909,] 1.751467     2
#>  [910,] 1.751467     2
#>  [911,] 1.751467     2
#>  [912,] 1.751467     2
#>  [913,] 1.743203     2
#>  [914,] 1.743032     2
#>  [915,] 1.743032     2
#>  [916,] 1.743032     2
#>  [917,] 1.740077     2
#>  [918,] 1.740077     2
#>  [919,] 1.740077     2
#>  [920,] 1.740077     2
#>  [921,] 1.740077     2
#>  [922,] 1.743652     2
#>  [923,] 1.743652     2
#>  [924,] 1.750952     2
#>  [925,] 1.750952     2
#>  [926,] 1.742787     2
#>  [927,] 1.740160     2
#>  [928,] 1.740160     2
#>  [929,] 1.742592     2
#>  [930,] 1.737631     2
#>  [931,] 1.745714     2
#>  [932,] 1.745714     2
#>  [933,] 1.741688     2
#>  [934,] 1.729180     2
#>  [935,] 1.729180     2
#>  [936,] 1.753648     2
#>  [937,] 1.753730     2
#>  [938,] 1.747444     2
#>  [939,] 1.750858     2
#>  [940,] 1.748139     2
#>  [941,] 1.748139     2
#>  [942,] 1.737637     2
#>  [943,] 1.737637     2
#>  [944,] 1.730594     2
#>  [945,] 1.730594     2
#>  [946,] 1.741706     2
#>  [947,] 1.741413     2
#>  [948,] 1.741413     2
#>  [949,] 1.755945     2
#>  [950,] 1.731847     2
#>  [951,] 1.736846     2
#>  [952,] 1.744867     2
#>  [953,] 1.736389     2
#>  [954,] 1.740618     2
#>  [955,] 1.740618     2
#>  [956,] 1.740618     2
#>  [957,] 1.748154     2
#>  [958,] 1.733936     2
#>  [959,] 1.733936     2
#>  [960,] 1.733936     2
#>  [961,] 1.733936     2
#>  [962,] 1.737056     2
#>  [963,] 1.742749     2
#>  [964,] 1.749368     2
#>  [965,] 1.746483     2
#>  [966,] 1.752471     2
#>  [967,] 1.742467     2
#>  [968,] 1.752678     2
#>  [969,] 1.752678     2
#>  [970,] 1.756814     2
#>  [971,] 1.751236     2
#>  [972,] 1.746240     2
#>  [973,] 1.746240     2
#>  [974,] 1.745229     2
#>  [975,] 1.745229     2
#>  [976,] 1.745709     2
#>  [977,] 1.742110     2
#>  [978,] 1.742110     2
#>  [979,] 1.742110     2
#>  [980,] 1.742110     2
#>  [981,] 1.742110     2
#>  [982,] 1.732409     2
#>  [983,] 1.746747     2
#>  [984,] 1.746747     2
#>  [985,] 1.746747     2
#>  [986,] 1.725131     2
#>  [987,] 1.725131     2
#>  [988,] 1.725131     2
#>  [989,] 1.761732     2
#>  [990,] 1.762878     2
#>  [991,] 1.749957     2
#>  [992,] 1.736775     2
#>  [993,] 1.742619     2
#>  [994,] 1.743712     2
#>  [995,] 1.743712     2
#>  [996,] 1.748606     2
#>  [997,] 1.748606     2
#>  [998,] 1.748606     2
#>  [999,] 1.755060     2
#> [1000,] 1.729918     2
#> [1001,] 1.730378     2
#> [1002,] 1.749928     2
#> [1003,] 1.737804     2
#> [1004,] 1.748371     2
#> [1005,] 1.754850     2
#> [1006,] 1.744097     2
#> [1007,] 1.744097     2
#> [1008,] 1.736437     2
#> [1009,] 1.744709     2
#> [1010,] 1.737315     2
#> [1011,] 1.751179     2
#> [1012,] 1.745733     2
#> [1013,] 1.738163     2
#> [1014,] 1.738163     2
#> [1015,] 1.752594     2
#> [1016,] 1.752594     2
#> [1017,] 1.740480     2
#> [1018,] 1.742766     2
#> [1019,] 1.742766     2
#> [1020,] 1.742766     2
#> [1021,] 1.742766     2
#> [1022,] 1.742766     2
#> [1023,] 1.741991     2
#> [1024,] 1.751383     2
#> [1025,] 1.729382     2
#> [1026,] 1.729382     2
#> [1027,] 1.736990     2
#> [1028,] 1.736990     2
#> [1029,] 1.730514     2
#> [1030,] 1.748232     2
#> [1031,] 1.748232     2
#> [1032,] 1.743624     2
#> [1033,] 1.743624     2
#> [1034,] 1.734911     2
#> [1035,] 1.748132     2
#> [1036,] 1.748132     2
#> [1037,] 1.748132     2
#> [1038,] 1.742171     2
#> [1039,] 1.736020     2
#> [1040,] 1.736020     2
#> [1041,] 1.736020     2
#> [1042,] 1.741456     2
#> [1043,] 1.753376     2
#> [1044,] 1.759722     2
#> [1045,] 1.753339     2
#> [1046,] 1.740398     2
#> [1047,] 1.746590     2
#> [1048,] 1.730649     2
#> [1049,] 1.750052     2
#> [1050,] 1.750052     2
#> [1051,] 1.750052     2
#> [1052,] 1.743327     2
#> [1053,] 1.743327     2
#> [1054,] 1.743327     2
#> [1055,] 1.752013     2
#> [1056,] 1.726792     2
#> [1057,] 1.754414     2
#> [1058,] 1.745890     2
#> [1059,] 1.746791     2
#> [1060,] 1.745701     2
#> [1061,] 1.745701     2
#> [1062,] 1.745701     2
#> [1063,] 1.732099     2
#> [1064,] 1.743568     2
#> [1065,] 1.745119     2
#> [1066,] 1.745119     2
#> [1067,] 1.737363     2
#> [1068,] 1.739245     2
#> [1069,] 1.735217     2
#> [1070,] 1.735217     2
#> [1071,] 1.735217     2
#> [1072,] 1.730691     2
#> [1073,] 1.730691     2
#> [1074,] 1.730447     2
#> [1075,] 1.743644     2
#> [1076,] 1.743644     2
#> [1077,] 1.743644     2
#> [1078,] 1.737748     2
#> [1079,] 1.737748     2
#> [1080,] 1.737748     2
#> [1081,] 1.754890     2
#> [1082,] 1.754890     2
#> [1083,] 1.748708     2
#> [1084,] 1.748708     2
#> [1085,] 1.748708     2
#> [1086,] 1.740248     2
#> [1087,] 1.746256     2
#> [1088,] 1.746256     2
#> [1089,] 1.746256     2
#> [1090,] 1.746256     2
#> [1091,] 1.734540     2
#> [1092,] 1.739564     2
#> [1093,] 1.739564     2
#> [1094,] 1.739564     2
#> [1095,] 1.746880     2
#> [1096,] 1.742135     2
#> [1097,] 1.752240     2
#> [1098,] 1.752672     2
#> [1099,] 1.752672     2
#> [1100,] 1.752042     2
#> [1101,] 1.746880     2
#> [1102,] 1.732778     2
#> [1103,] 1.732778     2
#> [1104,] 1.763145     2
#> [1105,] 1.751608     2
#> [1106,] 1.745827     2
#> [1107,] 1.738890     2
#> [1108,] 1.738890     2
#> [1109,] 1.738890     2
#> [1110,] 1.738890     2
#> [1111,] 1.738890     2
#> [1112,] 1.750394     2
#> [1113,] 1.750394     2
#> [1114,] 1.755150     2
#> [1115,] 1.755150     2
#> [1116,] 1.736565     2
#> [1117,] 1.736565     2
#> [1118,] 1.754969     2
#> [1119,] 1.754969     2
#> [1120,] 1.738711     2
#> [1121,] 1.742729     2
#> [1122,] 1.742729     2
#> [1123,] 1.742729     2
#> [1124,] 1.744472     2
#> [1125,] 1.744472     2
#> [1126,] 1.744472     2
#> [1127,] 1.733687     2
#> [1128,] 1.742052     2
#> [1129,] 1.742052     2
#> [1130,] 1.745221     2
#> [1131,] 1.745221     2
#> [1132,] 1.745221     2
#> [1133,] 1.738357     2
#> [1134,] 1.738357     2
#> [1135,] 1.742168     2
#> [1136,] 1.746705     2
#> [1137,] 1.755705     2
#> [1138,] 1.743634     2
#> [1139,] 1.743634     2
#> [1140,] 1.753575     2
#> [1141,] 1.732503     2
#> [1142,] 1.741349     2
#> [1143,] 1.745903     2
#> [1144,] 1.745903     2
#> [1145,] 1.745903     2
#> [1146,] 1.744677     2
#> [1147,] 1.744677     2
#> [1148,] 1.744677     2
#> [1149,] 1.743995     2
#> [1150,] 1.758945     2
#> [1151,] 1.744294     2
#> [1152,] 1.744294     2
#> [1153,] 1.744294     2
#> [1154,] 1.744294     2
#> [1155,] 1.744294     2
#> [1156,] 1.743398     2
#> [1157,] 1.743398     2
#> [1158,] 1.721217     2
#> [1159,] 1.730642     2
#> [1160,] 1.738440     2
#> [1161,] 1.739297     2
#> [1162,] 1.750838     2
#> [1163,] 1.748920     2
#> [1164,] 1.732298     2
#> [1165,] 1.751704     2
#> [1166,] 1.740514     2
#> [1167,] 1.740514     2
#> [1168,] 1.736533     2
#> [1169,] 1.761124     2
#> [1170,] 1.719360     2
#> [1171,] 1.740796     2
#> [1172,] 1.744644     2
#> [1173,] 1.744697     2
#> [1174,] 1.746402     2
#> [1175,] 1.743724     2
#> [1176,] 1.734452     2
#> [1177,] 1.764630     2
#> [1178,] 1.750124     2
#> [1179,] 1.753179     2
#> [1180,] 1.754422     2
#> [1181,] 1.754422     2
#> [1182,] 1.757521     2
#> [1183,] 1.757521     2
#> [1184,] 1.757521     2
#> [1185,] 1.750094     2
#> [1186,] 1.750094     2
#> [1187,] 1.740899     2
#> [1188,] 1.733406     2
#> [1189,] 1.746695     2
#> [1190,] 1.746695     2
#> [1191,] 1.742086     2
#> [1192,] 1.742796     2
#> [1193,] 1.742796     2
#> [1194,] 1.749958     2
#> [1195,] 1.743217     2
#> [1196,] 1.747099     2
#> [1197,] 1.747099     2
#> [1198,] 1.747099     2
#> [1199,] 1.747888     2
#> [1200,] 1.747888     2
#> [1201,] 1.738083     2
#> [1202,] 1.744320     2
#> [1203,] 1.744320     2
#> [1204,] 1.744320     2
#> [1205,] 1.755697     2
#> [1206,] 1.747547     2
#> [1207,] 1.743240     2
#> [1208,] 1.742379     2
#> [1209,] 1.742379     2
#> [1210,] 1.742379     2
#> [1211,] 1.742379     2
#> [1212,] 1.729899     2
#> [1213,] 1.740948     2
#> [1214,] 1.749694     2
#> [1215,] 1.749694     2
#> [1216,] 1.749694     2
#> [1217,] 1.749694     2
#> [1218,] 1.728606     2
#> [1219,] 1.742412     2
#> [1220,] 1.735730     2
#> [1221,] 1.751396     2
#> [1222,] 1.751396     2
#> [1223,] 1.740545     2
#> [1224,] 1.740545     2
#> [1225,] 1.740545     2
#> [1226,] 1.742161     2
#> [1227,] 1.742161     2
#> [1228,] 1.742161     2
#> [1229,] 1.742161     2
#> [1230,] 1.739426     2
#> [1231,] 1.747399     2
#> [1232,] 1.751936     2
#> [1233,] 1.751936     2
#> [1234,] 1.737833     2
#> [1235,] 1.737833     2
#> [1236,] 1.737833     2
#> [1237,] 1.737833     2
#> [1238,] 1.739559     2
#> [1239,] 1.739559     2
#> [1240,] 1.739559     2
#> [1241,] 1.739559     2
#> [1242,] 1.739559     2
#> [1243,] 1.736658     2
#> [1244,] 1.736658     2
#> [1245,] 1.752800     2
#> [1246,] 1.738137     2
#> [1247,] 1.738137     2
#> [1248,] 1.737409     2
#> [1249,] 1.741340     2
#> [1250,] 1.757090     2
#> [1251,] 1.756747     2
#> [1252,] 1.740235     2
#> [1253,] 1.740235     2
#> [1254,] 1.753383     2
#> [1255,] 1.753144     2
#> [1256,] 1.753144     2
#> [1257,] 1.743633     2
#> [1258,] 1.743633     2
#> [1259,] 1.741701     2
#> [1260,] 1.741701     2
#> [1261,] 1.735778     2
#> [1262,] 1.733233     2
#> [1263,] 1.762896     2
#> [1264,] 1.730434     2
#> [1265,] 1.730434     2
#> [1266,] 1.740697     2
#> [1267,] 1.740697     2
#> [1268,] 1.730815     2
#> [1269,] 1.741040     2
#> [1270,] 1.741040     2
#> [1271,] 1.753993     2
#> [1272,] 1.756195     2
#> [1273,] 1.747051     2
#> [1274,] 1.737867     2
#> [1275,] 1.735143     2
#> [1276,] 1.735143     2
#> [1277,] 1.750077     2
#> [1278,] 1.757408     2
#> [1279,] 1.735612     2
#> [1280,] 1.735612     2
#> [1281,] 1.735612     2
#> [1282,] 1.735612     2
#> [1283,] 1.742473     2
#> [1284,] 1.728360     2
#> [1285,] 1.738964     2
#> [1286,] 1.747162     2
#> [1287,] 1.747162     2
#> [1288,] 1.747162     2
#> [1289,] 1.747172     2
#> [1290,] 1.747172     2
#> [1291,] 1.737438     2
#> [1292,] 1.737438     2
#> [1293,] 1.743429     2
#> [1294,] 1.743429     2
#> [1295,] 1.743429     2
#> [1296,] 1.744789     2
#> [1297,] 1.744789     2
#> [1298,] 1.741201     2
#> [1299,] 1.741201     2
#> [1300,] 1.746600     2
#> [1301,] 1.740811     2
#> [1302,] 1.750391     2
#> [1303,] 1.750391     2
#> [1304,] 1.744431     2
#> [1305,] 1.744894     2
#> [1306,] 1.744894     2
#> [1307,] 1.736767     2
#> [1308,] 1.739104     2
#> [1309,] 1.729010     2
#> [1310,] 1.751566     2
#> [1311,] 1.751566     2
#> [1312,] 1.741670     2
#> [1313,] 1.744221     2
#> [1314,] 1.736304     2
#> [1315,] 1.736304     2
#> [1316,] 1.736304     2
#> [1317,] 1.754324     2
#> [1318,] 1.734746     2
#> [1319,] 1.744866     2
#> [1320,] 1.751327     2
#> [1321,] 1.762448     2
#> [1322,] 1.762448     2
#> [1323,] 1.743993     2
#> [1324,] 1.755185     2
#> [1325,] 1.746726     2
#> [1326,] 1.746726     2
#> [1327,] 1.755093     2
#> [1328,] 1.755093     2
#> [1329,] 1.730429     2
#> [1330,] 1.739069     2
#> [1331,] 1.739069     2
#> [1332,] 1.741679     2
#> [1333,] 1.741679     2
#> [1334,] 1.741679     2
#> [1335,] 1.741679     2
#> [1336,] 1.741679     2
#> [1337,] 1.741679     2
#> [1338,] 1.741679     2
#> [1339,] 1.741662     2
#> [1340,] 1.741662     2
#> [1341,] 1.741662     2
#> [1342,] 1.741662     2
#> [1343,] 1.741662     2
#> [1344,] 1.746945     2
#> [1345,] 1.744774     2
#> [1346,] 1.744774     2
#> [1347,] 1.744774     2
#> [1348,] 1.754243     2
#> [1349,] 1.754243     2
#> [1350,] 1.746550     2
#> [1351,] 1.744942     2
#> [1352,] 1.759276     2
#> [1353,] 1.751627     2
#> [1354,] 1.736193     2
#> [1355,] 1.736193     2
#> [1356,] 1.741501     2
#> [1357,] 1.750329     2
#> [1358,] 1.740470     2
#> [1359,] 1.740470     2
#> [1360,] 1.740470     2
#> [1361,] 1.740470     2
#> [1362,] 1.740470     2
#> [1363,] 1.746245     2
#> [1364,] 1.746245     2
#> [1365,] 1.745524     2
#> [1366,] 1.738891     2
#> [1367,] 1.744252     2
#> [1368,] 1.744252     2
#> [1369,] 1.732146     2
#> [1370,] 1.735895     2
#> [1371,] 1.737927     2
#> [1372,] 1.737927     2
#> [1373,] 1.723546     2
#> [1374,] 1.748075     2
#> [1375,] 1.748075     2
#> [1376,] 1.766233     2
#> [1377,] 1.765117     2
#> [1378,] 1.758327     2
#> [1379,] 1.727417     2
#> [1380,] 1.727417     2
#> [1381,] 1.743027     2
#> [1382,] 1.731486     2
#> [1383,] 1.731486     2
#> [1384,] 1.746364     2
#> [1385,] 1.746364     2
#> [1386,] 1.741487     2
#> [1387,] 1.741487     2
#> [1388,] 1.741000     2
#> [1389,] 1.740026     2
#> [1390,] 1.735586     2
#> [1391,] 1.734771     2
#> [1392,] 1.763191     2
#> [1393,] 1.761006     2
#> [1394,] 1.737901     2
#> [1395,] 1.737901     2
#> [1396,] 1.737901     2
#> [1397,] 1.745899     2
#> [1398,] 1.735788     2
#> [1399,] 1.748521     2
#> [1400,] 1.748521     2
#> [1401,] 1.742014     2
#> [1402,] 1.742014     2
#> [1403,] 1.745150     2
#> [1404,] 1.758125     2
#> [1405,] 1.755893     2
#> [1406,] 1.755893     2
#> [1407,] 1.748043     2
#> [1408,] 1.748043     2
#> [1409,] 1.748043     2
#> [1410,] 1.748043     2
#> [1411,] 1.732585     2
#> [1412,] 1.732585     2
#> [1413,] 1.732585     2
#> [1414,] 1.758204     2
#> [1415,] 1.745763     2
#> [1416,] 1.731027     2
#> [1417,] 1.745593     2
#> [1418,] 1.745593     2
#> [1419,] 1.749255     2
#> [1420,] 1.736816     2
#> [1421,] 1.743994     2
#> [1422,] 1.743994     2
#> [1423,] 1.741720     2
#> [1424,] 1.753129     2
#> [1425,] 1.753129     2
#> [1426,] 1.748769     2
#> [1427,] 1.747905     2
#> [1428,] 1.747532     2
#> [1429,] 1.747532     2
#> [1430,] 1.766958     2
#> [1431,] 1.727601     2
#> [1432,] 1.745222     2
#> [1433,] 1.745222     2
#> [1434,] 1.745222     2
#> [1435,] 1.733039     2
#> [1436,] 1.758186     2
#> [1437,] 1.750537     2
#> [1438,] 1.750537     2
#> [1439,] 1.750537     2
#> [1440,] 1.749565     2
#> [1441,] 1.759451     2
#> [1442,] 1.735179     2
#> [1443,] 1.749516     2
#> [1444,] 1.738275     2
#> [1445,] 1.741249     2
#> [1446,] 1.757336     2
#> [1447,] 1.749666     2
#> [1448,] 1.728483     2
#> [1449,] 1.727813     2
#> [1450,] 1.755486     2
#> [1451,] 1.755486     2
#> [1452,] 1.758227     2
#> [1453,] 1.726237     2
#> [1454,] 1.726237     2
#> [1455,] 1.726237     2
#> [1456,] 1.735518     2
#> [1457,] 1.735518     2
#> [1458,] 1.735518     2
#> [1459,] 1.753295     2
#> [1460,] 1.742046     2
#> [1461,] 1.736376     2
#> [1462,] 1.758887     2
#> [1463,] 1.734194     2
#> [1464,] 1.743721     2
#> [1465,] 1.743721     2
#> [1466,] 1.744435     2
#> [1467,] 1.744435     2
#> [1468,] 1.749864     2
#> [1469,] 1.749864     2
#> [1470,] 1.749864     2
#> [1471,] 1.741381     2
#> [1472,] 1.741381     2
#> [1473,] 1.738539     2
#> [1474,] 1.733715     2
#> [1475,] 1.733715     2
#> [1476,] 1.733715     2
#> [1477,] 1.752447     2
#> [1478,] 1.752447     2
#> [1479,] 1.745064     2
#> [1480,] 1.745064     2
#> [1481,] 1.744840     2
#> [1482,] 1.760933     2
#> [1483,] 1.760933     2
#> [1484,] 1.760933     2
#> [1485,] 1.737178     2
#> [1486,] 1.747119     2
#> [1487,] 1.748211     2
#> [1488,] 1.752311     2
#> [1489,] 1.752311     2
#> [1490,] 1.740081     2
#> [1491,] 1.743711     2
#> [1492,] 1.729397     2
#> [1493,] 1.747929     2
#> [1494,] 1.753732     2
#> [1495,] 1.753492     2
#> [1496,] 1.746703     2
#> [1497,] 1.746703     2
#> [1498,] 1.736102     2
#> [1499,] 1.735211     2
#> [1500,] 1.735211     2
#> [1501,] 1.747984     2
#> [1502,] 1.747984     2
#> [1503,] 1.737570     2
#> [1504,] 1.737570     2
#> [1505,] 1.734234     2
#> [1506,] 1.750092     2
#> [1507,] 1.731342     2
#> [1508,] 1.755676     2
#> [1509,] 1.736865     2
#> [1510,] 1.737144     2
#> [1511,] 1.745953     2
#> [1512,] 1.740064     2
#> [1513,] 1.739292     2
#> [1514,] 1.753597     2
#> [1515,] 1.748438     2
#> [1516,] 1.748438     2
#> [1517,] 1.750828     2
#> [1518,] 1.754590     2
#> [1519,] 1.745934     2
#> [1520,] 1.745934     2
#> [1521,] 1.750460     2
#> [1522,] 1.742118     2
#> [1523,] 1.742118     2
#> [1524,] 1.742434     2
#> [1525,] 1.744953     2
#> [1526,] 1.744953     2
#> [1527,] 1.744953     2
#> [1528,] 1.744019     2
#> [1529,] 1.747863     2
#> [1530,] 1.725524     2
#> [1531,] 1.748998     2
#> [1532,] 1.751927     2
#> [1533,] 1.751927     2
#> [1534,] 1.747532     2
#> [1535,] 1.736282     2
#> [1536,] 1.745080     2
#> [1537,] 1.745080     2
#> [1538,] 1.736422     2
#> [1539,] 1.736422     2
#> [1540,] 1.737035     2
#> [1541,] 1.737035     2
#> [1542,] 1.744867     2
#> [1543,] 1.741410     2
#> [1544,] 1.741104     2
#> [1545,] 1.741104     2
#> [1546,] 1.741104     2
#> [1547,] 1.741104     2
#> [1548,] 1.741104     2
#> [1549,] 1.731532     2
#> [1550,] 1.754766     2
#> [1551,] 1.727358     2
#> [1552,] 1.717861     2
#> [1553,] 1.717861     2
#> [1554,] 1.735452     2
#> [1555,] 1.753691     2
#> [1556,] 1.741834     2
#> [1557,] 1.752925     2
#> [1558,] 1.752925     2
#> [1559,] 1.748946     2
#> [1560,] 1.753195     2
#> [1561,] 1.729688     2
#> [1562,] 1.754950     2
#> [1563,] 1.754950     2
#> [1564,] 1.754950     2
#> [1565,] 1.757829     2
#> [1566,] 1.754154     2
#> [1567,] 1.732793     2
#> [1568,] 1.755840     2
#> [1569,] 1.755840     2
#> [1570,] 1.739870     2
#> [1571,] 1.739870     2
#> [1572,] 1.746266     2
#> [1573,] 1.746266     2
#> [1574,] 1.748910     2
#> [1575,] 1.743438     2
#> [1576,] 1.743438     2
#> [1577,] 1.743438     2
#> [1578,] 1.743438     2
#> [1579,] 1.753260     2
#> [1580,] 1.730019     2
#> [1581,] 1.730019     2
#> [1582,] 1.730019     2
#> [1583,] 1.746438     2
#> [1584,] 1.746438     2
#> [1585,] 1.725528     2
#> [1586,] 1.750360     2
#> [1587,] 1.746403     2
#> [1588,] 1.737614     2
#> [1589,] 1.735634     2
#> [1590,] 1.740907     2
#> [1591,] 1.749866     2
#> [1592,] 1.747187     2
#> [1593,] 1.745420     2
#> [1594,] 1.745420     2
#> [1595,] 1.745420     2
#> [1596,] 1.754666     2
#> [1597,] 1.733141     2
#> [1598,] 1.733765     2
#> [1599,] 1.733765     2
#> [1600,] 1.757437     2
#> [1601,] 1.733536     2
#> [1602,] 1.740603     2
#> [1603,] 1.736364     2
#> [1604,] 1.736364     2
#> [1605,] 1.733636     2
#> [1606,] 1.744328     2
#> [1607,] 1.741876     2
#> [1608,] 1.741876     2
#> [1609,] 1.741876     2
#> [1610,] 1.741876     2
#> [1611,] 1.741876     2
#> [1612,] 1.741876     2
#> [1613,] 1.741905     2
#> [1614,] 1.738775     2
#> [1615,] 1.748632     2
#> [1616,] 1.747915     2
#> [1617,] 1.741236     2
#> [1618,] 1.739661     2
#> [1619,] 1.752626     2
#> [1620,] 1.752626     2
#> [1621,] 1.752626     2
#> [1622,] 1.746246     2
#> [1623,] 1.726474     2
#> [1624,] 1.728895     2
#> [1625,] 1.740844     2
#> [1626,] 1.736814     2
#> [1627,] 1.745781     2
#> [1628,] 1.745781     2
#> [1629,] 1.745781     2
#> [1630,] 1.755376     2
#> [1631,] 1.742413     2
#> [1632,] 1.754244     2
#> [1633,] 1.754244     2
#> [1634,] 1.740862     2
#> [1635,] 1.743466     2
#> [1636,] 1.743466     2
#> [1637,] 1.743466     2
#> [1638,] 1.737726     2
#> [1639,] 1.731261     2
#> [1640,] 1.729363     2
#> [1641,] 1.738162     2
#> [1642,] 1.738162     2
#> [1643,] 1.738162     2
#> [1644,] 1.736409     2
#> [1645,] 1.736409     2
#> [1646,] 1.744672     2
#> [1647,] 1.744672     2
#> [1648,] 1.732413     2
#> [1649,] 1.740640     2
#> [1650,] 1.742211     2
#> [1651,] 1.732759     2
#> [1652,] 1.732759     2
#> [1653,] 1.750293     2
#> [1654,] 1.750293     2
#> [1655,] 1.740731     2
#> [1656,] 1.745657     2
#> [1657,] 1.749436     2
#> [1658,] 1.749151     2
#> [1659,] 1.743271     2
#> [1660,] 1.737665     2
#> [1661,] 1.753988     2
#> [1662,] 1.747201     2
#> [1663,] 1.746578     2
#> [1664,] 1.746578     2
#> [1665,] 1.746578     2
#> [1666,] 1.746578     2
#> [1667,] 1.746578     2
#> [1668,] 1.740551     2
#> [1669,] 1.738360     2
#> [1670,] 1.737078     2
#> [1671,] 1.741098     2
#> [1672,] 1.736155     2
#> [1673,] 1.754736     2
#> [1674,] 1.734526     2
#> [1675,] 1.739478     2
#> [1676,] 1.750743     2
#> [1677,] 1.750743     2
#> [1678,] 1.750743     2
#> [1679,] 1.750743     2
#> [1680,] 1.733021     2
#> [1681,] 1.733021     2
#> [1682,] 1.733021     2
#> [1683,] 1.752395     2
#> [1684,] 1.727937     2
#> [1685,] 1.755390     2
#> [1686,] 1.746105     2
#> [1687,] 1.746105     2
#> [1688,] 1.738717     2
#> [1689,] 1.738717     2
#> [1690,] 1.738717     2
#> [1691,] 1.738717     2
#> [1692,] 1.742884     2
#> [1693,] 1.753281     2
#> [1694,] 1.733058     2
#> [1695,] 1.733058     2
#> [1696,] 1.733058     2
#> [1697,] 1.744599     2
#> [1698,] 1.722085     2
#> [1699,] 1.726910     2
#> [1700,] 1.742114     2
#> [1701,] 1.754230     2
#> [1702,] 1.761712     2
#> [1703,] 1.724681     2
#> [1704,] 1.747501     2
#> [1705,] 1.747501     2
#> [1706,] 1.747501     2
#> [1707,] 1.740559     2
#> [1708,] 1.748916     2
#> [1709,] 1.748916     2
#> [1710,] 1.740771     2
#> [1711,] 1.740771     2
#> [1712,] 1.740771     2
#> [1713,] 1.740771     2
#> [1714,] 1.735798     2
#> [1715,] 1.742722     2
#> [1716,] 1.738089     2
#> [1717,] 1.738089     2
#> [1718,] 1.754689     2
#> [1719,] 1.722893     2
#> [1720,] 1.745334     2
#> [1721,] 1.751459     2
#> [1722,] 1.747688     2
#> [1723,] 1.755182     2
#> [1724,] 1.737763     2
#> [1725,] 1.735450     2
#> [1726,] 1.733829     2
#> [1727,] 1.735834     2
#> [1728,] 1.740934     2
#> [1729,] 1.740934     2
#> [1730,] 1.740934     2
#> [1731,] 1.740934     2
#> [1732,] 1.758906     2
#> [1733,] 1.719017     2
#> [1734,] 1.747649     2
#> [1735,] 1.747649     2
#> [1736,] 1.747649     2
#> [1737,] 1.747649     2
#> [1738,] 1.727549     2
#> [1739,] 1.756423     2
#> [1740,] 1.749252     2
#> [1741,] 1.734040     2
#> [1742,] 1.745904     2
#> [1743,] 1.745904     2
#> [1744,] 1.745904     2
#> [1745,] 1.745904     2
#> [1746,] 1.745904     2
#> [1747,] 1.745904     2
#> [1748,] 1.733016     2
#> [1749,] 1.737292     2
#> [1750,] 1.730968     2
#> [1751,] 1.743362     2
#> [1752,] 1.737874     2
#> [1753,] 1.737874     2
#> [1754,] 1.755981     2
#> [1755,] 1.755193     2
#> [1756,] 1.740808     2
#> [1757,] 1.738205     2
#> [1758,] 1.751862     2
#> [1759,] 1.748722     2
#> [1760,] 1.752747     2
#> [1761,] 1.759811     2
#> [1762,] 1.744792     2
#> [1763,] 1.741999     2
#> [1764,] 1.741148     2
#> [1765,] 1.754830     2
#> [1766,] 1.739132     2
#> [1767,] 1.739132     2
#> [1768,] 1.739132     2
#> [1769,] 1.744146     2
#> [1770,] 1.744146     2
#> [1771,] 1.735488     2
#> [1772,] 1.746352     2
#> [1773,] 1.746352     2
#> [1774,] 1.735807     2
#> [1775,] 1.735807     2
#> [1776,] 1.735807     2
#> [1777,] 1.735807     2
#> [1778,] 1.743420     2
#> [1779,] 1.743420     2
#> [1780,] 1.743420     2
#> [1781,] 1.743420     2
#> [1782,] 1.743420     2
#> [1783,] 1.743420     2
#> [1784,] 1.744529     2
#> [1785,] 1.744529     2
#> [1786,] 1.744529     2
#> [1787,] 1.714906     2
#> [1788,] 1.739511     2
#> [1789,] 1.751370     2
#> [1790,] 1.751370     2
#> [1791,] 1.751370     2
#> [1792,] 1.753878     2
#> [1793,] 1.738696     2
#> [1794,] 1.735998     2
#> [1795,] 1.740178     2
#> [1796,] 1.740178     2
#> [1797,] 1.743659     2
#> [1798,] 1.747129     2
#> [1799,] 1.752263     2
#> [1800,] 1.736754     2
#> [1801,] 1.736754     2
#> [1802,] 1.743731     2
#> [1803,] 1.734567     2
#> [1804,] 1.734567     2
#> [1805,] 1.737713     2
#> [1806,] 1.737713     2
#> [1807,] 1.737713     2
#> [1808,] 1.746853     2
#> [1809,] 1.740596     2
#> [1810,] 1.735128     2
#> [1811,] 1.734408     2
#> [1812,] 1.731621     2
#> [1813,] 1.739103     2
#> [1814,] 1.742536     2
#> [1815,] 1.740400     2
#> [1816,] 1.742636     2
#> [1817,] 1.748968     2
#> [1818,] 1.736423     2
#> [1819,] 1.736423     2
#> [1820,] 1.736423     2
#> [1821,] 1.747614     2
#> [1822,] 1.747614     2
#> [1823,] 1.747614     2
#> [1824,] 1.747614     2
#> [1825,] 1.747577     2
#> [1826,] 1.755130     2
#> [1827,] 1.748624     2
#> [1828,] 1.747845     2
#> [1829,] 1.747845     2
#> [1830,] 1.740908     2
#> [1831,] 1.735733     2
#> [1832,] 1.735733     2
#> [1833,] 1.735733     2
#> [1834,] 1.743621     2
#> [1835,] 1.740123     2
#> [1836,] 1.743976     2
#> [1837,] 1.743976     2
#> [1838,] 1.743976     2
#> [1839,] 1.737134     2
#> [1840,] 1.736388     2
#> [1841,] 1.736388     2
#> [1842,] 1.736388     2
#> [1843,] 1.734420     2
#> [1844,] 1.734090     2
#> [1845,] 1.734090     2
#> [1846,] 1.743094     2
#> [1847,] 1.743094     2
#> [1848,] 1.743094     2
#> [1849,] 1.744408     2
#> [1850,] 1.744408     2
#> [1851,] 1.748693     2
#> [1852,] 1.748693     2
#> [1853,] 1.748693     2
#> [1854,] 1.732975     2
#> [1855,] 1.748940     2
#> [1856,] 1.748940     2
#> [1857,] 1.734801     2
#> [1858,] 1.734801     2
#> [1859,] 1.754631     2
#> [1860,] 1.734768     2
#> [1861,] 1.741977     2
#> [1862,] 1.741977     2
#> [1863,] 1.741977     2
#> [1864,] 1.750572     2
#> [1865,] 1.741357     2
#> [1866,] 1.741255     2
#> [1867,] 1.741255     2
#> [1868,] 1.739936     2
#> [1869,] 1.754490     2
#> [1870,] 1.751671     2
#> [1871,] 1.751671     2
#> [1872,] 1.753056     2
#> [1873,] 1.747950     2
#> [1874,] 1.742470     2
#> [1875,] 1.742470     2
#> [1876,] 1.742470     2
#> [1877,] 1.741613     2
#> [1878,] 1.746282     2
#> [1879,] 1.746282     2
#> [1880,] 1.746282     2
#> [1881,] 1.741247     2
#> [1882,] 1.750565     2
#> [1883,] 1.729354     2
#> [1884,] 1.734937     2
#> [1885,] 1.756076     2
#> [1886,] 1.754870     2
#> [1887,] 1.740882     2
#> [1888,] 1.740041     2
#> [1889,] 1.740041     2
#> [1890,] 1.764835     2
#> [1891,] 1.742909     2
#> [1892,] 1.742912     2
#> [1893,] 1.760872     2
#> [1894,] 1.749061     2
#> [1895,] 1.749061     2
#> [1896,] 1.749061     2
#> [1897,] 1.747464     2
#> [1898,] 1.747464     2
#> [1899,] 1.747464     2
#> [1900,] 1.745046     2
#> [1901,] 1.745046     2
#> [1902,] 1.741082     2
#> [1903,] 1.746366     2
#> [1904,] 1.739640     2
#> [1905,] 1.750637     2
#> [1906,] 1.745678     2
#> [1907,] 1.745678     2
#> [1908,] 1.752609     2
#> [1909,] 1.752609     2
#> [1910,] 1.738092     2
#> [1911,] 1.736303     2
#> [1912,] 1.736303     2
#> [1913,] 1.743188     2
#> [1914,] 1.743188     2
#> [1915,] 1.743188     2
#> [1916,] 1.739349     2
#> [1917,] 1.738092     2
#> [1918,] 1.745980     2
#> [1919,] 1.744349     2
#> [1920,] 1.742969     2
#> [1921,] 1.742969     2
#> [1922,] 1.742969     2
#> [1923,] 1.742969     2
#> [1924,] 1.742969     2
#> [1925,] 1.742969     2
#> [1926,] 1.750832     2
#> [1927,] 1.735121     2
#> [1928,] 1.735121     2
#> [1929,] 1.735121     2
#> [1930,] 1.758392     2
#> [1931,] 1.735153     2
#> [1932,] 1.736039     2
#> [1933,] 1.746619     2
#> [1934,] 1.754402     2
#> [1935,] 1.723457     2
#> [1936,] 1.727913     2
#> [1937,] 1.727913     2
#> [1938,] 1.726159     2
#> [1939,] 1.731191     2
#> [1940,] 1.747745     2
#> [1941,] 1.747745     2
#> [1942,] 1.747745     2
#> [1943,] 1.734534     2
#> [1944,] 1.743576     2
#> [1945,] 1.745774     2
#> [1946,] 1.744433     2
#> [1947,] 1.748383     2
#> [1948,] 1.747619     2
#> [1949,] 1.734506     2
#> [1950,] 1.731495     2
#> [1951,] 1.731495     2
#> [1952,] 1.742359     2
#> [1953,] 1.741427     2
#> [1954,] 1.741427     2
#> [1955,] 1.731130     2
#> [1956,] 1.752127     2
#> [1957,] 1.752127     2
#> [1958,] 1.729154     2
#> [1959,] 1.765158     2
#> [1960,] 1.746810     2
#> [1961,] 1.746876     2
#> [1962,] 1.744684     2
#> [1963,] 1.750231     2
#> [1964,] 1.750231     2
#> [1965,] 1.736192     2
#> [1966,] 1.743088     2
#> [1967,] 1.743088     2
#> [1968,] 1.745763     2
#> [1969,] 1.745763     2
#> [1970,] 1.745763     2
#> [1971,] 1.745763     2
#> [1972,] 1.745763     2
#> [1973,] 1.748203     2
#> [1974,] 1.734468     2
#> [1975,] 1.734468     2
#> [1976,] 1.734468     2
#> [1977,] 1.746297     2
#> [1978,] 1.739001     2
#> [1979,] 1.754004     2
#> [1980,] 1.754004     2
#> [1981,] 1.756090     2