Burn and thin MCMC chain

Return a burned and thinned trace of the chain.

burnAndThin(trace, burn = 0, thin = 0)

Arguments

trace

either a data.frame or a list of data.frame with all variables in column, as returned by mcmcMh. Accept also an mcmc or mcmc.list object.

burn

proportion of the chain to burn.

thin

number of samples to discard per sample that is being kept

Value

an object with the same format as trace (data.frame or

list of data.frame or mcmc object or mcmc.list

object)

Examples

data(mcmcEpi)
burnAndThin(mcmcEpi1, burn = 100, thin = 4)
#>              R_0 D_inf
#>    [1,] 1.750468     2
#>    [2,] 1.750468     2
#>    [3,] 1.750468     2
#>    [4,] 1.745574     2
#>    [5,] 1.755889     2
#>    [6,] 1.746943     2
#>    [7,] 1.745609     2
#>    [8,] 1.745609     2
#>    [9,] 1.745609     2
#>   [10,] 1.745609     2
#>   [11,] 1.734027     2
#>   [12,] 1.735428     2
#>   [13,] 1.735428     2
#>   [14,] 1.757522     2
#>   [15,] 1.757522     2
#>   [16,] 1.730526     2
#>   [17,] 1.736771     2
#>   [18,] 1.738201     2
#>   [19,] 1.748309     2
#>   [20,] 1.749246     2
#>   [21,] 1.749405     2
#>   [22,] 1.733280     2
#>   [23,] 1.733280     2
#>   [24,] 1.737395     2
#>   [25,] 1.750700     2
#>   [26,] 1.747502     2
#>   [27,] 1.747502     2
#>   [28,] 1.743367     2
#>   [29,] 1.743367     2
#>   [30,] 1.745071     2
#>   [31,] 1.733379     2
#>   [32,] 1.755867     2
#>   [33,] 1.755867     2
#>   [34,] 1.757384     2
#>   [35,] 1.742921     2
#>   [36,] 1.742921     2
#>   [37,] 1.735965     2
#>   [38,] 1.742803     2
#>   [39,] 1.737143     2
#>   [40,] 1.749213     2
#>   [41,] 1.749213     2
#>   [42,] 1.749213     2
#>   [43,] 1.744731     2
#>   [44,] 1.745956     2
#>   [45,] 1.745956     2
#>   [46,] 1.745956     2
#>   [47,] 1.743891     2
#>   [48,] 1.747214     2
#>   [49,] 1.744151     2
#>   [50,] 1.753995     2
#>   [51,] 1.755747     2
#>   [52,] 1.743840     2
#>   [53,] 1.748954     2
#>   [54,] 1.748954     2
#>   [55,] 1.748954     2
#>   [56,] 1.753343     2
#>   [57,] 1.758594     2
#>   [58,] 1.726656     2
#>   [59,] 1.719489     2
#>   [60,] 1.719489     2
#>   [61,] 1.747347     2
#>   [62,] 1.747347     2
#>   [63,] 1.755681     2
#>   [64,] 1.743392     2
#>   [65,] 1.743392     2
#>   [66,] 1.750007     2
#>   [67,] 1.750007     2
#>   [68,] 1.746175     2
#>   [69,] 1.723281     2
#>   [70,] 1.737931     2
#>   [71,] 1.734994     2
#>   [72,] 1.754178     2
#>   [73,] 1.734523     2
#>   [74,] 1.736074     2
#>   [75,] 1.748045     2
#>   [76,] 1.748045     2
#>   [77,] 1.738740     2
#>   [78,] 1.740142     2
#>   [79,] 1.729955     2
#>   [80,] 1.729955     2
#>   [81,] 1.740058     2
#>   [82,] 1.748768     2
#>   [83,] 1.742749     2
#>   [84,] 1.755160     2
#>   [85,] 1.757353     2
#>   [86,] 1.745540     2
#>   [87,] 1.745540     2
#>   [88,] 1.745540     2
#>   [89,] 1.745525     2
#>   [90,] 1.745525     2
#>   [91,] 1.759674     2
#>   [92,] 1.759674     2
#>   [93,] 1.752643     2
#>   [94,] 1.734699     2
#>   [95,] 1.740134     2
#>   [96,] 1.747448     2
#>   [97,] 1.736395     2
#>   [98,] 1.736395     2
#>   [99,] 1.736395     2
#>  [100,] 1.735932     2
#>  [101,] 1.735932     2
#>  [102,] 1.735932     2
#>  [103,] 1.754971     2
#>  [104,] 1.748701     2
#>  [105,] 1.741437     2
#>  [106,] 1.745712     2
#>  [107,] 1.745712     2
#>  [108,] 1.754842     2
#>  [109,] 1.754842     2
#>  [110,] 1.745037     2
#>  [111,] 1.745037     2
#>  [112,] 1.745037     2
#>  [113,] 1.735347     2
#>  [114,] 1.731273     2
#>  [115,] 1.731273     2
#>  [116,] 1.744603     2
#>  [117,] 1.744603     2
#>  [118,] 1.757853     2
#>  [119,] 1.726826     2
#>  [120,] 1.738173     2
#>  [121,] 1.738173     2
#>  [122,] 1.738173     2
#>  [123,] 1.738173     2
#>  [124,] 1.738173     2
#>  [125,] 1.737427     2
#>  [126,] 1.750711     2
#>  [127,] 1.737050     2
#>  [128,] 1.737050     2
#>  [129,] 1.731608     2
#>  [130,] 1.731608     2
#>  [131,] 1.737949     2
#>  [132,] 1.737949     2
#>  [133,] 1.737949     2
#>  [134,] 1.747613     2
#>  [135,] 1.747613     2
#>  [136,] 1.747613     2
#>  [137,] 1.735024     2
#>  [138,] 1.730233     2
#>  [139,] 1.743631     2
#>  [140,] 1.743631     2
#>  [141,] 1.737284     2
#>  [142,] 1.748604     2
#>  [143,] 1.748604     2
#>  [144,] 1.738779     2
#>  [145,] 1.737890     2
#>  [146,] 1.739394     2
#>  [147,] 1.739394     2
#>  [148,] 1.739394     2
#>  [149,] 1.750427     2
#>  [150,] 1.750427     2
#>  [151,] 1.750427     2
#>  [152,] 1.750427     2
#>  [153,] 1.724145     2
#>  [154,] 1.758472     2
#>  [155,] 1.734711     2
#>  [156,] 1.731315     2
#>  [157,] 1.741557     2
#>  [158,] 1.741436     2
#>  [159,] 1.749423     2
#>  [160,] 1.749423     2
#>  [161,] 1.732650     2
#>  [162,] 1.758949     2
#>  [163,] 1.738714     2
#>  [164,] 1.746924     2
#>  [165,] 1.746924     2
#>  [166,] 1.734594     2
#>  [167,] 1.734594     2
#>  [168,] 1.730309     2
#>  [169,] 1.729300     2
#>  [170,] 1.752147     2
#>  [171,] 1.733937     2
#>  [172,] 1.733937     2
#>  [173,] 1.748246     2
#>  [174,] 1.734647     2
#>  [175,] 1.756827     2
#>  [176,] 1.749864     2
#>  [177,] 1.749864     2
#>  [178,] 1.749864     2
#>  [179,] 1.752508     2
#>  [180,] 1.752508     2
#>  [181,] 1.745135     2
#>  [182,] 1.736265     2
#>  [183,] 1.757376     2
#>  [184,] 1.732422     2
#>  [185,] 1.739609     2
#>  [186,] 1.739609     2
#>  [187,] 1.742968     2
#>  [188,] 1.742968     2
#>  [189,] 1.735854     2
#>  [190,] 1.753338     2
#>  [191,] 1.753338     2
#>  [192,] 1.753338     2
#>  [193,] 1.728538     2
#>  [194,] 1.730450     2
#>  [195,] 1.733334     2
#>  [196,] 1.742309     2
#>  [197,] 1.742309     2
#>  [198,] 1.742309     2
#>  [199,] 1.734544     2
#>  [200,] 1.749514     2
#>  [201,] 1.741173     2
#>  [202,] 1.741173     2
#>  [203,] 1.753665     2
#>  [204,] 1.732169     2
#>  [205,] 1.734275     2
#>  [206,] 1.753525     2
#>  [207,] 1.753525     2
#>  [208,] 1.734997     2
#>  [209,] 1.755930     2
#>  [210,] 1.721760     2
#>  [211,] 1.733738     2
#>  [212,] 1.732895     2
#>  [213,] 1.733734     2
#>  [214,] 1.750734     2
#>  [215,] 1.745330     2
#>  [216,] 1.735736     2
#>  [217,] 1.725174     2
#>  [218,] 1.750061     2
#>  [219,] 1.750061     2
#>  [220,] 1.731412     2
#>  [221,] 1.744566     2
#>  [222,] 1.736841     2
#>  [223,] 1.755770     2
#>  [224,] 1.755770     2
#>  [225,] 1.755770     2
#>  [226,] 1.745398     2
#>  [227,] 1.745398     2
#>  [228,] 1.745398     2
#>  [229,] 1.738028     2
#>  [230,] 1.742492     2
#>  [231,] 1.742469     2
#>  [232,] 1.742469     2
#>  [233,] 1.757736     2
#>  [234,] 1.738645     2
#>  [235,] 1.738645     2
#>  [236,] 1.741946     2
#>  [237,] 1.735190     2
#>  [238,] 1.743075     2
#>  [239,] 1.737393     2
#>  [240,] 1.742800     2
#>  [241,] 1.742800     2
#>  [242,] 1.742800     2
#>  [243,] 1.742800     2
#>  [244,] 1.742338     2
#>  [245,] 1.742338     2
#>  [246,] 1.721545     2
#>  [247,] 1.741049     2
#>  [248,] 1.735264     2
#>  [249,] 1.742736     2
#>  [250,] 1.735483     2
#>  [251,] 1.735483     2
#>  [252,] 1.746350     2
#>  [253,] 1.755113     2
#>  [254,] 1.758336     2
#>  [255,] 1.738545     2
#>  [256,] 1.738545     2
#>  [257,] 1.734602     2
#>  [258,] 1.735036     2
#>  [259,] 1.735604     2
#>  [260,] 1.735604     2
#>  [261,] 1.730226     2
#>  [262,] 1.730226     2
#>  [263,] 1.730226     2
#>  [264,] 1.747084     2
#>  [265,] 1.754263     2
#>  [266,] 1.736582     2
#>  [267,] 1.739244     2
#>  [268,] 1.739244     2
#>  [269,] 1.739244     2
#>  [270,] 1.748579     2
#>  [271,] 1.748579     2
#>  [272,] 1.732208     2
#>  [273,] 1.732948     2
#>  [274,] 1.745372     2
#>  [275,] 1.745372     2
#>  [276,] 1.739341     2
#>  [277,] 1.739341     2
#>  [278,] 1.745836     2
#>  [279,] 1.752218     2
#>  [280,] 1.735366     2
#>  [281,] 1.735366     2
#>  [282,] 1.735366     2
#>  [283,] 1.764853     2
#>  [284,] 1.764853     2
#>  [285,] 1.764853     2
#>  [286,] 1.759524     2
#>  [287,] 1.759524     2
#>  [288,] 1.759524     2
#>  [289,] 1.755506     2
#>  [290,] 1.743825     2
#>  [291,] 1.745205     2
#>  [292,] 1.745205     2
#>  [293,] 1.754208     2
#>  [294,] 1.740379     2
#>  [295,] 1.740379     2
#>  [296,] 1.740104     2
#>  [297,] 1.750898     2
#>  [298,] 1.750898     2
#>  [299,] 1.748153     2
#>  [300,] 1.742524     2
#>  [301,] 1.742536     2
#>  [302,] 1.741238     2
#>  [303,] 1.744287     2
#>  [304,] 1.730070     2
#>  [305,] 1.742092     2
#>  [306,] 1.746248     2
#>  [307,] 1.746248     2
#>  [308,] 1.746248     2
#>  [309,] 1.765767     2
#>  [310,] 1.765767     2
#>  [311,] 1.759093     2
#>  [312,] 1.744619     2
#>  [313,] 1.746313     2
#>  [314,] 1.746313     2
#>  [315,] 1.740152     2
#>  [316,] 1.746759     2
#>  [317,] 1.746759     2
#>  [318,] 1.746152     2
#>  [319,] 1.746152     2
#>  [320,] 1.723960     2
#>  [321,] 1.744003     2
#>  [322,] 1.744003     2
#>  [323,] 1.744003     2
#>  [324,] 1.749576     2
#>  [325,] 1.744906     2
#>  [326,] 1.744906     2
#>  [327,] 1.744906     2
#>  [328,] 1.741629     2
#>  [329,] 1.747052     2
#>  [330,] 1.747136     2
#>  [331,] 1.742168     2
#>  [332,] 1.741748     2
#>  [333,] 1.741748     2
#>  [334,] 1.753304     2
#>  [335,] 1.739023     2
#>  [336,] 1.739023     2
#>  [337,] 1.739023     2
#>  [338,] 1.747997     2
#>  [339,] 1.739193     2
#>  [340,] 1.747996     2
#>  [341,] 1.747996     2
#>  [342,] 1.747996     2
#>  [343,] 1.734633     2
#>  [344,] 1.734633     2
#>  [345,] 1.738014     2
#>  [346,] 1.738014     2
#>  [347,] 1.740947     2
#>  [348,] 1.740947     2
#>  [349,] 1.750389     2
#>  [350,] 1.745779     2
#>  [351,] 1.745779     2
#>  [352,] 1.739567     2
#>  [353,] 1.740468     2
#>  [354,] 1.740468     2
#>  [355,] 1.747239     2
#>  [356,] 1.747239     2
#>  [357,] 1.747239     2
#>  [358,] 1.751772     2
#>  [359,] 1.740814     2
#>  [360,] 1.740814     2
#>  [361,] 1.744468     2
#>  [362,] 1.744468     2
#>  [363,] 1.744468     2
#>  [364,] 1.744468     2
#>  [365,] 1.738293     2
#>  [366,] 1.738293     2
#>  [367,] 1.738293     2
#>  [368,] 1.726890     2
#>  [369,] 1.739488     2
#>  [370,] 1.740705     2
#>  [371,] 1.734031     2
#>  [372,] 1.745762     2
#>  [373,] 1.741070     2
#>  [374,] 1.741070     2
#>  [375,] 1.721994     2
#>  [376,] 1.742616     2
#>  [377,] 1.733268     2
#>  [378,] 1.751049     2
#>  [379,] 1.750382     2
#>  [380,] 1.729591     2
#>  [381,] 1.739451     2
#>  [382,] 1.739451     2
#>  [383,] 1.748225     2
#>  [384,] 1.748225     2
#>  [385,] 1.733845     2
#>  [386,] 1.739653     2
#>  [387,] 1.739653     2
#>  [388,] 1.762775     2
#>  [389,] 1.744550     2
#>  [390,] 1.744550     2
#>  [391,] 1.749261     2
#>  [392,] 1.749261     2
#>  [393,] 1.744077     2
#>  [394,] 1.757849     2
#>  [395,] 1.762939     2
#>  [396,] 1.746668     2
#>  [397,] 1.746668     2
#>  [398,] 1.747998     2
#>  [399,] 1.747998     2
#>  [400,] 1.747998     2
#>  [401,] 1.747209     2
#>  [402,] 1.748209     2
#>  [403,] 1.745551     2
#>  [404,] 1.745551     2
#>  [405,] 1.743429     2
#>  [406,] 1.732680     2
#>  [407,] 1.748805     2
#>  [408,] 1.748805     2
#>  [409,] 1.748805     2
#>  [410,] 1.748805     2
#>  [411,] 1.742036     2
#>  [412,] 1.737502     2
#>  [413,] 1.741420     2
#>  [414,] 1.741420     2
#>  [415,] 1.741420     2
#>  [416,] 1.743513     2
#>  [417,] 1.756911     2
#>  [418,] 1.755895     2
#>  [419,] 1.755895     2
#>  [420,] 1.742009     2
#>  [421,] 1.738335     2
#>  [422,] 1.756803     2
#>  [423,] 1.756803     2
#>  [424,] 1.748457     2
#>  [425,] 1.731590     2
#>  [426,] 1.731590     2
#>  [427,] 1.731590     2
#>  [428,] 1.731590     2
#>  [429,] 1.731590     2
#>  [430,] 1.745327     2
#>  [431,] 1.747454     2
#>  [432,] 1.747454     2
#>  [433,] 1.747454     2
#>  [434,] 1.735597     2
#>  [435,] 1.744433     2
#>  [436,] 1.751262     2
#>  [437,] 1.751801     2
#>  [438,] 1.751805     2
#>  [439,] 1.725357     2
#>  [440,] 1.734095     2
#>  [441,] 1.753425     2
#>  [442,] 1.753425     2
#>  [443,] 1.745108     2
#>  [444,] 1.750269     2
#>  [445,] 1.750269     2
#>  [446,] 1.736910     2
#>  [447,] 1.741743     2
#>  [448,] 1.741743     2
#>  [449,] 1.741743     2
#>  [450,] 1.741743     2
#>  [451,] 1.741743     2
#>  [452,] 1.741743     2
#>  [453,] 1.749201     2
#>  [454,] 1.749201     2
#>  [455,] 1.736966     2
#>  [456,] 1.736966     2
#>  [457,] 1.736966     2
#>  [458,] 1.752014     2
#>  [459,] 1.752014     2
#>  [460,] 1.738604     2
#>  [461,] 1.738604     2
#>  [462,] 1.743814     2
#>  [463,] 1.733726     2
#>  [464,] 1.733726     2
#>  [465,] 1.736533     2
#>  [466,] 1.736533     2
#>  [467,] 1.743963     2
#>  [468,] 1.749872     2
#>  [469,] 1.749872     2
#>  [470,] 1.744974     2
#>  [471,] 1.739665     2
#>  [472,] 1.741820     2
#>  [473,] 1.741820     2
#>  [474,] 1.757495     2
#>  [475,] 1.734900     2
#>  [476,] 1.751639     2
#>  [477,] 1.751639     2
#>  [478,] 1.751639     2
#>  [479,] 1.756067     2
#>  [480,] 1.750519     2
#>  [481,] 1.746148     2
#>  [482,] 1.740493     2
#>  [483,] 1.745208     2
#>  [484,] 1.745208     2
#>  [485,] 1.745208     2
#>  [486,] 1.746430     2
#>  [487,] 1.752150     2
#>  [488,] 1.752150     2
#>  [489,] 1.752150     2
#>  [490,] 1.732754     2
#>  [491,] 1.732754     2
#>  [492,] 1.732754     2
#>  [493,] 1.734499     2
#>  [494,] 1.758730     2
#>  [495,] 1.746710     2
#>  [496,] 1.742500     2
#>  [497,] 1.738157     2
#>  [498,] 1.737377     2
#>  [499,] 1.737377     2
#>  [500,] 1.750461     2
#>  [501,] 1.746347     2
#>  [502,] 1.746347     2
#>  [503,] 1.746347     2
#>  [504,] 1.746347     2
#>  [505,] 1.733339     2
#>  [506,] 1.731570     2
#>  [507,] 1.748485     2
#>  [508,] 1.740800     2
#>  [509,] 1.737551     2
#>  [510,] 1.743828     2
#>  [511,] 1.743828     2
#>  [512,] 1.751815     2
#>  [513,] 1.733420     2
#>  [514,] 1.733420     2
#>  [515,] 1.751370     2
#>  [516,] 1.733081     2
#>  [517,] 1.737346     2
#>  [518,] 1.761030     2
#>  [519,] 1.732425     2
#>  [520,] 1.732425     2
#>  [521,] 1.739925     2
#>  [522,] 1.746811     2
#>  [523,] 1.746811     2
#>  [524,] 1.746811     2
#>  [525,] 1.746811     2
#>  [526,] 1.746811     2
#>  [527,] 1.738820     2
#>  [528,] 1.744235     2
#>  [529,] 1.744235     2
#>  [530,] 1.743538     2
#>  [531,] 1.731815     2
#>  [532,] 1.731815     2
#>  [533,] 1.743502     2
#>  [534,] 1.743502     2
#>  [535,] 1.743502     2
#>  [536,] 1.736184     2
#>  [537,] 1.736184     2
#>  [538,] 1.750000     2
#>  [539,] 1.755451     2
#>  [540,] 1.755451     2
#>  [541,] 1.735249     2
#>  [542,] 1.736023     2
#>  [543,] 1.736023     2
#>  [544,] 1.733886     2
#>  [545,] 1.731956     2
#>  [546,] 1.752361     2
#>  [547,] 1.752361     2
#>  [548,] 1.749205     2
#>  [549,] 1.744760     2
#>  [550,] 1.741338     2
#>  [551,] 1.756389     2
#>  [552,] 1.746895     2
#>  [553,] 1.740471     2
#>  [554,] 1.754162     2
#>  [555,] 1.732932     2
#>  [556,] 1.747339     2
#>  [557,] 1.747339     2
#>  [558,] 1.750894     2
#>  [559,] 1.750894     2
#>  [560,] 1.750894     2
#>  [561,] 1.745928     2
#>  [562,] 1.734563     2
#>  [563,] 1.737096     2
#>  [564,] 1.752638     2
#>  [565,] 1.750791     2
#>  [566,] 1.737897     2
#>  [567,] 1.742409     2
#>  [568,] 1.742409     2
#>  [569,] 1.744647     2
#>  [570,] 1.746567     2
#>  [571,] 1.746567     2
#>  [572,] 1.746567     2
#>  [573,] 1.738187     2
#>  [574,] 1.754945     2
#>  [575,] 1.736824     2
#>  [576,] 1.722506     2
#>  [577,] 1.726661     2
#>  [578,] 1.737144     2
#>  [579,] 1.747349     2
#>  [580,] 1.744526     2
#>  [581,] 1.744526     2
#>  [582,] 1.736477     2
#>  [583,] 1.753515     2
#>  [584,] 1.753515     2
#>  [585,] 1.753515     2
#>  [586,] 1.731319     2
#>  [587,] 1.749894     2
#>  [588,] 1.745974     2
#>  [589,] 1.737624     2
#>  [590,] 1.736793     2
#>  [591,] 1.754782     2
#>  [592,] 1.759011     2
#>  [593,] 1.736713     2
#>  [594,] 1.736713     2
#>  [595,] 1.736713     2
#>  [596,] 1.746437     2
#>  [597,] 1.733711     2
#>  [598,] 1.735325     2
#>  [599,] 1.753741     2
#>  [600,] 1.735317     2
#>  [601,] 1.743305     2
#>  [602,] 1.743305     2
#>  [603,] 1.748847     2
#>  [604,] 1.748847     2
#>  [605,] 1.744455     2
#>  [606,] 1.744455     2
#>  [607,] 1.744455     2
#>  [608,] 1.733802     2
#>  [609,] 1.753391     2
#>  [610,] 1.753391     2
#>  [611,] 1.754493     2
#>  [612,] 1.754493     2
#>  [613,] 1.745961     2
#>  [614,] 1.737956     2
#>  [615,] 1.742938     2
#>  [616,] 1.727184     2
#>  [617,] 1.743704     2
#>  [618,] 1.760820     2
#>  [619,] 1.730678     2
#>  [620,] 1.748016     2
#>  [621,] 1.748016     2
#>  [622,] 1.748016     2
#>  [623,] 1.732229     2
#>  [624,] 1.732229     2
#>  [625,] 1.732229     2
#>  [626,] 1.746987     2
#>  [627,] 1.746397     2
#>  [628,] 1.746397     2
#>  [629,] 1.757100     2
#>  [630,] 1.731599     2
#>  [631,] 1.749735     2
#>  [632,] 1.746439     2
#>  [633,] 1.746439     2
#>  [634,] 1.743033     2
#>  [635,] 1.735757     2
#>  [636,] 1.740114     2
#>  [637,] 1.740114     2
#>  [638,] 1.732636     2
#>  [639,] 1.747016     2
#>  [640,] 1.738336     2
#>  [641,] 1.742898     2
#>  [642,] 1.742898     2
#>  [643,] 1.742898     2
#>  [644,] 1.742898     2
#>  [645,] 1.742445     2
#>  [646,] 1.757412     2
#>  [647,] 1.749235     2
#>  [648,] 1.749235     2
#>  [649,] 1.728426     2
#>  [650,] 1.735177     2
#>  [651,] 1.748155     2
#>  [652,] 1.734256     2
#>  [653,] 1.759051     2
#>  [654,] 1.759051     2
#>  [655,] 1.759051     2
#>  [656,] 1.768554     2
#>  [657,] 1.739701     2
#>  [658,] 1.751752     2
#>  [659,] 1.741487     2
#>  [660,] 1.752442     2
#>  [661,] 1.753793     2
#>  [662,] 1.753793     2
#>  [663,] 1.735310     2
#>  [664,] 1.735310     2
#>  [665,] 1.755511     2
#>  [666,] 1.776537     2
#>  [667,] 1.776537     2
#>  [668,] 1.764450     2
#>  [669,] 1.718074     2
#>  [670,] 1.718074     2
#>  [671,] 1.748609     2
#>  [672,] 1.748609     2
#>  [673,] 1.741911     2
#>  [674,] 1.741911     2
#>  [675,] 1.741911     2
#>  [676,] 1.739165     2
#>  [677,] 1.748441     2
#>  [678,] 1.748441     2
#>  [679,] 1.735238     2
#>  [680,] 1.736377     2
#>  [681,] 1.755336     2
#>  [682,] 1.753311     2
#>  [683,] 1.731294     2
#>  [684,] 1.757725     2
#>  [685,] 1.723086     2
#>  [686,] 1.740895     2
#>  [687,] 1.740895     2
#>  [688,] 1.740895     2
#>  [689,] 1.745460     2
#>  [690,] 1.731953     2
#>  [691,] 1.740078     2
#>  [692,] 1.740078     2
#>  [693,] 1.740078     2
#>  [694,] 1.735064     2
#>  [695,] 1.735064     2
#>  [696,] 1.739228     2
#>  [697,] 1.739228     2
#>  [698,] 1.734362     2
#>  [699,] 1.734362     2
#>  [700,] 1.738942     2
#>  [701,] 1.744191     2
#>  [702,] 1.744191     2
#>  [703,] 1.750451     2
#>  [704,] 1.750451     2
#>  [705,] 1.750451     2
#>  [706,] 1.733533     2
#>  [707,] 1.754408     2
#>  [708,] 1.754408     2
#>  [709,] 1.754408     2
#>  [710,] 1.754408     2
#>  [711,] 1.752266     2
#>  [712,] 1.750106     2
#>  [713,] 1.748441     2
#>  [714,] 1.748441     2
#>  [715,] 1.731204     2
#>  [716,] 1.731204     2
#>  [717,] 1.756504     2
#>  [718,] 1.757433     2
#>  [719,] 1.742340     2
#>  [720,] 1.749148     2
#>  [721,] 1.749148     2
#>  [722,] 1.733458     2
#>  [723,] 1.733458     2
#>  [724,] 1.745826     2
#>  [725,] 1.745826     2
#>  [726,] 1.751535     2
#>  [727,] 1.749115     2
#>  [728,] 1.757318     2
#>  [729,] 1.754280     2
#>  [730,] 1.746798     2
#>  [731,] 1.746798     2
#>  [732,] 1.746798     2
#>  [733,] 1.731637     2
#>  [734,] 1.746421     2
#>  [735,] 1.748815     2
#>  [736,] 1.749861     2
#>  [737,] 1.735721     2
#>  [738,] 1.742424     2
#>  [739,] 1.742424     2
#>  [740,] 1.741551     2
#>  [741,] 1.742229     2
#>  [742,] 1.739609     2
#>  [743,] 1.739609     2
#>  [744,] 1.746049     2
#>  [745,] 1.746049     2
#>  [746,] 1.755461     2
#>  [747,] 1.728556     2
#>  [748,] 1.728556     2
#>  [749,] 1.728556     2
#>  [750,] 1.741824     2
#>  [751,] 1.750844     2
#>  [752,] 1.746593     2
#>  [753,] 1.746593     2
#>  [754,] 1.748123     2
#>  [755,] 1.732845     2
#>  [756,] 1.732845     2
#>  [757,] 1.748595     2
#>  [758,] 1.731817     2
#>  [759,] 1.743118     2
#>  [760,] 1.749886     2
#>  [761,] 1.743237     2
#>  [762,] 1.740130     2
#>  [763,] 1.740130     2
#>  [764,] 1.740130     2
#>  [765,] 1.740130     2
#>  [766,] 1.734172     2
#>  [767,] 1.741701     2
#>  [768,] 1.741701     2
#>  [769,] 1.745365     2
#>  [770,] 1.743758     2
#>  [771,] 1.743758     2
#>  [772,] 1.743758     2
#>  [773,] 1.740730     2
#>  [774,] 1.740730     2
#>  [775,] 1.752827     2
#>  [776,] 1.752827     2
#>  [777,] 1.752827     2
#>  [778,] 1.743675     2
#>  [779,] 1.752769     2
#>  [780,] 1.728310     2
#>  [781,] 1.762407     2
#>  [782,] 1.754458     2
#>  [783,] 1.743171     2
#>  [784,] 1.737769     2
#>  [785,] 1.732640     2
#>  [786,] 1.744450     2
#>  [787,] 1.744450     2
#>  [788,] 1.750675     2
#>  [789,] 1.748320     2
#>  [790,] 1.732648     2
#>  [791,] 1.758554     2
#>  [792,] 1.758554     2
#>  [793,] 1.758554     2
#>  [794,] 1.758554     2
#>  [795,] 1.758554     2
#>  [796,] 1.745660     2
#>  [797,] 1.731566     2
#>  [798,] 1.734838     2
#>  [799,] 1.734838     2
#>  [800,] 1.734838     2
#>  [801,] 1.734838     2
#>  [802,] 1.756887     2
#>  [803,] 1.744242     2
#>  [804,] 1.731926     2
#>  [805,] 1.731926     2
#>  [806,] 1.731926     2
#>  [807,] 1.753362     2
#>  [808,] 1.743922     2
#>  [809,] 1.734579     2
#>  [810,] 1.733005     2
#>  [811,] 1.733005     2
#>  [812,] 1.733005     2
#>  [813,] 1.725896     2
#>  [814,] 1.742292     2
#>  [815,] 1.751963     2
#>  [816,] 1.737229     2
#>  [817,] 1.747490     2
#>  [818,] 1.747490     2
#>  [819,] 1.740734     2
#>  [820,] 1.731485     2
#>  [821,] 1.723194     2
#>  [822,] 1.744302     2
#>  [823,] 1.742042     2
#>  [824,] 1.748498     2
#>  [825,] 1.758231     2
#>  [826,] 1.729513     2
#>  [827,] 1.750887     2
#>  [828,] 1.732075     2
#>  [829,] 1.731499     2
#>  [830,] 1.736909     2
#>  [831,] 1.736664     2
#>  [832,] 1.733041     2
#>  [833,] 1.733041     2
#>  [834,] 1.733041     2
#>  [835,] 1.742125     2
#>  [836,] 1.742125     2
#>  [837,] 1.742125     2
#>  [838,] 1.742125     2
#>  [839,] 1.742125     2
#>  [840,] 1.742125     2
#>  [841,] 1.742346     2
#>  [842,] 1.742346     2
#>  [843,] 1.742346     2
#>  [844,] 1.742346     2
#>  [845,] 1.731715     2
#>  [846,] 1.731715     2
#>  [847,] 1.739517     2
#>  [848,] 1.741329     2
#>  [849,] 1.744843     2
#>  [850,] 1.744843     2
#>  [851,] 1.744843     2
#>  [852,] 1.741896     2
#>  [853,] 1.741896     2
#>  [854,] 1.741896     2
#>  [855,] 1.741896     2
#>  [856,] 1.746815     2
#>  [857,] 1.746386     2
#>  [858,] 1.749771     2
#>  [859,] 1.749771     2
#>  [860,] 1.752600     2
#>  [861,] 1.752600     2
#>  [862,] 1.752600     2
#>  [863,] 1.732867     2
#>  [864,] 1.753006     2
#>  [865,] 1.753006     2
#>  [866,] 1.743568     2
#>  [867,] 1.738704     2
#>  [868,] 1.742542     2
#>  [869,] 1.742542     2
#>  [870,] 1.742542     2
#>  [871,] 1.749872     2
#>  [872,] 1.745116     2
#>  [873,] 1.745116     2
#>  [874,] 1.745116     2
#>  [875,] 1.749534     2
#>  [876,] 1.740733     2
#>  [877,] 1.737891     2
#>  [878,] 1.737891     2
#>  [879,] 1.746756     2
#>  [880,] 1.743208     2
#>  [881,] 1.743208     2
#>  [882,] 1.743208     2
#>  [883,] 1.743208     2
#>  [884,] 1.743208     2
#>  [885,] 1.736034     2
#>  [886,] 1.736034     2
#>  [887,] 1.732885     2
#>  [888,] 1.743311     2
#>  [889,] 1.740881     2
#>  [890,] 1.733345     2
#>  [891,] 1.748002     2
#>  [892,] 1.728786     2
#>  [893,] 1.760766     2
#>  [894,] 1.760766     2
#>  [895,] 1.750103     2
#>  [896,] 1.746072     2
#>  [897,] 1.746072     2
#>  [898,] 1.749399     2
#>  [899,] 1.736423     2
#>  [900,] 1.753553     2
#>  [901,] 1.751101     2
#>  [902,] 1.749324     2
#>  [903,] 1.747607     2
#>  [904,] 1.748208     2
#>  [905,] 1.748208     2
#>  [906,] 1.751623     2
#>  [907,] 1.752713     2
#>  [908,] 1.752713     2
#>  [909,] 1.752713     2
#>  [910,] 1.748788     2
#>  [911,] 1.732967     2
#>  [912,] 1.731586     2
#>  [913,] 1.742313     2
#>  [914,] 1.742313     2
#>  [915,] 1.742313     2
#>  [916,] 1.742313     2
#>  [917,] 1.736449     2
#>  [918,] 1.736449     2
#>  [919,] 1.757912     2
#>  [920,] 1.738657     2
#>  [921,] 1.743136     2
#>  [922,] 1.743136     2
#>  [923,] 1.750276     2
#>  [924,] 1.746966     2
#>  [925,] 1.738603     2
#>  [926,] 1.744291     2
#>  [927,] 1.734585     2
#>  [928,] 1.757573     2
#>  [929,] 1.750836     2
#>  [930,] 1.750836     2
#>  [931,] 1.750836     2
#>  [932,] 1.750836     2
#>  [933,] 1.744377     2
#>  [934,] 1.733670     2
#>  [935,] 1.745011     2
#>  [936,] 1.745011     2
#>  [937,] 1.745011     2
#>  [938,] 1.743230     2
#>  [939,] 1.743230     2
#>  [940,] 1.743230     2
#>  [941,] 1.743230     2
#>  [942,] 1.729653     2
#>  [943,] 1.729653     2
#>  [944,] 1.755780     2
#>  [945,] 1.741466     2
#>  [946,] 1.756950     2
#>  [947,] 1.756950     2
#>  [948,] 1.747836     2
#>  [949,] 1.755214     2
#>  [950,] 1.731999     2
#>  [951,] 1.755355     2
#>  [952,] 1.749147     2
#>  [953,] 1.757651     2
#>  [954,] 1.757651     2
#>  [955,] 1.757651     2
#>  [956,] 1.757651     2
#>  [957,] 1.756991     2
#>  [958,] 1.739381     2
#>  [959,] 1.739973     2
#>  [960,] 1.739973     2
#>  [961,] 1.763564     2
#>  [962,] 1.730210     2
#>  [963,] 1.743675     2
#>  [964,] 1.743675     2
#>  [965,] 1.734571     2
#>  [966,] 1.734571     2
#>  [967,] 1.734571     2
#>  [968,] 1.736515     2
#>  [969,] 1.744636     2
#>  [970,] 1.745052     2
#>  [971,] 1.745052     2
#>  [972,] 1.745052     2
#>  [973,] 1.740346     2
#>  [974,] 1.742951     2
#>  [975,] 1.742951     2
#>  [976,] 1.742951     2
#>  [977,] 1.742951     2
#>  [978,] 1.751724     2
#>  [979,] 1.751724     2
#>  [980,] 1.764740     2
#>  [981,] 1.758650     2
#>  [982,] 1.758650     2
#>  [983,] 1.751072     2
#>  [984,] 1.734728     2
#>  [985,] 1.734728     2
#>  [986,] 1.756523     2
#>  [987,] 1.756523     2
#>  [988,] 1.756523     2
#>  [989,] 1.736452     2
#>  [990,] 1.757006     2
#>  [991,] 1.757006     2
#>  [992,] 1.739804     2
#>  [993,] 1.743719     2
#>  [994,] 1.740247     2
#>  [995,] 1.740454     2
#>  [996,] 1.740454     2
#>  [997,] 1.726814     2
#>  [998,] 1.726814     2
#>  [999,] 1.755479     2
#> [1000,] 1.756478     2
#> [1001,] 1.741672     2
#> [1002,] 1.741672     2
#> [1003,] 1.722826     2
#> [1004,] 1.740091     2
#> [1005,] 1.749150     2
#> [1006,] 1.735405     2
#> [1007,] 1.743379     2
#> [1008,] 1.743379     2
#> [1009,] 1.750965     2
#> [1010,] 1.748415     2
#> [1011,] 1.748415     2
#> [1012,] 1.732973     2
#> [1013,] 1.732973     2
#> [1014,] 1.725335     2
#> [1015,] 1.739940     2
#> [1016,] 1.739940     2
#> [1017,] 1.739940     2
#> [1018,] 1.752521     2
#> [1019,] 1.752521     2
#> [1020,] 1.752521     2
#> [1021,] 1.752521     2
#> [1022,] 1.754127     2
#> [1023,] 1.734097     2
#> [1024,] 1.735108     2
#> [1025,] 1.734738     2
#> [1026,] 1.734738     2
#> [1027,] 1.751540     2
#> [1028,] 1.733509     2
#> [1029,] 1.743753     2
#> [1030,] 1.743753     2
#> [1031,] 1.735069     2
#> [1032,] 1.730824     2
#> [1033,] 1.751247     2
#> [1034,] 1.755920     2
#> [1035,] 1.738336     2
#> [1036,] 1.758929     2
#> [1037,] 1.758929     2
#> [1038,] 1.758929     2
#> [1039,] 1.758929     2
#> [1040,] 1.724028     2
#> [1041,] 1.725914     2
#> [1042,] 1.742464     2
#> [1043,] 1.735515     2
#> [1044,] 1.748791     2
#> [1045,] 1.735890     2
#> [1046,] 1.755733     2
#> [1047,] 1.763767     2
#> [1048,] 1.743093     2
#> [1049,] 1.742798     2
#> [1050,] 1.732301     2
#> [1051,] 1.728234     2
#> [1052,] 1.729366     2
#> [1053,] 1.729366     2
#> [1054,] 1.739159     2
#> [1055,] 1.749650     2
#> [1056,] 1.735330     2
#> [1057,] 1.756713     2
#> [1058,] 1.743320     2
#> [1059,] 1.748524     2
#> [1060,] 1.736557     2
#> [1061,] 1.741283     2
#> [1062,] 1.739546     2
#> [1063,] 1.739546     2
#> [1064,] 1.746186     2
#> [1065,] 1.741170     2
#> [1066,] 1.749868     2
#> [1067,] 1.757035     2
#> [1068,] 1.745907     2
#> [1069,] 1.745907     2
#> [1070,] 1.758558     2
#> [1071,] 1.744815     2
#> [1072,] 1.743606     2
#> [1073,] 1.743606     2
#> [1074,] 1.743606     2
#> [1075,] 1.743606     2
#> [1076,] 1.753318     2
#> [1077,] 1.724327     2
#> [1078,] 1.724327     2
#> [1079,] 1.724327     2
#> [1080,] 1.727704     2
#> [1081,] 1.746666     2
#> [1082,] 1.740911     2
#> [1083,] 1.740911     2
#> [1084,] 1.757300     2
#> [1085,] 1.748976     2
#> [1086,] 1.749430     2
#> [1087,] 1.749430     2
#> [1088,] 1.753482     2
#> [1089,] 1.752029     2
#> [1090,] 1.752896     2
#> [1091,] 1.727700     2
#> [1092,] 1.758002     2
#> [1093,] 1.741552     2
#> [1094,] 1.753273     2
#> [1095,] 1.743505     2
#> [1096,] 1.730187     2
#> [1097,] 1.732651     2
#> [1098,] 1.725852     2
#> [1099,] 1.742286     2
#> [1100,] 1.742286     2
#> [1101,] 1.750552     2
#> [1102,] 1.750451     2
#> [1103,] 1.739874     2
#> [1104,] 1.756493     2
#> [1105,] 1.746206     2
#> [1106,] 1.746206     2
#> [1107,] 1.749862     2
#> [1108,] 1.747034     2
#> [1109,] 1.747241     2
#> [1110,] 1.731339     2
#> [1111,] 1.733709     2
#> [1112,] 1.733709     2
#> [1113,] 1.743759     2
#> [1114,] 1.743759     2
#> [1115,] 1.731024     2
#> [1116,] 1.741843     2
#> [1117,] 1.752394     2
#> [1118,] 1.752394     2
#> [1119,] 1.734741     2
#> [1120,] 1.736031     2
#> [1121,] 1.734749     2
#> [1122,] 1.733020     2
#> [1123,] 1.741756     2
#> [1124,] 1.733205     2
#> [1125,] 1.755391     2
#> [1126,] 1.755280     2
#> [1127,] 1.731630     2
#> [1128,] 1.731630     2
#> [1129,] 1.746024     2
#> [1130,] 1.731255     2
#> [1131,] 1.753931     2
#> [1132,] 1.753931     2
#> [1133,] 1.748115     2
#> [1134,] 1.741832     2
#> [1135,] 1.741832     2
#> [1136,] 1.741832     2
#> [1137,] 1.741832     2
#> [1138,] 1.733244     2
#> [1139,] 1.733244     2
#> [1140,] 1.744979     2
#> [1141,] 1.744979     2
#> [1142,] 1.744979     2
#> [1143,] 1.744525     2
#> [1144,] 1.744525     2
#> [1145,] 1.744525     2
#> [1146,] 1.744525     2
#> [1147,] 1.744525     2
#> [1148,] 1.742582     2
#> [1149,] 1.743343     2
#> [1150,] 1.740781     2
#> [1151,] 1.747110     2
#> [1152,] 1.746065     2
#> [1153,] 1.741932     2
#> [1154,] 1.741277     2
#> [1155,] 1.741277     2
#> [1156,] 1.746154     2
#> [1157,] 1.748288     2
#> [1158,] 1.746416     2
#> [1159,] 1.750453     2
#> [1160,] 1.750453     2
#> [1161,] 1.753489     2
#> [1162,] 1.764927     2
#> [1163,] 1.730445     2
#> [1164,] 1.742738     2
#> [1165,] 1.748313     2
#> [1166,] 1.748313     2
#> [1167,] 1.745075     2
#> [1168,] 1.745075     2
#> [1169,] 1.726839     2
#> [1170,] 1.726839     2
#> [1171,] 1.729193     2
#> [1172,] 1.729193     2
#> [1173,] 1.752503     2
#> [1174,] 1.741092     2
#> [1175,] 1.738401     2
#> [1176,] 1.738401     2
#> [1177,] 1.735717     2
#> [1178,] 1.732904     2
#> [1179,] 1.732904     2
#> [1180,] 1.761778     2
#> [1181,] 1.743974     2
#> [1182,] 1.743974     2
#> [1183,] 1.755761     2
#> [1184,] 1.742169     2
#> [1185,] 1.742169     2
#> [1186,] 1.756819     2
#> [1187,] 1.732659     2
#> [1188,] 1.734929     2
#> [1189,] 1.735546     2
#> [1190,] 1.735546     2
#> [1191,] 1.735546     2
#> [1192,] 1.743103     2
#> [1193,] 1.734461     2
#> [1194,] 1.738028     2
#> [1195,] 1.743764     2
#> [1196,] 1.740060     2
#> [1197,] 1.748816     2
#> [1198,] 1.745220     2
#> [1199,] 1.751276     2
#> [1200,] 1.756856     2
#> [1201,] 1.743641     2
#> [1202,] 1.743641     2
#> [1203,] 1.746501     2
#> [1204,] 1.746501     2
#> [1205,] 1.746501     2
#> [1206,] 1.730199     2
#> [1207,] 1.730968     2
#> [1208,] 1.743588     2
#> [1209,] 1.743588     2
#> [1210,] 1.750332     2
#> [1211,] 1.750332     2
#> [1212,] 1.750332     2
#> [1213,] 1.750332     2
#> [1214,] 1.750143     2
#> [1215,] 1.735870     2
#> [1216,] 1.755076     2
#> [1217,] 1.724904     2
#> [1218,] 1.736251     2
#> [1219,] 1.727893     2
#> [1220,] 1.737913     2
#> [1221,] 1.737913     2
#> [1222,] 1.756851     2
#> [1223,] 1.742450     2
#> [1224,] 1.750055     2
#> [1225,] 1.750055     2
#> [1226,] 1.754231     2
#> [1227,] 1.754231     2
#> [1228,] 1.757848     2
#> [1229,] 1.753817     2
#> [1230,] 1.753817     2
#> [1231,] 1.756590     2
#> [1232,] 1.756590     2
#> [1233,] 1.753232     2
#> [1234,] 1.750404     2
#> [1235,] 1.739411     2
#> [1236,] 1.733558     2
#> [1237,] 1.740945     2
#> [1238,] 1.746721     2
#> [1239,] 1.746721     2
#> [1240,] 1.746721     2
#> [1241,] 1.737317     2
#> [1242,] 1.745621     2
#> [1243,] 1.751669     2
#> [1244,] 1.743280     2
#> [1245,] 1.753735     2
#> [1246,] 1.757584     2
#> [1247,] 1.753112     2
#> [1248,] 1.752205     2
#> [1249,] 1.734194     2
#> [1250,] 1.745893     2
#> [1251,] 1.745692     2
#> [1252,] 1.750074     2
#> [1253,] 1.750074     2
#> [1254,] 1.750074     2
#> [1255,] 1.750074     2
#> [1256,] 1.750074     2
#> [1257,] 1.735021     2
#> [1258,] 1.735021     2
#> [1259,] 1.735042     2
#> [1260,] 1.735005     2
#> [1261,] 1.723263     2
#> [1262,] 1.742691     2
#> [1263,] 1.746648     2
#> [1264,] 1.746648     2
#> [1265,] 1.754919     2
#> [1266,] 1.738378     2
#> [1267,] 1.745384     2
#> [1268,] 1.745384     2
#> [1269,] 1.739109     2
#> [1270,] 1.739109     2
#> [1271,] 1.743997     2
#> [1272,] 1.754325     2
#> [1273,] 1.741536     2
#> [1274,] 1.744295     2
#> [1275,] 1.744295     2
#> [1276,] 1.744295     2
#> [1277,] 1.744295     2
#> [1278,] 1.730113     2
#> [1279,] 1.730113     2
#> [1280,] 1.735985     2
#> [1281,] 1.735985     2
#> [1282,] 1.752447     2
#> [1283,] 1.741030     2
#> [1284,] 1.741030     2
#> [1285,] 1.736116     2
#> [1286,] 1.736116     2
#> [1287,] 1.750964     2
#> [1288,] 1.757642     2
#> [1289,] 1.741453     2
#> [1290,] 1.741453     2
#> [1291,] 1.733589     2
#> [1292,] 1.747386     2
#> [1293,] 1.747386     2
#> [1294,] 1.747386     2
#> [1295,] 1.747386     2
#> [1296,] 1.753960     2
#> [1297,] 1.753960     2
#> [1298,] 1.753464     2
#> [1299,] 1.761156     2
#> [1300,] 1.745243     2
#> [1301,] 1.750815     2
#> [1302,] 1.758144     2
#> [1303,] 1.729989     2
#> [1304,] 1.730955     2
#> [1305,] 1.730955     2
#> [1306,] 1.736161     2
#> [1307,] 1.752565     2
#> [1308,] 1.757603     2
#> [1309,] 1.730579     2
#> [1310,] 1.737273     2
#> [1311,] 1.737273     2
#> [1312,] 1.743833     2
#> [1313,] 1.743833     2
#> [1314,] 1.743972     2
#> [1315,] 1.754003     2
#> [1316,] 1.747852     2
#> [1317,] 1.737119     2
#> [1318,] 1.750966     2
#> [1319,] 1.750966     2
#> [1320,] 1.750966     2
#> [1321,] 1.753171     2
#> [1322,] 1.747388     2
#> [1323,] 1.734899     2
#> [1324,] 1.741189     2
#> [1325,] 1.739521     2
#> [1326,] 1.743524     2
#> [1327,] 1.743524     2
#> [1328,] 1.755302     2
#> [1329,] 1.746102     2
#> [1330,] 1.735578     2
#> [1331,] 1.744227     2
#> [1332,] 1.744227     2
#> [1333,] 1.744227     2
#> [1334,] 1.744227     2
#> [1335,] 1.744227     2
#> [1336,] 1.744227     2
#> [1337,] 1.743535     2
#> [1338,] 1.743535     2
#> [1339,] 1.741886     2
#> [1340,] 1.762800     2
#> [1341,] 1.762800     2
#> [1342,] 1.762800     2
#> [1343,] 1.762800     2
#> [1344,] 1.743991     2
#> [1345,] 1.738201     2
#> [1346,] 1.750606     2
#> [1347,] 1.737429     2
#> [1348,] 1.751707     2
#> [1349,] 1.751707     2
#> [1350,] 1.751707     2
#> [1351,] 1.751707     2
#> [1352,] 1.742976     2
#> [1353,] 1.742976     2
#> [1354,] 1.737066     2
#> [1355,] 1.737066     2
#> [1356,] 1.748692     2
#> [1357,] 1.734907     2
#> [1358,] 1.733109     2
#> [1359,] 1.748727     2
#> [1360,] 1.751627     2
#> [1361,] 1.721728     2
#> [1362,] 1.720267     2
#> [1363,] 1.738036     2
#> [1364,] 1.742798     2
#> [1365,] 1.742798     2
#> [1366,] 1.742798     2
#> [1367,] 1.734058     2
#> [1368,] 1.734058     2
#> [1369,] 1.746016     2
#> [1370,] 1.746016     2
#> [1371,] 1.746016     2
#> [1372,] 1.746016     2
#> [1373,] 1.746016     2
#> [1374,] 1.746016     2
#> [1375,] 1.746016     2
#> [1376,] 1.749367     2
#> [1377,] 1.737017     2
#> [1378,] 1.733565     2
#> [1379,] 1.733565     2
#> [1380,] 1.738993     2
#> [1381,] 1.738993     2
#> [1382,] 1.746275     2
#> [1383,] 1.739712     2
#> [1384,] 1.762378     2
#> [1385,] 1.751377     2
#> [1386,] 1.751377     2
#> [1387,] 1.749895     2
#> [1388,] 1.744387     2
#> [1389,] 1.747026     2
#> [1390,] 1.742887     2
#> [1391,] 1.742887     2
#> [1392,] 1.742887     2
#> [1393,] 1.742887     2
#> [1394,] 1.742887     2
#> [1395,] 1.735133     2
#> [1396,] 1.729760     2
#> [1397,] 1.729760     2
#> [1398,] 1.729760     2
#> [1399,] 1.746028     2
#> [1400,] 1.746028     2
#> [1401,] 1.746028     2
#> [1402,] 1.746028     2
#> [1403,] 1.730199     2
#> [1404,] 1.745449     2
#> [1405,] 1.745449     2
#> [1406,] 1.741770     2
#> [1407,] 1.741770     2
#> [1408,] 1.751130     2
#> [1409,] 1.751130     2
#> [1410,] 1.758069     2
#> [1411,] 1.758069     2
#> [1412,] 1.736169     2
#> [1413,] 1.736169     2
#> [1414,] 1.736330     2
#> [1415,] 1.736330     2
#> [1416,] 1.753652     2
#> [1417,] 1.753652     2
#> [1418,] 1.747375     2
#> [1419,] 1.747375     2
#> [1420,] 1.747375     2
#> [1421,] 1.742623     2
#> [1422,] 1.742623     2
#> [1423,] 1.742623     2
#> [1424,] 1.750192     2
#> [1425,] 1.735340     2
#> [1426,] 1.748636     2
#> [1427,] 1.746634     2
#> [1428,] 1.746634     2
#> [1429,] 1.748659     2
#> [1430,] 1.744190     2
#> [1431,] 1.752575     2
#> [1432,] 1.752588     2
#> [1433,] 1.746587     2
#> [1434,] 1.748020     2
#> [1435,] 1.743841     2
#> [1436,] 1.753538     2
#> [1437,] 1.748299     2
#> [1438,] 1.747611     2
#> [1439,] 1.747611     2
#> [1440,] 1.747611     2
#> [1441,] 1.723719     2
#> [1442,] 1.748288     2
#> [1443,] 1.748288     2
#> [1444,] 1.748288     2
#> [1445,] 1.742239     2
#> [1446,] 1.742239     2
#> [1447,] 1.742239     2
#> [1448,] 1.738624     2
#> [1449,] 1.735130     2
#> [1450,] 1.733577     2
#> [1451,] 1.746814     2
#> [1452,] 1.735222     2
#> [1453,] 1.748878     2
#> [1454,] 1.748878     2
#> [1455,] 1.748560     2
#> [1456,] 1.748560     2
#> [1457,] 1.748560     2
#> [1458,] 1.746553     2
#> [1459,] 1.746553     2
#> [1460,] 1.753396     2
#> [1461,] 1.737631     2
#> [1462,] 1.757590     2
#> [1463,] 1.733008     2
#> [1464,] 1.733008     2
#> [1465,] 1.733008     2
#> [1466,] 1.736473     2
#> [1467,] 1.753439     2
#> [1468,] 1.748063     2
#> [1469,] 1.748063     2
#> [1470,] 1.748554     2
#> [1471,] 1.741873     2
#> [1472,] 1.743187     2
#> [1473,] 1.741056     2
#> [1474,] 1.747160     2
#> [1475,] 1.747160     2
#> [1476,] 1.743260     2
#> [1477,] 1.743260     2
#> [1478,] 1.743260     2
#> [1479,] 1.743260     2
#> [1480,] 1.751699     2
#> [1481,] 1.750127     2
#> [1482,] 1.750127     2
#> [1483,] 1.741877     2
#> [1484,] 1.727949     2
#> [1485,] 1.740878     2
#> [1486,] 1.749972     2
#> [1487,] 1.747248     2
#> [1488,] 1.733084     2
#> [1489,] 1.748234     2
#> [1490,] 1.735646     2
#> [1491,] 1.725676     2
#> [1492,] 1.731589     2
#> [1493,] 1.749472     2
#> [1494,] 1.744180     2
#> [1495,] 1.732049     2
#> [1496,] 1.732049     2
#> [1497,] 1.756876     2
#> [1498,] 1.737469     2
#> [1499,] 1.749366     2
#> [1500,] 1.749366     2
#> [1501,] 1.751359     2
#> [1502,] 1.751359     2
#> [1503,] 1.746601     2
#> [1504,] 1.752343     2
#> [1505,] 1.739250     2
#> [1506,] 1.739250     2
#> [1507,] 1.750536     2
#> [1508,] 1.749091     2
#> [1509,] 1.734669     2
#> [1510,] 1.745889     2
#> [1511,] 1.745889     2
#> [1512,] 1.759490     2
#> [1513,] 1.755813     2
#> [1514,] 1.750514     2
#> [1515,] 1.750514     2
#> [1516,] 1.735471     2
#> [1517,] 1.735471     2
#> [1518,] 1.736810     2
#> [1519,] 1.745887     2
#> [1520,] 1.748071     2
#> [1521,] 1.748071     2
#> [1522,] 1.748071     2
#> [1523,] 1.744276     2
#> [1524,] 1.736622     2
#> [1525,] 1.736622     2
#> [1526,] 1.743787     2
#> [1527,] 1.743272     2
#> [1528,] 1.743272     2
#> [1529,] 1.739141     2
#> [1530,] 1.741651     2
#> [1531,] 1.746490     2
#> [1532,] 1.740284     2
#> [1533,] 1.731968     2
#> [1534,] 1.731968     2
#> [1535,] 1.732065     2
#> [1536,] 1.732065     2
#> [1537,] 1.747554     2
#> [1538,] 1.736707     2
#> [1539,] 1.767007     2
#> [1540,] 1.738095     2
#> [1541,] 1.744640     2
#> [1542,] 1.740317     2
#> [1543,] 1.740317     2
#> [1544,] 1.746634     2
#> [1545,] 1.746634     2
#> [1546,] 1.746634     2
#> [1547,] 1.744695     2
#> [1548,] 1.744695     2
#> [1549,] 1.750470     2
#> [1550,] 1.750470     2
#> [1551,] 1.720847     2
#> [1552,] 1.727355     2
#> [1553,] 1.756326     2
#> [1554,] 1.745777     2
#> [1555,] 1.751048     2
#> [1556,] 1.751048     2
#> [1557,] 1.751048     2
#> [1558,] 1.751555     2
#> [1559,] 1.743008     2
#> [1560,] 1.741597     2
#> [1561,] 1.742670     2
#> [1562,] 1.742670     2
#> [1563,] 1.735747     2
#> [1564,] 1.748046     2
#> [1565,] 1.748046     2
#> [1566,] 1.748046     2
#> [1567,] 1.735398     2
#> [1568,] 1.735398     2
#> [1569,] 1.730046     2
#> [1570,] 1.739884     2
#> [1571,] 1.752734     2
#> [1572,] 1.752734     2
#> [1573,] 1.770450     2
#> [1574,] 1.745713     2
#> [1575,] 1.745713     2
#> [1576,] 1.736409     2
#> [1577,] 1.737861     2
#> [1578,] 1.740741     2
#> [1579,] 1.737111     2
#> [1580,] 1.744072     2
#> [1581,] 1.739922     2
#> [1582,] 1.736470     2
#> [1583,] 1.736470     2
#> [1584,] 1.742454     2
#> [1585,] 1.755252     2
#> [1586,] 1.740842     2
#> [1587,] 1.740842     2
#> [1588,] 1.765119     2
#> [1589,] 1.746062     2
#> [1590,] 1.746062     2
#> [1591,] 1.746062     2
#> [1592,] 1.732265     2
#> [1593,] 1.744149     2
#> [1594,] 1.744149     2
#> [1595,] 1.744149     2
#> [1596,] 1.737454     2
#> [1597,] 1.736093     2
#> [1598,] 1.739475     2
#> [1599,] 1.746716     2
#> [1600,] 1.746716     2
#> [1601,] 1.743800     2
#> [1602,] 1.743800     2
#> [1603,] 1.743800     2
#> [1604,] 1.736512     2
#> [1605,] 1.736512     2
#> [1606,] 1.748929     2
#> [1607,] 1.735602     2
#> [1608,] 1.747141     2
#> [1609,] 1.743434     2
#> [1610,] 1.758941     2
#> [1611,] 1.755777     2
#> [1612,] 1.755777     2
#> [1613,] 1.752021     2
#> [1614,] 1.752021     2
#> [1615,] 1.752021     2
#> [1616,] 1.750749     2
#> [1617,] 1.756510     2
#> [1618,] 1.727905     2
#> [1619,] 1.746637     2
#> [1620,] 1.731728     2
#> [1621,] 1.731728     2
#> [1622,] 1.752781     2
#> [1623,] 1.749181     2
#> [1624,] 1.742442     2
#> [1625,] 1.742442     2
#> [1626,] 1.742442     2
#> [1627,] 1.752860     2
#> [1628,] 1.755340     2
#> [1629,] 1.755340     2
#> [1630,] 1.755340     2
#> [1631,] 1.733727     2
#> [1632,] 1.743989     2
#> [1633,] 1.743989     2
#> [1634,] 1.750549     2
#> [1635,] 1.753099     2
#> [1636,] 1.751037     2
#> [1637,] 1.751037     2
#> [1638,] 1.745589     2
#> [1639,] 1.740724     2
#> [1640,] 1.740724     2
#> [1641,] 1.740724     2
#> [1642,] 1.747004     2
#> [1643,] 1.731657     2
#> [1644,] 1.731173     2
#> [1645,] 1.737901     2
#> [1646,] 1.750227     2
#> [1647,] 1.757175     2
#> [1648,] 1.757175     2
#> [1649,] 1.727555     2
#> [1650,] 1.755323     2
#> [1651,] 1.750318     2
#> [1652,] 1.739888     2
#> [1653,] 1.765733     2
#> [1654,] 1.736566     2
#> [1655,] 1.744016     2
#> [1656,] 1.749112     2
#> [1657,] 1.740211     2
#> [1658,] 1.740211     2
#> [1659,] 1.745088     2
#> [1660,] 1.752711     2
#> [1661,] 1.746115     2
#> [1662,] 1.738451     2
#> [1663,] 1.745575     2
#> [1664,] 1.745575     2
#> [1665,] 1.751708     2
#> [1666,] 1.751708     2
#> [1667,] 1.751708     2
#> [1668,] 1.738663     2
#> [1669,] 1.735057     2
#> [1670,] 1.735057     2
#> [1671,] 1.735057     2
#> [1672,] 1.735057     2
#> [1673,] 1.737237     2
#> [1674,] 1.741016     2
#> [1675,] 1.741016     2
#> [1676,] 1.741629     2
#> [1677,] 1.741629     2
#> [1678,] 1.741629     2
#> [1679,] 1.741629     2
#> [1680,] 1.741629     2
#> [1681,] 1.741629     2
#> [1682,] 1.739267     2
#> [1683,] 1.735833     2
#> [1684,] 1.735833     2
#> [1685,] 1.735833     2
#> [1686,] 1.747202     2
#> [1687,] 1.743539     2
#> [1688,] 1.743539     2
#> [1689,] 1.743980     2
#> [1690,] 1.743980     2
#> [1691,] 1.743980     2
#> [1692,] 1.743980     2
#> [1693,] 1.743980     2
#> [1694,] 1.743980     2
#> [1695,] 1.754721     2
#> [1696,] 1.754721     2
#> [1697,] 1.759716     2
#> [1698,] 1.741783     2
#> [1699,] 1.741783     2
#> [1700,] 1.741831     2
#> [1701,] 1.741831     2
#> [1702,] 1.738049     2
#> [1703,] 1.742206     2
#> [1704,] 1.742206     2
#> [1705,] 1.734020     2
#> [1706,] 1.740937     2
#> [1707,] 1.749946     2
#> [1708,] 1.754459     2
#> [1709,] 1.754459     2
#> [1710,] 1.745869     2
#> [1711,] 1.745869     2
#> [1712,] 1.730126     2
#> [1713,] 1.747394     2
#> [1714,] 1.744658     2
#> [1715,] 1.750168     2
#> [1716,] 1.758927     2
#> [1717,] 1.742133     2
#> [1718,] 1.740616     2
#> [1719,] 1.741022     2
#> [1720,] 1.748057     2
#> [1721,] 1.748057     2
#> [1722,] 1.748057     2
#> [1723,] 1.741971     2
#> [1724,] 1.741971     2
#> [1725,] 1.741971     2
#> [1726,] 1.751221     2
#> [1727,] 1.751221     2
#> [1728,] 1.736733     2
#> [1729,] 1.752490     2
#> [1730,] 1.727667     2
#> [1731,] 1.727667     2
#> [1732,] 1.754754     2
#> [1733,] 1.749013     2
#> [1734,] 1.750549     2
#> [1735,] 1.741282     2
#> [1736,] 1.741595     2
#> [1737,] 1.741595     2
#> [1738,] 1.737150     2
#> [1739,] 1.737150     2
#> [1740,] 1.742180     2
#> [1741,] 1.749570     2
#> [1742,] 1.745235     2
#> [1743,] 1.735245     2
#> [1744,] 1.751985     2
#> [1745,] 1.751985     2
#> [1746,] 1.737486     2
#> [1747,] 1.750972     2
#> [1748,] 1.750972     2
#> [1749,] 1.735118     2
#> [1750,] 1.741068     2
#> [1751,] 1.741068     2
#> [1752,] 1.739757     2
#> [1753,] 1.739757     2
#> [1754,] 1.739647     2
#> [1755,] 1.739647     2
#> [1756,] 1.752178     2
#> [1757,] 1.737342     2
#> [1758,] 1.753738     2
#> [1759,] 1.743882     2
#> [1760,] 1.740578     2
#> [1761,] 1.740578     2
#> [1762,] 1.740578     2
#> [1763,] 1.740578     2
#> [1764,] 1.740578     2
#> [1765,] 1.737160     2
#> [1766,] 1.736503     2
#> [1767,] 1.734718     2
#> [1768,] 1.736317     2
#> [1769,] 1.736317     2
#> [1770,] 1.736317     2
#> [1771,] 1.736317     2
#> [1772,] 1.736317     2
#> [1773,] 1.736229     2
#> [1774,] 1.736229     2
#> [1775,] 1.736229     2
#> [1776,] 1.748151     2
#> [1777,] 1.742736     2
#> [1778,] 1.739792     2
#> [1779,] 1.739792     2
#> [1780,] 1.768860     2
#> [1781,] 1.766562     2
#> [1782,] 1.753847     2
#> [1783,] 1.753847     2
#> [1784,] 1.753847     2
#> [1785,] 1.738289     2
#> [1786,] 1.735667     2
#> [1787,] 1.735667     2
#> [1788,] 1.756743     2
#> [1789,] 1.742878     2
#> [1790,] 1.748617     2
#> [1791,] 1.748617     2
#> [1792,] 1.746443     2
#> [1793,] 1.746092     2
#> [1794,] 1.746092     2
#> [1795,] 1.746092     2
#> [1796,] 1.736725     2
#> [1797,] 1.742922     2
#> [1798,] 1.732416     2
#> [1799,] 1.744700     2
#> [1800,] 1.738895     2
#> [1801,] 1.738895     2
#> [1802,] 1.744379     2
#> [1803,] 1.741359     2
#> [1804,] 1.746442     2
#> [1805,] 1.743741     2
#> [1806,] 1.731638     2
#> [1807,] 1.749527     2
#> [1808,] 1.755769     2
#> [1809,] 1.728573     2
#> [1810,] 1.756520     2
#> [1811,] 1.735820     2
#> [1812,] 1.756500     2
#> [1813,] 1.724379     2
#> [1814,] 1.750092     2
#> [1815,] 1.750305     2
#> [1816,] 1.750305     2
#> [1817,] 1.739404     2
#> [1818,] 1.735815     2
#> [1819,] 1.750140     2
#> [1820,] 1.750140     2
#> [1821,] 1.753687     2
#> [1822,] 1.737735     2
#> [1823,] 1.737735     2
#> [1824,] 1.753662     2
#> [1825,] 1.747212     2
#> [1826,] 1.748461     2
#> [1827,] 1.745765     2
#> [1828,] 1.747314     2
#> [1829,] 1.747314     2
#> [1830,] 1.746265     2
#> [1831,] 1.742597     2
#> [1832,] 1.742597     2
#> [1833,] 1.742839     2
#> [1834,] 1.742839     2
#> [1835,] 1.748152     2
#> [1836,] 1.738355     2
#> [1837,] 1.738355     2
#> [1838,] 1.737758     2
#> [1839,] 1.737758     2
#> [1840,] 1.737758     2
#> [1841,] 1.735411     2
#> [1842,] 1.735411     2
#> [1843,] 1.735411     2
#> [1844,] 1.743162     2
#> [1845,] 1.737035     2
#> [1846,] 1.738395     2
#> [1847,] 1.740899     2
#> [1848,] 1.740899     2
#> [1849,] 1.740899     2
#> [1850,] 1.731312     2
#> [1851,] 1.734865     2
#> [1852,] 1.734865     2
#> [1853,] 1.750867     2
#> [1854,] 1.748396     2
#> [1855,] 1.747105     2
#> [1856,] 1.745681     2
#> [1857,] 1.754740     2
#> [1858,] 1.730347     2
#> [1859,] 1.728992     2
#> [1860,] 1.728992     2
#> [1861,] 1.731961     2
#> [1862,] 1.731961     2
#> [1863,] 1.731961     2
#> [1864,] 1.750082     2
#> [1865,] 1.750082     2
#> [1866,] 1.748752     2
#> [1867,] 1.740633     2
#> [1868,] 1.748415     2
#> [1869,] 1.748415     2
#> [1870,] 1.746247     2
#> [1871,] 1.749387     2
#> [1872,] 1.739664     2
#> [1873,] 1.739664     2
#> [1874,] 1.730579     2
#> [1875,] 1.761876     2
#> [1876,] 1.724373     2
#> [1877,] 1.748402     2
#> [1878,] 1.740123     2
#> [1879,] 1.740123     2
#> [1880,] 1.747336     2
#> [1881,] 1.748540     2
#> [1882,] 1.745695     2
#> [1883,] 1.745695     2
#> [1884,] 1.742536     2
#> [1885,] 1.746010     2
#> [1886,] 1.744822     2
#> [1887,] 1.733427     2
#> [1888,] 1.745340     2
#> [1889,] 1.741355     2
#> [1890,] 1.737112     2
#> [1891,] 1.744472     2
#> [1892,] 1.744472     2
#> [1893,] 1.744472     2
#> [1894,] 1.744472     2
#> [1895,] 1.744472     2
#> [1896,] 1.744472     2
#> [1897,] 1.744991     2
#> [1898,] 1.744991     2
#> [1899,] 1.739344     2
#> [1900,] 1.748678     2
#> [1901,] 1.748678     2
#> [1902,] 1.748678     2
#> [1903,] 1.736089     2
#> [1904,] 1.750629     2
#> [1905,] 1.730219     2
#> [1906,] 1.730219     2
#> [1907,] 1.740706     2
#> [1908,] 1.740706     2
#> [1909,] 1.740706     2
#> [1910,] 1.742120     2
#> [1911,] 1.758552     2
#> [1912,] 1.756055     2
#> [1913,] 1.725531     2
#> [1914,] 1.751706     2
#> [1915,] 1.759279     2
#> [1916,] 1.748279     2
#> [1917,] 1.745756     2
#> [1918,] 1.745756     2
#> [1919,] 1.743411     2
#> [1920,] 1.730101     2
#> [1921,] 1.730101     2
#> [1922,] 1.733717     2
#> [1923,] 1.736140     2
#> [1924,] 1.731945     2
#> [1925,] 1.731945     2
#> [1926,] 1.731945     2
#> [1927,] 1.735098     2
#> [1928,] 1.756730     2
#> [1929,] 1.751791     2
#> [1930,] 1.738434     2
#> [1931,] 1.719962     2
#> [1932,] 1.741198     2
#> [1933,] 1.757491     2
#> [1934,] 1.737321     2
#> [1935,] 1.743576     2
#> [1936,] 1.738897     2
#> [1937,] 1.734843     2
#> [1938,] 1.744487     2
#> [1939,] 1.744487     2
#> [1940,] 1.744367     2
#> [1941,] 1.733840     2
#> [1942,] 1.758335     2
#> [1943,] 1.739515     2
#> [1944,] 1.750682     2
#> [1945,] 1.748930     2
#> [1946,] 1.742362     2
#> [1947,] 1.742999     2
#> [1948,] 1.757086     2
#> [1949,] 1.757086     2
#> [1950,] 1.741906     2
#> [1951,] 1.752415     2
#> [1952,] 1.752415     2
#> [1953,] 1.743465     2
#> [1954,] 1.731079     2
#> [1955,] 1.753778     2
#> [1956,] 1.740250     2
#> [1957,] 1.738403     2
#> [1958,] 1.738403     2
#> [1959,] 1.735027     2
#> [1960,] 1.735027     2
#> [1961,] 1.738827     2
#> [1962,] 1.754969     2
#> [1963,] 1.727269     2
#> [1964,] 1.734726     2
#> [1965,] 1.745800     2
#> [1966,] 1.745800     2
#> [1967,] 1.745800     2
#> [1968,] 1.731120     2
#> [1969,] 1.731120     2
#> [1970,] 1.746833     2
#> [1971,] 1.746567     2
#> [1972,] 1.743919     2
#> [1973,] 1.739412     2
#> [1974,] 1.759316     2
#> [1975,] 1.759316     2
#> [1976,] 1.754116     2
#> [1977,] 1.747199     2
#> [1978,] 1.747199     2
#> [1979,] 1.741767     2
#> [1980,] 1.741767     2
#> [1981,] 1.738607     2