Constructor of fitmodel object

A fitmodel object is a list that stores some variables and functions that will be useful to simulate and fit your model during the course.

fitmodel(
  name = NULL,
  stateNames = NULL,
  thetaNames = NULL,
  simulate = NULL,
  rPointObs = NULL,
  dprior = NULL,
  dPointObs = NULL
)

Arguments

name

character. Name of the model (required).

stateNames

character vector. Names of the state variables i.e. c("S","I","R") (required).

thetaNames

character vector. Names of the parameters i.e. c("R_0","infectiousPeriod") (required).

simulate

R-function to simulate forward the model (required). This function takes 3 arguments:

theta named numeric vector. Values of the parameters. Names should match thetaNames.

initState named numeric vector. Initial values of the state variables. Names should match stateNames.

times numeric vector. Time sequence for which the state of the model is wanted; the first value of times must be the initial time, i.e. the time of initState.

and returns a data.fame containing the simulated trajectories that is the values of the state variables (1 per column) at each observation time (1 per row). The first column is time.

rPointObs

R-function that generates a (randomly sampled) observation point from a model point, using an observation model (optional). It thus acts as an inverse of dPointObs (see below). This function takes 2 arguments

modelPoint named numeric vector. State of the model at a given point in time.

theta named numeric vector. Values of the parameters. Names should match thetaNames.

and returns an observation point

dprior

R-function that evaluates the prior density of the parameters at a given theta (optional). The function should take 2 arguments:

theta named numeric vector. Values of the parameters. Names should match thetaNames.

log boolean. determines whether the logarithm of the prior density should be returned.

and returns the (logged, if requested) value of the prior density distribution.

dPointObs

R-function that evaluates the likelihood of one data point given the state of the model at the same time point. This function takes 4 arguments:

dataPoint named numeric vector. Observation time and observed data point.

modelPoint named numeric vector containing the state of the model at the observation time point.

theta named numeric vector. Parameter values. Useful since parameters are usually needed to compute the likelihood (i.e. reporting rate).

log boolean. determines whether the logarithm of the likelihood should be returned.

and returns the (log-)likelihood. (optional)

Value

a fitmodel object that is a list of 7 elements:

name character, name of the model

stateNames vector, names of the state variables.

thetaNames vector, names of the parameters.

simulate R-function to simulate forward the model; usage: simulate(theta,initState,times).

rPointObs R-function to generate simulated observations; usage: rPointObs(modelPoint, theta).

dprior R-function to evaluate the log-prior of the parameter values; usage: dprior(theta).

dPointObs R-function to evaluate the log-likelihood of one data point; usage: dPointObs(dataPoint, modelPoint, theta, log).

See also

Examples

## create a simple deterministic SIR model with constant population size

SIR_name <- "SIR with constant population size" # nolint
SIR_stateNames <- c("S", "I", "R") # nolint
SIR_thetaNames <- c("R_0", "D_inf") # nolint

SIR_simulateDeterministic <- function(theta, initState, times) { # nolint
  SIR_ode <- function(time, state, parameters) { # nolint
    ## parameters
    beta <- parameters[["R_0"]] / parameters[["D_inf"]]
    nu <- 1 / parameters[["D_inf"]]

    ## states
    S <- state[["S"]] # nolint
    I <- state[["I"]] # nolint
    R <- state[["R"]] # nolint

    N <- S + I + R # nolint

    dS <- -beta * S * I / N
    dI <- beta * S * I / N - nu * I
    dR <- nu * I

    return(list(c(dS, dI, dR)))
  }

  trajectory <- data.frame(deSolve::ode(
    y = initState,
    times = times,
    func = SIR_ode,
    parms = theta,
    method = "ode45"
  ))

  return(trajectory)
}

## function to compute log-prior
SIR_prior <- function(theta, log = FALSE) { # nolint
  ## uniform prior on R_0: U[1,100]
  logPrior_R_0 <- dunif(theta[["R_0"]], min = 1, max = 100, log = TRUE) # nolint
  ## uniform prior on infectious period: U[0,30]
  logPrior_D <- dunif(theta[["D_inf"]], min = 0, max = 30, log = TRUE) # nolint

  logSum <- logPrior_R_0 + logPrior_D

  return(ifelse(log, logSum, exp(logSum)))
}

## function to compute the likelihood of one data point
SIR_pointLike <- function(dataPoint, modelPoint, theta, log = FALSE) { # nolint
  ## the prevalence is observed through a Poisson process
  return(dpois(
    x = dataPoint[["obs"]],
    lambda = modelPoint[["I"]],
    log = log
  ))
}

## function to generate observation from a model simulation
SIR_genObsPoint <- function(modelPoint, theta) { # nolint
  ## the prevalence is observed through a Poisson process
  obsPoint <- rpois(n = 1, lambda = modelPoint[["I"]])

  return(c(obs = obsPoint))
}

## create deterministic SIR fitmodel
SIR_deter <- fitmodel( # nolint
  name = SIR_name,
  stateNames = SIR_stateNames,
  thetaNames = SIR_thetaNames,
  simulate = SIR_simulateDeterministic,
  dprior = SIR_prior,
  rPointObs = SIR_genObsPoint,
  dPointObs = SIR_pointLike
)